Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3157076 0.4266751
#> 2     yes      1 0.2814584 0.2145039 0.3350099
#> 3      no      2 0.3927672 0.3431952 0.4441189
#> 4     yes      2 0.3039568 0.2398541 0.3568450
#> 5      no      3 0.4189750 0.3702653 0.4628073
#> 6     yes      3 0.3274340 0.2675692 0.3805087
#> 7      no      4 0.4456481 0.3979825 0.4932096
#> 8     yes      4 0.3518080 0.2939305 0.4021034
#> 9      no      5 0.4726379 0.4239279 0.5244761
#> 10    yes      5 0.3769795 0.3156372 0.4294703
#> 11     no      6 0.4997887 0.4514512 0.5608996
#> 12    yes      6 0.4028326 0.3419883 0.4615442
#> 13     no      7 0.5269407 0.4718897 0.5996728
#> 14    yes      7 0.4292372 0.3645204 0.4989373
#> 15     no      8 0.5539343 0.4920820 0.6351647
#> 16    yes      8 0.4560509 0.3824886 0.5347129
#> 17     no      9 0.5806134 0.5098458 0.6733263
#> 18    yes      9 0.4831217 0.4016140 0.5698258
#> 19     no     10 0.6068295 0.5256975 0.7093287
#> 20    yes     10 0.5102919 0.4189228 0.6101802
#> 21     no     11 0.6324443 0.5425307 0.7428825
#> 22    yes     11 0.5374015 0.4364403 0.6476004
#> 23     no     12 0.6573325 0.5601201 0.7738008
#> 24    yes     12 0.5642918 0.4543351 0.6832930
#> 25     no     13 0.6813846 0.5767847 0.8019934
#> 26    yes     13 0.5908091 0.4693192 0.7169500
#> 27     no     14 0.7045073 0.5925080 0.8274561
#> 28    yes     14 0.6168084 0.4843591 0.7481850
#> 29     no     15 0.7266252 0.6080438 0.8502556
#> 30    yes     15 0.6421564 0.4994274 0.7770173
#> 31     no     16 0.7476805 0.6235304 0.8705132
#> 32    yes     16 0.6667336 0.5144968 0.8042436
#> 33     no     17 0.7676327 0.6387787 0.8883891
#> 34    yes     17 0.6904365 0.5296561 0.8294722
#> 35     no     18 0.7864578 0.6537517 0.9040680
#> 36    yes     18 0.7131789 0.5447626 0.8519459
#> 37     no     19 0.8041469 0.6684254 0.9177471
#> 38    yes     19 0.7348919 0.5597871 0.8718662
#> 39     no     20 0.8207047 0.6827779 0.9295516
#> 40    yes     20 0.7555247 0.5747025 0.8894524
#> 41     no     21 0.8361480 0.6967900 0.9397681
#> 42    yes     21 0.7750433 0.5894830 0.9048866
#> 43     no     22 0.8505034 0.7104445 0.9485839
#> 44    yes     22 0.7934296 0.6041033 0.9183622
#> 45     no     23 0.8638059 0.7237266 0.9561688
#> 46    yes     23 0.8106802 0.6185394 0.9300748
#> 47     no     24 0.8760972 0.7366241 0.9626780
#> 48    yes     24 0.8268046 0.6327687 0.9402151
#> 49     no     25 0.8874238 0.7491269 0.9682521
#> 50    yes     25 0.8418238 0.6467698 0.9489646
#> 51     no     26 0.8978357 0.7612272 0.9730164
#> 52    yes     26 0.8557676 0.6605230 0.9564920
#> 53     no     27 0.9073851 0.7729193 0.9770822
#> 54    yes     27 0.8686741 0.6740100 0.9629517
#> 55     no     28 0.9161253 0.7841999 0.9805473
#> 56    yes     28 0.8805867 0.6872143 0.9684832
#> 57     no     29 0.9241097 0.7950673 0.9834969
#> 58    yes     29 0.8915537 0.7001211 0.9732113
#> 59     no     30 0.9313909 0.8055218 0.9860054
#> 60    yes     30 0.9016260 0.7127174 0.9772463
#> 61     no     31 0.9380204 0.8155653 0.9881370
#> 62    yes     31 0.9108563 0.7249917 0.9806852
#> 63     no     32 0.9440478 0.8252013 0.9899470
#> 64    yes     32 0.9192981 0.7369345 0.9836127
#> 65     no     33 0.9495206 0.8344349 0.9914830
#> 66    yes     33 0.9270045 0.7485379 0.9861025
#> 67     no     34 0.9544839 0.8432721 0.9927860
#> 68    yes     34 0.9340278 0.7597956 0.9882184
#> 69     no     35 0.9589802 0.8517202 0.9938907
#> 70    yes     35 0.9404188 0.7707031 0.9900151
#> 71     no     36 0.9630496 0.8593301 0.9948270
#> 72    yes     36 0.9462263 0.7812572 0.9915400
#> 73     no     37 0.9667293 0.8664674 0.9956203
#> 74    yes     37 0.9514970 0.7914564 0.9928336
#> 75     no     38 0.9700539 0.8732962 0.9962924
#> 76    yes     38 0.9562749 0.8013003 0.9939305
#> 77     no     39 0.9730556 0.8798242 0.9968616
#> 78    yes     39 0.9606016 0.8107900 0.9948603
#> 79     no     40 0.9757639 0.8860599 0.9973436
#> 80    yes     40 0.9645161 0.8199278 0.9956481
#> 81     no     41 0.9782061 0.8920118 0.9977517
#> 82    yes     41 0.9680546 0.8287169 0.9963156
#> 83     no     42 0.9804072 0.8976887 0.9980972
#> 84    yes     42 0.9712507 0.8371617 0.9968810
#> 85     no     43 0.9823899 0.9030995 0.9983896
#> 86    yes     43 0.9741357 0.8452673 0.9973597
#> 87     no     44 0.9841753 0.9082534 0.9986372
#> 88    yes     44 0.9767380 0.8530398 0.9977652
#> 89     no     45 0.9857822 0.9131595 0.9988467
#> 90    yes     45 0.9790842 0.8604858 0.9981084
#> 91     no     46 0.9872281 0.9178269 0.9990240
#> 92    yes     46 0.9811982 0.8676126 0.9983990
#> 93     no     47 0.9885287 0.9222647 0.9991740
#> 94    yes     47 0.9831023 0.8744281 0.9986450
#> 95     no     48 0.9896982 0.9264820 0.9993010
#> 96    yes     48 0.9848165 0.8809404 0.9988532
#> 97     no     49 0.9907496 0.9304876 0.9994085
#> 98    yes     49 0.9863593 0.8871582 0.9990294
#> 99     no     50 0.9916946 0.9342904 0.9994994
#> 100   yes     50 0.9877472 0.8930902 0.9991785
#> 101    no     51 0.9925437 0.9378989 0.9995764
#> 102   yes     51 0.9889955 0.8987457 0.9993048
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)