Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3092909 0.4242321
#> 2     yes      1 0.2814584 0.2147912 0.3560626
#> 3      no      2 0.3927672 0.3383129 0.4411452
#> 4     yes      2 0.3039568 0.2368779 0.3762468
#> 5      no      3 0.4189750 0.3646717 0.4638270
#> 6     yes      3 0.3274340 0.2598613 0.3968690
#> 7      no      4 0.4456481 0.3918071 0.4899550
#> 8     yes      4 0.3518080 0.2842486 0.4190455
#> 9      no      5 0.4726379 0.4122421 0.5203989
#> 10    yes      5 0.3769795 0.3099687 0.4510899
#> 11     no      6 0.4997887 0.4330017 0.5508338
#> 12    yes      6 0.4028326 0.3369210 0.4750030
#> 13     no      7 0.5269407 0.4550893 0.5861130
#> 14    yes      7 0.4292372 0.3609080 0.4979961
#> 15     no      8 0.5539343 0.4769059 0.6211278
#> 16    yes      8 0.4560509 0.3878051 0.5250738
#> 17     no      9 0.5806134 0.4957001 0.6566955
#> 18    yes      9 0.4831217 0.4084798 0.5605610
#> 19     no     10 0.6068295 0.5149126 0.6894285
#> 20    yes     10 0.5102919 0.4326191 0.6008043
#> 21     no     11 0.6324443 0.5346593 0.7235154
#> 22    yes     11 0.5374015 0.4537261 0.6418949
#> 23     no     12 0.6573325 0.5542887 0.7555302
#> 24    yes     12 0.5642918 0.4750010 0.6814651
#> 25     no     13 0.6813846 0.5718761 0.7849421
#> 26    yes     13 0.5908091 0.4963672 0.7185343
#> 27     no     14 0.7045073 0.5869594 0.8116991
#> 28    yes     14 0.6168084 0.5170204 0.7528073
#> 29     no     15 0.7266252 0.6020716 0.8358246
#> 30    yes     15 0.6421564 0.5358841 0.7841150
#> 31     no     16 0.7476805 0.6176556 0.8574032
#> 32    yes     16 0.6667336 0.5514571 0.8124029
#> 33     no     17 0.7676327 0.6330027 0.8765651
#> 34    yes     17 0.6904365 0.5669300 0.8377129
#> 35     no     18 0.7864578 0.6480856 0.8934725
#> 36    yes     18 0.7131789 0.5822735 0.8598478
#> 37     no     19 0.8041469 0.6628792 0.9083066
#> 38    yes     19 0.7348919 0.5974594 0.8786105
#> 39     no     20 0.8207047 0.6773604 0.9212573
#> 40    yes     20 0.7555247 0.6124606 0.8951707
#> 41     no     21 0.8361480 0.6915085 0.9325149
#> 42    yes     21 0.7750433 0.6272515 0.9097066
#> 43     no     22 0.8505034 0.7053053 0.9422639
#> 44    yes     22 0.7934296 0.6418716 0.9224038
#> 45     no     23 0.8638059 0.7187348 0.9506791
#> 46    yes     23 0.8106802 0.6562404 0.9334479
#> 47     no     24 0.8760972 0.7317834 0.9579225
#> 48    yes     24 0.8268046 0.6703286 0.9430185
#> 49     no     25 0.8874238 0.7446548 0.9641422
#> 50    yes     25 0.8418238 0.6841173 0.9512858
#> 51     no     26 0.8978357 0.7571300 0.9694718
#> 52    yes     26 0.8557676 0.6975892 0.9584072
#> 53     no     27 0.9073851 0.7691824 0.9740305
#> 54    yes     27 0.8686741 0.7107294 0.9645271
#> 55     no     28 0.9161253 0.7808074 0.9779240
#> 56    yes     28 0.8805867 0.7235247 0.9697753
#> 57     no     29 0.9241097 0.7914907 0.9812449
#> 58    yes     29 0.8915537 0.7359641 0.9742682
#> 59     no     30 0.9313909 0.8014824 0.9840744
#> 60    yes     30 0.9016260 0.7480385 0.9781085
#> 61     no     31 0.9380204 0.8111095 0.9864828
#> 62    yes     31 0.9108563 0.7597408 0.9813868
#> 63     no     32 0.9440478 0.8203744 0.9885313
#> 64    yes     32 0.9192981 0.7710656 0.9841822
#> 65     no     33 0.9495206 0.8292806 0.9902723
#> 66    yes     33 0.9270045 0.7820096 0.9865637
#> 67     no     34 0.9544839 0.8378324 0.9917512
#> 68    yes     34 0.9340278 0.7925710 0.9885908
#> 69     no     35 0.9589802 0.8460354 0.9930069
#> 70    yes     35 0.9404188 0.8027495 0.9903151
#> 71     no     36 0.9630496 0.8538958 0.9940725
#> 72    yes     36 0.9462263 0.8125465 0.9917811
#> 73     no     37 0.9667293 0.8614207 0.9949766
#> 74    yes     37 0.9514970 0.8219646 0.9930267
#> 75     no     38 0.9700539 0.8686176 0.9957433
#> 76    yes     38 0.9562749 0.8310076 0.9940847
#> 77     no     39 0.9730556 0.8754948 0.9963932
#> 78    yes     39 0.9606016 0.8396808 0.9949830
#> 79     no     40 0.9757639 0.8820609 0.9969430
#> 80    yes     40 0.9645161 0.8479900 0.9957454
#> 81     no     41 0.9782061 0.8883249 0.9974092
#> 82    yes     41 0.9680546 0.8559423 0.9963925
#> 83     no     42 0.9804072 0.8942961 0.9978045
#> 84    yes     42 0.9712507 0.8635453 0.9969414
#> 85     no     43 0.9823899 0.8999840 0.9981395
#> 86    yes     43 0.9741357 0.8708074 0.9974070
#> 87     no     44 0.9841753 0.9053981 0.9984235
#> 88    yes     44 0.9767380 0.8777375 0.9978019
#> 89     no     45 0.9857822 0.9105484 0.9986643
#> 90    yes     45 0.9790842 0.8843452 0.9981367
#> 91     no     46 0.9872281 0.9154444 0.9988682
#> 92    yes     46 0.9811982 0.8906400 0.9984207
#> 93     no     47 0.9885287 0.9200960 0.9990411
#> 94    yes     47 0.9831023 0.8966322 0.9986614
#> 95     no     48 0.9896982 0.9245127 0.9991876
#> 96    yes     48 0.9848165 0.9023318 0.9988655
#> 97     no     49 0.9907496 0.9287043 0.9993117
#> 98    yes     49 0.9863593 0.9077494 0.9990384
#> 99     no     50 0.9916946 0.9326801 0.9994169
#> 100   yes     50 0.9877472 0.9128954 0.9991851
#> 101    no     51 0.9925437 0.9364493 0.9995060
#> 102   yes     51 0.9889955 0.9177802 0.9993094
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)