Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3149241 0.4191075
#> 2     yes      1 0.2814584 0.2224971 0.3454611
#> 3      no      2 0.3927672 0.3443340 0.4380272
#> 4     yes      2 0.3039568 0.2443399 0.3654071
#> 5      no      3 0.4189750 0.3728443 0.4581700
#> 6     yes      3 0.3274340 0.2691401 0.3861385
#> 7      no      4 0.4456481 0.3960284 0.4843954
#> 8     yes      4 0.3518080 0.2960267 0.4137219
#> 9      no      5 0.4726379 0.4203153 0.5131247
#> 10    yes      5 0.3769795 0.3198713 0.4363652
#> 11     no      6 0.4997887 0.4435596 0.5492190
#> 12    yes      6 0.4028326 0.3425235 0.4662505
#> 13     no      7 0.5269407 0.4632728 0.5851719
#> 14    yes      7 0.4292372 0.3631384 0.5018259
#> 15     no      8 0.5539343 0.4821533 0.6216809
#> 16    yes      8 0.4560509 0.3835005 0.5446885
#> 17     no      9 0.5806134 0.5029905 0.6590614
#> 18    yes      9 0.4831217 0.4033433 0.5869077
#> 19     no     10 0.6068295 0.5230212 0.6945631
#> 20    yes     10 0.5102919 0.4235075 0.6278907
#> 21     no     11 0.6324443 0.5430170 0.7278943
#> 22    yes     11 0.5374015 0.4433891 0.6671137
#> 23     no     12 0.6573325 0.5593138 0.7592172
#> 24    yes     12 0.5642918 0.4614186 0.7013837
#> 25     no     13 0.6813846 0.5759420 0.7889940
#> 26    yes     13 0.5908091 0.4799854 0.7324755
#> 27     no     14 0.7045073 0.5924186 0.8158578
#> 28    yes     14 0.6168084 0.4986252 0.7655309
#> 29     no     15 0.7266252 0.6086152 0.8400362
#> 30    yes     15 0.6421564 0.5166651 0.7952663
#> 31     no     16 0.7476805 0.6228394 0.8620296
#> 32    yes     16 0.6667336 0.5332344 0.8219525
#> 33     no     17 0.7676327 0.6368549 0.8814235
#> 34    yes     17 0.6904365 0.5497308 0.8458199
#> 35     no     18 0.7864578 0.6506416 0.8984102
#> 36    yes     18 0.7131789 0.5661187 0.8669908
#> 37     no     19 0.8041469 0.6647939 0.9132010
#> 38    yes     19 0.7348919 0.5823634 0.8856338
#> 39     no     20 0.8207047 0.6786848 0.9260140
#> 40    yes     20 0.7555247 0.5984315 0.9019463
#> 41     no     21 0.8361480 0.6922665 0.9370645
#> 42    yes     21 0.7750433 0.6142908 0.9161262
#> 43     no     22 0.8505034 0.7055231 0.9465584
#> 44    yes     22 0.7934296 0.6299113 0.9283785
#> 45     no     23 0.8638059 0.7184406 0.9546884
#> 46    yes     23 0.8106802 0.6452646 0.9390865
#> 47     no     24 0.8760972 0.7310072 0.9616308
#> 48    yes     24 0.8268046 0.6598553 0.9482816
#> 49     no     25 0.8874238 0.7432129 0.9675449
#> 50    yes     25 0.8418238 0.6737892 0.9561525
#> 51     no     26 0.8978357 0.7550498 0.9725729
#> 52    yes     26 0.8557676 0.6874220 0.9628717
#> 53     no     27 0.9073851 0.7665779 0.9768400
#> 54    yes     27 0.8686741 0.7007373 0.9685944
#> 55     no     28 0.9161253 0.7777856 0.9804562
#> 56    yes     28 0.8805867 0.7138516 0.9734588
#> 57     no     29 0.9241097 0.7886027 0.9835169
#> 58    yes     29 0.8915537 0.7266837 0.9775868
#> 59     no     30 0.9313909 0.7990285 0.9861047
#> 60    yes     30 0.9016260 0.7391493 0.9810848
#> 61     no     31 0.9380204 0.8090638 0.9882909
#> 62    yes     31 0.9108563 0.7512395 0.9840455
#> 63     no     32 0.9440478 0.8187109 0.9901363
#> 64    yes     32 0.9192981 0.7629474 0.9865489
#> 65     no     33 0.9495206 0.8279732 0.9916932
#> 66    yes     33 0.9270045 0.7742681 0.9886638
#> 67     no     34 0.9544839 0.8368558 0.9930059
#> 68    yes     34 0.9340278 0.7851986 0.9904492
#> 69     no     35 0.9589802 0.8453645 0.9941123
#> 70    yes     35 0.9404188 0.7957376 0.9919555
#> 71     no     36 0.9630496 0.8535062 0.9950444
#> 72    yes     36 0.9462263 0.8058854 0.9932258
#> 73     no     37 0.9667293 0.8612887 0.9958295
#> 74    yes     37 0.9514970 0.8156441 0.9942965
#> 75     no     38 0.9700539 0.8687205 0.9964906
#> 76    yes     38 0.9562749 0.8250167 0.9951987
#> 77     no     39 0.9730556 0.8758108 0.9970471
#> 78    yes     39 0.9606016 0.8340080 0.9959587
#> 79     no     40 0.9757639 0.8825690 0.9975156
#> 80    yes     40 0.9645161 0.8426235 0.9965988
#> 81     no     41 0.9782061 0.8890053 0.9979098
#> 82    yes     41 0.9680546 0.8508700 0.9971377
#> 83     no     42 0.9804072 0.8951300 0.9982416
#> 84    yes     42 0.9712507 0.8587551 0.9975913
#> 85     no     43 0.9823899 0.9009536 0.9985207
#> 86    yes     43 0.9741357 0.8662870 0.9979732
#> 87     no     44 0.9841753 0.9064870 0.9987556
#> 88    yes     44 0.9767380 0.8734748 0.9982946
#> 89     no     45 0.9857822 0.9117408 0.9989532
#> 90    yes     45 0.9790842 0.8803280 0.9985651
#> 91     no     46 0.9872281 0.9167260 0.9991194
#> 92    yes     46 0.9811982 0.8868565 0.9987927
#> 93     no     47 0.9885287 0.9214532 0.9992592
#> 94    yes     47 0.9831023 0.8930707 0.9989842
#> 95     no     48 0.9896982 0.9259332 0.9993769
#> 96    yes     48 0.9848165 0.8989810 0.9991454
#> 97     no     49 0.9907496 0.9301765 0.9994758
#> 98    yes     49 0.9863593 0.9045981 0.9992810
#> 99     no     50 0.9916946 0.9341934 0.9995591
#> 100   yes     50 0.9877472 0.9099329 0.9993951
#> 101    no     51 0.9925437 0.9379942 0.9996291
#> 102   yes     51 0.9889955 0.9149963 0.9994910
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)