Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3053247 0.4193049
#> 2     yes      1 0.2814584 0.2277914 0.3407395
#> 3      no      2 0.3927672 0.3396703 0.4444520
#> 4     yes      2 0.3039568 0.2511071 0.3579101
#> 5      no      3 0.4189750 0.3702594 0.4713430
#> 6     yes      3 0.3274340 0.2718575 0.3768924
#> 7      no      4 0.4456481 0.3978461 0.5016816
#> 8     yes      4 0.3518080 0.2898953 0.4046541
#> 9      no      5 0.4726379 0.4233291 0.5329393
#> 10    yes      5 0.3769795 0.3087698 0.4330178
#> 11     no      6 0.4997887 0.4481529 0.5624413
#> 12    yes      6 0.4028326 0.3283179 0.4645052
#> 13     no      7 0.5269407 0.4748997 0.5915074
#> 14    yes      7 0.4292372 0.3519501 0.4985278
#> 15     no      8 0.5539343 0.4945382 0.6226635
#> 16    yes      8 0.4560509 0.3791640 0.5264974
#> 17     no      9 0.5806134 0.5133336 0.6529069
#> 18    yes      9 0.4831217 0.4071863 0.5559985
#> 19     no     10 0.6068295 0.5331189 0.6814704
#> 20    yes     10 0.5102919 0.4362181 0.5898380
#> 21     no     11 0.6324443 0.5521132 0.7090346
#> 22    yes     11 0.5374015 0.4604553 0.6228467
#> 23     no     12 0.6573325 0.5684209 0.7369149
#> 24    yes     12 0.5642918 0.4823750 0.6547522
#> 25     no     13 0.6813846 0.5846709 0.7630169
#> 26    yes     13 0.5908091 0.5018405 0.6853603
#> 27     no     14 0.7045073 0.6007365 0.7872768
#> 28    yes     14 0.6168084 0.5193507 0.7174670
#> 29     no     15 0.7266252 0.6165850 0.8113456
#> 30    yes     15 0.6421564 0.5359400 0.7473884
#> 31     no     16 0.7476805 0.6321858 0.8332283
#> 32    yes     16 0.6667336 0.5515954 0.7747689
#> 33     no     17 0.7676327 0.6475100 0.8528747
#> 34    yes     17 0.6904365 0.5667746 0.7999658
#> 35     no     18 0.7864578 0.6625313 0.8717174
#> 36    yes     18 0.7131789 0.5818288 0.8229768
#> 37     no     19 0.8041469 0.6772257 0.8886968
#> 38    yes     19 0.7348919 0.5967316 0.8430699
#> 39     no     20 0.8207047 0.6915718 0.9036771
#> 40    yes     20 0.7555247 0.6114572 0.8612608
#> 41     no     21 0.8361480 0.7058631 0.9168299
#> 42    yes     21 0.7750433 0.6259813 0.8784013
#> 43     no     22 0.8505034 0.7198062 0.9283292
#> 44    yes     22 0.7934296 0.6402812 0.8936784
#> 45     no     23 0.8638059 0.7333339 0.9383455
#> 46    yes     23 0.8106802 0.6543354 0.9071641
#> 47     no     24 0.8760972 0.7464343 0.9470418
#> 48    yes     24 0.8268046 0.6681242 0.9189297
#> 49     no     25 0.8874238 0.7590985 0.9545709
#> 50    yes     25 0.8418238 0.6816298 0.9293135
#> 51     no     26 0.8978357 0.7713197 0.9610497
#> 52    yes     26 0.8557676 0.6948358 0.9384501
#> 53     no     27 0.9073851 0.7827278 0.9664450
#> 54    yes     27 0.8686741 0.7077280 0.9464681
#> 55     no     28 0.9161253 0.7934605 0.9711141
#> 56    yes     28 0.8805867 0.7202938 0.9534882
#> 57     no     29 0.9241097 0.8037957 0.9751491
#> 58    yes     29 0.8915537 0.7325227 0.9596225
#> 59     no     30 0.9313909 0.8137349 0.9786088
#> 60    yes     30 0.9016260 0.7444056 0.9649734
#> 61     no     31 0.9380204 0.8232812 0.9815799
#> 62    yes     31 0.9108563 0.7559356 0.9696342
#> 63     no     32 0.9440478 0.8324389 0.9841443
#> 64    yes     32 0.9192981 0.7671073 0.9736887
#> 65     no     33 0.9495206 0.8412134 0.9863559
#> 66    yes     33 0.9270045 0.7779170 0.9772120
#> 67     no     34 0.9544839 0.8496114 0.9882622
#> 68    yes     34 0.9340278 0.7883625 0.9802707
#> 69     no     35 0.9589802 0.8576403 0.9899043
#> 70    yes     35 0.9404188 0.7985929 0.9829242
#> 71     no     36 0.9630496 0.8653084 0.9913183
#> 72    yes     36 0.9462263 0.8091819 0.9852245
#> 73     no     37 0.9667293 0.8726248 0.9925353
#> 74    yes     37 0.9514970 0.8193399 0.9872389
#> 75     no     38 0.9700539 0.8795990 0.9935825
#> 76    yes     38 0.9562749 0.8290707 0.9890133
#> 77     no     39 0.9730556 0.8862411 0.9944834
#> 78    yes     39 0.9606016 0.8383800 0.9905430
#> 79     no     40 0.9757639 0.8925615 0.9952581
#> 80    yes     40 0.9645161 0.8472747 0.9918610
#> 81     no     41 0.9782061 0.8985709 0.9959243
#> 82    yes     41 0.9680546 0.8557631 0.9929963
#> 83     no     42 0.9804072 0.9042802 0.9964971
#> 84    yes     42 0.9712507 0.8638544 0.9939739
#> 85     no     43 0.9823899 0.9097005 0.9969894
#> 86    yes     43 0.9741357 0.8715587 0.9948156
#> 87     no     44 0.9841753 0.9148426 0.9974126
#> 88    yes     44 0.9767380 0.8788869 0.9955400
#> 89     no     45 0.9857822 0.9197178 0.9977764
#> 90    yes     45 0.9790842 0.8858504 0.9961634
#> 91     no     46 0.9872281 0.9243370 0.9980890
#> 92    yes     46 0.9811982 0.8924613 0.9966998
#> 93     no     47 0.9885287 0.9287110 0.9983576
#> 94    yes     47 0.9831023 0.8987927 0.9971613
#> 95     no     48 0.9896982 0.9328505 0.9985885
#> 96    yes     48 0.9848165 0.9048327 0.9975583
#> 97     no     49 0.9907496 0.9367660 0.9987869
#> 98    yes     49 0.9863593 0.9105468 0.9978998
#> 99     no     50 0.9916946 0.9404677 0.9989574
#> 100   yes     50 0.9877472 0.9159484 0.9981935
#> 101    no     51 0.9925437 0.9439657 0.9991039
#> 102   yes     51 0.9889955 0.9210509 0.9984462
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)