Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3070686 0.4219829
#> 2     yes      1 0.2814584 0.2282485 0.3669317
#> 3      no      2 0.3927672 0.3376746 0.4371179
#> 4     yes      2 0.3039568 0.2533091 0.3789576
#> 5      no      3 0.4189750 0.3668891 0.4578862
#> 6     yes      3 0.3274340 0.2760016 0.3965159
#> 7      no      4 0.4456481 0.3971332 0.4847075
#> 8     yes      4 0.3518080 0.2978619 0.4225840
#> 9      no      5 0.4726379 0.4257610 0.5163110
#> 10    yes      5 0.3769795 0.3183614 0.4506115
#> 11     no      6 0.4997887 0.4510446 0.5523628
#> 12    yes      6 0.4028326 0.3396016 0.4789379
#> 13     no      7 0.5269407 0.4751900 0.5855178
#> 14    yes      7 0.4292372 0.3627956 0.5072854
#> 15     no      8 0.5539343 0.4983677 0.6193430
#> 16    yes      8 0.4560509 0.3892555 0.5353359
#> 17     no      9 0.5806134 0.5184266 0.6524008
#> 18    yes      9 0.4831217 0.4164116 0.5608555
#> 19     no     10 0.6068295 0.5330050 0.6846915
#> 20    yes     10 0.5102919 0.4393412 0.5898044
#> 21     no     11 0.6324443 0.5460311 0.7157243
#> 22    yes     11 0.5374015 0.4563158 0.6226915
#> 23     no     12 0.6573325 0.5589932 0.7448331
#> 24    yes     12 0.5642918 0.4703837 0.6574497
#> 25     no     13 0.6813846 0.5718739 0.7719035
#> 26    yes     13 0.5908091 0.4844960 0.6905161
#> 27     no     14 0.7045073 0.5846560 0.7977905
#> 28    yes     14 0.6168084 0.4986290 0.7215927
#> 29     no     15 0.7266252 0.5973232 0.8218823
#> 30    yes     15 0.6421564 0.5139531 0.7506737
#> 31     no     16 0.7476805 0.6098597 0.8436682
#> 32    yes     16 0.6667336 0.5313519 0.7787039
#> 33     no     17 0.7676327 0.6222501 0.8632346
#> 34    yes     17 0.6904365 0.5465733 0.8045189
#> 35     no     18 0.7864578 0.6344803 0.8806996
#> 36    yes     18 0.7131789 0.5617099 0.8279876
#> 37     no     19 0.8041469 0.6465366 0.8962034
#> 38    yes     19 0.7348919 0.5767491 0.8491668
#> 39     no     20 0.8207047 0.6584064 0.9098992
#> 40    yes     20 0.7555247 0.5926036 0.8681533
#> 41     no     21 0.8361480 0.6700778 0.9219457
#> 42    yes     21 0.7750433 0.6082542 0.8850516
#> 43     no     22 0.8505034 0.6815402 0.9325013
#> 44    yes     22 0.7934296 0.6234592 0.9000076
#> 45     no     23 0.8638059 0.6927836 0.9417199
#> 46    yes     23 0.8106802 0.6366195 0.9132083
#> 47     no     24 0.8760972 0.7037992 0.9497474
#> 48    yes     24 0.8268046 0.6495704 0.9248116
#> 49     no     25 0.8874238 0.7145791 0.9567201
#> 50    yes     25 0.8418238 0.6622960 0.9349740
#> 51     no     26 0.8978357 0.7251164 0.9627633
#> 52    yes     26 0.8557676 0.6747815 0.9438460
#> 53     no     27 0.9073851 0.7354053 0.9679908
#> 54    yes     27 0.8686741 0.6870138 0.9515515
#> 55     no     28 0.9161253 0.7454409 0.9725055
#> 56    yes     28 0.8805867 0.6989809 0.9581984
#> 57     no     29 0.9241097 0.7552190 0.9763988
#> 58    yes     29 0.8915537 0.7106726 0.9639680
#> 59     no     30 0.9313909 0.7647366 0.9797523
#> 60    yes     30 0.9016260 0.7220797 0.9689670
#> 61     no     31 0.9380204 0.7739916 0.9826378
#> 62    yes     31 0.9108563 0.7331947 0.9732916
#> 63     no     32 0.9440478 0.7829824 0.9851182
#> 64    yes     32 0.9192981 0.7440115 0.9770279
#> 65     no     33 0.9495206 0.7917085 0.9872489
#> 66    yes     33 0.9270045 0.7545252 0.9802521
#> 67     no     34 0.9544839 0.8001702 0.9890778
#> 68    yes     34 0.9340278 0.7647322 0.9830317
#> 69     no     35 0.9589802 0.8083681 0.9906469
#> 70    yes     35 0.9404188 0.7746302 0.9854258
#> 71     no     36 0.9630496 0.8163039 0.9919924
#> 72    yes     36 0.9462263 0.7837531 0.9874864
#> 73     no     37 0.9667293 0.8239797 0.9931456
#> 74    yes     37 0.9514970 0.7924249 0.9892589
#> 75     no     38 0.9700539 0.8313981 0.9941337
#> 76    yes     38 0.9562749 0.8008338 0.9907827
#> 77     no     39 0.9730556 0.8385624 0.9949801
#> 78    yes     39 0.9606016 0.8089805 0.9920920
#> 79     no     40 0.9757639 0.8454762 0.9957049
#> 80    yes     40 0.9645161 0.8168668 0.9932166
#> 81     no     41 0.9782061 0.8521435 0.9963254
#> 82    yes     41 0.9680546 0.8244949 0.9941822
#> 83     no     42 0.9804072 0.8585688 0.9968565
#> 84    yes     42 0.9712507 0.8318675 0.9950110
#> 85     no     43 0.9823899 0.8647569 0.9973111
#> 86    yes     43 0.9741357 0.8389880 0.9957223
#> 87     no     44 0.9841753 0.8707127 0.9977000
#> 88    yes     44 0.9767380 0.8458598 0.9963325
#> 89     no     45 0.9857822 0.8764415 0.9980328
#> 90    yes     45 0.9790842 0.8524873 0.9968560
#> 91     no     46 0.9872281 0.8819489 0.9983176
#> 92    yes     46 0.9811982 0.8588747 0.9973050
#> 93     no     47 0.9885287 0.8872404 0.9985611
#> 94    yes     47 0.9831023 0.8650269 0.9976899
#> 95     no     48 0.9896982 0.8923218 0.9987692
#> 96    yes     48 0.9848165 0.8709489 0.9980200
#> 97     no     49 0.9907496 0.8971989 0.9989467
#> 98    yes     49 0.9863593 0.8766459 0.9983031
#> 99     no     50 0.9916946 0.9018778 0.9990974
#> 100   yes     50 0.9877472 0.8821233 0.9985457
#> 101    no     51 0.9925437 0.9063643 0.9992265
#> 102   yes     51 0.9889955 0.8873868 0.9987536
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)