Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3102827 0.4258573
#> 2     yes      1 0.2814584 0.2337789 0.3494475
#> 3      no      2 0.3927672 0.3399091 0.4428406
#> 4     yes      2 0.3039568 0.2559953 0.3675565
#> 5      no      3 0.4189750 0.3708113 0.4654639
#> 6     yes      3 0.3274340 0.2809787 0.3879018
#> 7      no      4 0.4456481 0.4009931 0.4935059
#> 8     yes      4 0.3518080 0.3012079 0.4108474
#> 9      no      5 0.4726379 0.4211908 0.5217929
#> 10    yes      5 0.3769795 0.3220570 0.4355394
#> 11     no      6 0.4997887 0.4357650 0.5496003
#> 12    yes      6 0.4028326 0.3439981 0.4690067
#> 13     no      7 0.5269407 0.4469013 0.5844135
#> 14    yes      7 0.4292372 0.3669240 0.4971342
#> 15     no      8 0.5539343 0.4583164 0.6224452
#> 16    yes      8 0.4560509 0.3862582 0.5343794
#> 17     no      9 0.5806134 0.4698459 0.6588308
#> 18    yes      9 0.4831217 0.4020709 0.5720474
#> 19     no     10 0.6068295 0.4814076 0.6937899
#> 20    yes     10 0.5102919 0.4188344 0.6074359
#> 21     no     11 0.6324443 0.4929888 0.7264846
#> 22    yes     11 0.5374015 0.4295530 0.6416700
#> 23     no     12 0.6573325 0.5045767 0.7569076
#> 24    yes     12 0.5642918 0.4400287 0.6778454
#> 25     no     13 0.6813846 0.5161586 0.7849443
#> 26    yes     13 0.5908091 0.4505587 0.7133783
#> 27     no     14 0.7045073 0.5277626 0.8105529
#> 28    yes     14 0.6168084 0.4611337 0.7456572
#> 29     no     15 0.7266252 0.5396224 0.8337541
#> 30    yes     15 0.6421564 0.4723328 0.7754494
#> 31     no     16 0.7476805 0.5514362 0.8553240
#> 32    yes     16 0.6667336 0.4862890 0.8026790
#> 33     no     17 0.7676327 0.5631909 0.8745372
#> 34    yes     17 0.6904365 0.5009587 0.8273448
#> 35     no     18 0.7864578 0.5748736 0.8915229
#> 36    yes     18 0.7131789 0.5160635 0.8495078
#> 37     no     19 0.8041469 0.5864717 0.9064553
#> 38    yes     19 0.7348919 0.5311303 0.8692773
#> 39     no     20 0.8207047 0.5979729 0.9195181
#> 40    yes     20 0.7555247 0.5461258 0.8867973
#> 41     no     21 0.8361480 0.6093654 0.9308959
#> 42    yes     21 0.7750433 0.5610175 0.9022344
#> 43     no     22 0.8505034 0.6206381 0.9407690
#> 44    yes     22 0.7934296 0.5757739 0.9157673
#> 45     no     23 0.8638059 0.6317800 0.9493085
#> 46    yes     23 0.8106802 0.5872041 0.9275778
#> 47     no     24 0.8760972 0.6427811 0.9566735
#> 48    yes     24 0.8268046 0.5983472 0.9378453
#> 49     no     25 0.8874238 0.6536315 0.9630103
#> 50    yes     25 0.8418238 0.6093840 0.9467410
#> 51     no     26 0.8978357 0.6643224 0.9684508
#> 52    yes     26 0.8557676 0.6203045 0.9544256
#> 53     no     27 0.9073851 0.6748452 0.9731136
#> 54    yes     27 0.8686741 0.6310988 0.9610472
#> 55     no     28 0.9161253 0.6851923 0.9771035
#> 56    yes     28 0.8805867 0.6417575 0.9667403
#> 57     no     29 0.9241097 0.6953565 0.9805132
#> 58    yes     29 0.8915537 0.6522722 0.9716260
#> 59     no     30 0.9313909 0.7053314 0.9834237
#> 60    yes     30 0.9016260 0.6626344 0.9758120
#> 61     no     31 0.9380204 0.7151113 0.9859058
#> 62    yes     31 0.9108563 0.6728367 0.9793936
#> 63     no     32 0.9440478 0.7246910 0.9880208
#> 64    yes     32 0.9192981 0.6828722 0.9824544
#> 65     no     33 0.9495206 0.7340662 0.9898217
#> 66    yes     33 0.9270045 0.6929679 0.9850675
#> 67     no     34 0.9544839 0.7432330 0.9913542
#> 68    yes     34 0.9340278 0.7029400 0.9872964
#> 69     no     35 0.9589802 0.7521883 0.9926576
#> 70    yes     35 0.9404188 0.7127193 0.9891963
#> 71     no     36 0.9630496 0.7609296 0.9937659
#> 72    yes     36 0.9462263 0.7223005 0.9908147
#> 73     no     37 0.9667293 0.7694550 0.9947077
#> 74    yes     37 0.9514970 0.7316793 0.9921925
#> 75     no     38 0.9700539 0.7777631 0.9955079
#> 76    yes     38 0.9562749 0.7408520 0.9933651
#> 77     no     39 0.9730556 0.7858530 0.9961875
#> 78    yes     39 0.9606016 0.7498154 0.9943625
#> 79     no     40 0.9757639 0.7937246 0.9967647
#> 80    yes     40 0.9645161 0.7585670 0.9952107
#> 81     no     41 0.9782061 0.8013781 0.9972547
#> 82    yes     41 0.9680546 0.7671048 0.9959318
#> 83     no     42 0.9804072 0.8088141 0.9976707
#> 84    yes     42 0.9712507 0.7754274 0.9965447
#> 85     no     43 0.9823899 0.8160338 0.9980238
#> 86    yes     43 0.9741357 0.7835340 0.9970656
#> 87     no     44 0.9841753 0.8230388 0.9983234
#> 88    yes     44 0.9767380 0.7914243 0.9975081
#> 89     no     45 0.9857822 0.8298310 0.9985777
#> 90    yes     45 0.9790842 0.7990984 0.9978840
#> 91     no     46 0.9872281 0.8364127 0.9987935
#> 92    yes     46 0.9811982 0.8065570 0.9982033
#> 93     no     47 0.9885287 0.8427864 0.9989765
#> 94    yes     47 0.9831023 0.8138011 0.9984745
#> 95     no     48 0.9896982 0.8489552 0.9991318
#> 96    yes     48 0.9848165 0.8208322 0.9987048
#> 97     no     49 0.9907496 0.8549221 0.9992636
#> 98    yes     49 0.9863593 0.8276521 0.9989004
#> 99     no     50 0.9916946 0.8606905 0.9993754
#> 100   yes     50 0.9877472 0.8342630 0.9990665
#> 101    no     51 0.9925437 0.8662642 0.9994702
#> 102   yes     51 0.9889955 0.8406674 0.9992075
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)