Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3136298 0.4207573
#> 2     yes      1 0.2814584 0.2138580 0.3512377
#> 3      no      2 0.3927672 0.3393807 0.4455149
#> 4     yes      2 0.3039568 0.2385647 0.3677844
#> 5      no      3 0.4189750 0.3703649 0.4725156
#> 6     yes      3 0.3274340 0.2651873 0.3856367
#> 7      no      4 0.4456481 0.3970540 0.4983445
#> 8     yes      4 0.3518080 0.2936366 0.4106209
#> 9      no      5 0.4726379 0.4176120 0.5245468
#> 10    yes      5 0.3769795 0.3215453 0.4448869
#> 11     no      6 0.4997887 0.4395451 0.5510169
#> 12    yes      6 0.4028326 0.3455054 0.4700169
#> 13     no      7 0.5269407 0.4609055 0.5826071
#> 14    yes      7 0.4292372 0.3717436 0.4958268
#> 15     no      8 0.5539343 0.4805673 0.6147926
#> 16    yes      8 0.4560509 0.3939558 0.5293583
#> 17     no      9 0.5806134 0.4985384 0.6465673
#> 18    yes      9 0.4831217 0.4142190 0.5634550
#> 19     no     10 0.6068295 0.5142751 0.6776806
#> 20    yes     10 0.5102919 0.4339209 0.5969649
#> 21     no     11 0.6324443 0.5299862 0.7085839
#> 22    yes     11 0.5374015 0.4538340 0.6295964
#> 23     no     12 0.6573325 0.5456383 0.7378006
#> 24    yes     12 0.5642918 0.4742867 0.6610887
#> 25     no     13 0.6813846 0.5617999 0.7652613
#> 26    yes     13 0.5908091 0.4940433 0.6912194
#> 27     no     14 0.7045073 0.5778828 0.7906621
#> 28    yes     14 0.6168084 0.5142098 0.7198092
#> 29     no     15 0.7266252 0.5937956 0.8139826
#> 30    yes     15 0.6421564 0.5319461 0.7476181
#> 31     no     16 0.7476805 0.6095054 0.8352464
#> 32    yes     16 0.6667336 0.5474280 0.7738505
#> 33     no     17 0.7676327 0.6249809 0.8545141
#> 34    yes     17 0.6904365 0.5623417 0.7989184
#> 35     no     18 0.7864578 0.6401929 0.8718741
#> 36    yes     18 0.7131789 0.5771412 0.8218430
#> 37     no     19 0.8041469 0.6551146 0.8874355
#> 38    yes     19 0.7348919 0.5918007 0.8426621
#> 39     no     20 0.8207047 0.6697216 0.9013208
#> 40    yes     20 0.7555247 0.6062953 0.8614502
#> 41     no     21 0.8361480 0.6839919 0.9136600
#> 42    yes     21 0.7750433 0.6206013 0.8781236
#> 43     no     22 0.8505034 0.6979065 0.9245853
#> 44    yes     22 0.7934296 0.6346964 0.8929141
#> 45     no     23 0.8638059 0.7114486 0.9345358
#> 46    yes     23 0.8106802 0.6485596 0.9060928
#> 47     no     24 0.8760972 0.7246043 0.9430314
#> 48    yes     24 0.8268046 0.6621717 0.9177905
#> 49     no     25 0.8874238 0.7373624 0.9503953
#> 50    yes     25 0.8418238 0.6755150 0.9281389
#> 51     no     26 0.8978357 0.7495437 0.9568509
#> 52    yes     26 0.8557676 0.6885736 0.9376059
#> 53     no     27 0.9073851 0.7606027 0.9624995
#> 54    yes     27 0.8686741 0.7013331 0.9459328
#> 55     no     28 0.9161253 0.7713150 0.9674332
#> 56    yes     28 0.8805867 0.7137813 0.9532024
#> 57     no     29 0.9241097 0.7816784 0.9717334
#> 58    yes     29 0.8915537 0.7259073 0.9595351
#> 59     no     30 0.9313909 0.7916922 0.9754877
#> 60    yes     30 0.9016260 0.7377022 0.9650412
#> 61     no     31 0.9380204 0.8013568 0.9787712
#> 62    yes     31 0.9108563 0.7491587 0.9697436
#> 63     no     32 0.9440478 0.8106739 0.9816230
#> 64    yes     32 0.9192981 0.7602714 0.9737650
#> 65     no     33 0.9495206 0.8196462 0.9840962
#> 66    yes     33 0.9270045 0.7710360 0.9772646
#> 67     no     34 0.9544839 0.8282774 0.9862412
#> 68    yes     34 0.9340278 0.7814503 0.9803069
#> 69     no     35 0.9589802 0.8365724 0.9881003
#> 70    yes     35 0.9404188 0.7915130 0.9829492
#> 71     no     36 0.9630496 0.8445366 0.9897109
#> 72    yes     36 0.9462263 0.8012244 0.9852423
#> 73     no     37 0.9667293 0.8521762 0.9911055
#> 74    yes     37 0.9514970 0.8105861 0.9872311
#> 75     no     38 0.9700539 0.8594980 0.9923125
#> 76    yes     38 0.9562749 0.8196006 0.9889549
#> 77     no     39 0.9730556 0.8665095 0.9933568
#> 78    yes     39 0.9606016 0.8282717 0.9904482
#> 79     no     40 0.9757639 0.8732184 0.9942600
#> 80    yes     40 0.9645161 0.8366040 0.9917413
#> 81     no     41 0.9782061 0.8796329 0.9950411
#> 82    yes     41 0.9680546 0.8446029 0.9928606
#> 83     no     42 0.9804072 0.8857615 0.9957163
#> 84    yes     42 0.9712507 0.8522746 0.9938292
#> 85     no     43 0.9823899 0.8916129 0.9963000
#> 86    yes     43 0.9741357 0.8596261 0.9946670
#> 87     no     44 0.9841753 0.8971960 0.9968043
#> 88    yes     44 0.9767380 0.8666646 0.9953917
#> 89     no     45 0.9857822 0.9025198 0.9972401
#> 90    yes     45 0.9790842 0.8733980 0.9960182
#> 91     no     46 0.9872281 0.9075933 0.9976167
#> 92    yes     46 0.9811982 0.8798347 0.9965599
#> 93     no     47 0.9885287 0.9124256 0.9979419
#> 94    yes     47 0.9831023 0.8859831 0.9970281
#> 95     no     48 0.9896982 0.9170257 0.9982229
#> 96    yes     48 0.9848165 0.8918521 0.9974327
#> 97     no     49 0.9907496 0.9214025 0.9984655
#> 98    yes     49 0.9863593 0.8974506 0.9977824
#> 99     no     50 0.9916946 0.9255649 0.9986751
#> 100   yes     50 0.9877472 0.9027878 0.9980845
#> 101    no     51 0.9925437 0.9295216 0.9988561
#> 102   yes     51 0.9889955 0.9078727 0.9983456
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)