Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2967162 0.4394405
#> 2     yes      1 0.2814584 0.2108292 0.3553991
#> 3      no      2 0.3927672 0.3282651 0.4558844
#> 4     yes      2 0.3039568 0.2356886 0.3784076
#> 5      no      3 0.4189750 0.3578509 0.4719002
#> 6     yes      3 0.3274340 0.2611913 0.4002526
#> 7      no      4 0.4456481 0.3875011 0.4953247
#> 8     yes      4 0.3518080 0.2884167 0.4225011
#> 9      no      5 0.4726379 0.4178673 0.5223380
#> 10    yes      5 0.3769795 0.3172653 0.4450662
#> 11     no      6 0.4997887 0.4480033 0.5521883
#> 12    yes      6 0.4028326 0.3470303 0.4731354
#> 13     no      7 0.5269407 0.4724160 0.5839129
#> 14    yes      7 0.4292372 0.3716205 0.5060441
#> 15     no      8 0.5539343 0.4967039 0.6157378
#> 16    yes      8 0.4560509 0.3945606 0.5374115
#> 17     no      9 0.5806134 0.5155884 0.6487233
#> 18    yes      9 0.4831217 0.4080241 0.5698598
#> 19     no     10 0.6068295 0.5320102 0.6817895
#> 20    yes     10 0.5102919 0.4246832 0.6034284
#> 21     no     11 0.6324443 0.5457060 0.7153196
#> 22    yes     11 0.5374015 0.4420377 0.6360536
#> 23     no     12 0.6573325 0.5584921 0.7469154
#> 24    yes     12 0.5642918 0.4595358 0.6673143
#> 25     no     13 0.6813846 0.5711753 0.7760721
#> 26    yes     13 0.5908091 0.4771290 0.6972117
#> 27     no     14 0.7045073 0.5854127 0.8013492
#> 28    yes     14 0.6168084 0.4947675 0.7304695
#> 29     no     15 0.7266252 0.6001202 0.8244135
#> 30    yes     15 0.6421564 0.5124014 0.7613220
#> 31     no     16 0.7476805 0.6146361 0.8451453
#> 32    yes     16 0.6667336 0.5293350 0.7894137
#> 33     no     17 0.7676327 0.6289372 0.8641288
#> 34    yes     17 0.6904365 0.5445127 0.8149153
#> 35     no     18 0.7864578 0.6430015 0.8814875
#> 36    yes     18 0.7131789 0.5595961 0.8379567
#> 37     no     19 0.8041469 0.6568090 0.8968924
#> 38    yes     19 0.7348919 0.5745573 0.8586214
#> 39     no     20 0.8207047 0.6703414 0.9104978
#> 40    yes     20 0.7555247 0.5891861 0.8770318
#> 41     no     21 0.8361480 0.6835823 0.9224625
#> 42    yes     21 0.7750433 0.6032537 0.8933372
#> 43     no     22 0.8505034 0.6965169 0.9329452
#> 44    yes     22 0.7934296 0.6171432 0.9077032
#> 45     no     23 0.8638059 0.7091327 0.9420993
#> 46    yes     23 0.8106802 0.6308347 0.9203026
#> 47     no     24 0.8760972 0.7214189 0.9500702
#> 48    yes     24 0.8268046 0.6443092 0.9313083
#> 49     no     25 0.8874238 0.7333663 0.9570148
#> 50    yes     25 0.8418238 0.6575494 0.9408886
#> 51     no     26 0.8978357 0.7449680 0.9630563
#> 52    yes     26 0.8557676 0.6705393 0.9492027
#> 53     no     27 0.9073851 0.7562183 0.9682768
#> 54    yes     27 0.8686741 0.6832643 0.9563992
#> 55     no     28 0.9161253 0.7671137 0.9727803
#> 56    yes     28 0.8805867 0.6957117 0.9626142
#> 57     no     29 0.9241097 0.7776517 0.9766599
#> 58    yes     29 0.8915537 0.7078699 0.9679712
#> 59     no     30 0.9313909 0.7878317 0.9799978
#> 60    yes     30 0.9016260 0.7197291 0.9725809
#> 61     no     31 0.9380204 0.7976544 0.9828668
#> 62    yes     31 0.9108563 0.7312809 0.9764603
#> 63     no     32 0.9440478 0.8071215 0.9853303
#> 64    yes     32 0.9192981 0.7425185 0.9798186
#> 65     no     33 0.9495206 0.8162361 0.9874442
#> 66    yes     33 0.9270045 0.7534364 0.9827105
#> 67     no     34 0.9544839 0.8250024 0.9892567
#> 68    yes     34 0.9340278 0.7640306 0.9851943
#> 69     no     35 0.9589802 0.8334252 0.9908100
#> 70    yes     35 0.9404188 0.7742984 0.9873259
#> 71     no     36 0.9630496 0.8415105 0.9921405
#> 72    yes     36 0.9462263 0.7842383 0.9891540
#> 73     no     37 0.9667293 0.8492649 0.9932797
#> 74    yes     37 0.9514970 0.7938501 0.9907209
#> 75     no     38 0.9700539 0.8566955 0.9942547
#> 76    yes     38 0.9562749 0.8031346 0.9920632
#> 77     no     39 0.9730556 0.8638100 0.9950889
#> 78    yes     39 0.9606016 0.8120936 0.9932127
#> 79     no     40 0.9757639 0.8706167 0.9958025
#> 80    yes     40 0.9645161 0.8207300 0.9941967
#> 81     no     41 0.9782061 0.8771241 0.9964128
#> 82    yes     41 0.9680546 0.8290474 0.9950387
#> 83     no     42 0.9804072 0.8833409 0.9969346
#> 84    yes     42 0.9712507 0.8370501 0.9957591
#> 85     no     43 0.9823899 0.8892763 0.9973807
#> 86    yes     43 0.9741357 0.8447434 0.9963753
#> 87     no     44 0.9841753 0.8949393 0.9977620
#> 88    yes     44 0.9767380 0.8521329 0.9969022
#> 89     no     45 0.9857822 0.9003394 0.9980879
#> 90    yes     45 0.9790842 0.8592248 0.9973527
#> 91     no     46 0.9872281 0.9054857 0.9983665
#> 92    yes     46 0.9811982 0.8660260 0.9977378
#> 93     no     47 0.9885287 0.9103877 0.9986045
#> 94    yes     47 0.9831023 0.8725434 0.9980671
#> 95     no     48 0.9896982 0.9150544 0.9988078
#> 96    yes     48 0.9848165 0.8787844 0.9983484
#> 97     no     49 0.9907496 0.9194952 0.9989816
#> 98    yes     49 0.9863593 0.8847568 0.9985889
#> 99     no     50 0.9916946 0.9237190 0.9991301
#> 100   yes     50 0.9877472 0.8904684 0.9987944
#> 101    no     51 0.9925437 0.9277347 0.9992569
#> 102   yes     51 0.9889955 0.8959273 0.9989700
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)