Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3140235 0.4268983
#> 2     yes      1 0.2814584 0.2262367 0.3669178
#> 3      no      2 0.3927672 0.3411404 0.4444575
#> 4     yes      2 0.3039568 0.2511818 0.3852756
#> 5      no      3 0.4189750 0.3697156 0.4651207
#> 6     yes      3 0.3274340 0.2765373 0.4034391
#> 7      no      4 0.4456481 0.3965402 0.4890417
#> 8     yes      4 0.3518080 0.3019406 0.4261369
#> 9      no      5 0.4726379 0.4226307 0.5159255
#> 10    yes      5 0.3769795 0.3268265 0.4493202
#> 11     no      6 0.4997887 0.4459617 0.5496507
#> 12    yes      6 0.4028326 0.3491637 0.4745843
#> 13     no      7 0.5269407 0.4683955 0.5856087
#> 14    yes      7 0.4292372 0.3679074 0.4997206
#> 15     no      8 0.5539343 0.4909579 0.6196265
#> 16    yes      8 0.4560509 0.3820526 0.5289917
#> 17     no      9 0.5806134 0.5135571 0.6536970
#> 18    yes      9 0.4831217 0.3964229 0.5642994
#> 19     no     10 0.6068295 0.5343513 0.6862557
#> 20    yes     10 0.5102919 0.4167643 0.6019350
#> 21     no     11 0.6324443 0.5543467 0.7167260
#> 22    yes     11 0.5374015 0.4366166 0.6383638
#> 23     no     12 0.6573325 0.5706031 0.7447692
#> 24    yes     12 0.5642918 0.4527773 0.6732113
#> 25     no     13 0.6813846 0.5850338 0.7708925
#> 26    yes     13 0.5908091 0.4690399 0.7061757
#> 27     no     14 0.7045073 0.5995858 0.7950431
#> 28    yes     14 0.6168084 0.4853697 0.7370348
#> 29     no     15 0.7266252 0.6142040 0.8181553
#> 30    yes     15 0.6421564 0.5017314 0.7639674
#> 31     no     16 0.7476805 0.6286144 0.8403171
#> 32    yes     16 0.6667336 0.5183905 0.7908313
#> 33     no     17 0.7676327 0.6427938 0.8602363
#> 34    yes     17 0.6904365 0.5350389 0.8153885
#> 35     no     18 0.7864578 0.6567207 0.8780288
#> 36    yes     18 0.7131789 0.5516085 0.8376627
#> 37     no     19 0.8041469 0.6703754 0.8938336
#> 38    yes     19 0.7348919 0.5680626 0.8577234
#> 39     no     20 0.8207047 0.6837399 0.9078037
#> 40    yes     20 0.7555247 0.5843655 0.8756759
#> 41     no     21 0.8361480 0.6967984 0.9200984
#> 42    yes     21 0.7750433 0.6004830 0.8916504
#> 43     no     22 0.8505034 0.7095369 0.9308771
#> 44    yes     22 0.7934296 0.6156181 0.9057930
#> 45     no     23 0.8638059 0.7219433 0.9402951
#> 46    yes     23 0.8106802 0.6304631 0.9182580
#> 47     no     24 0.8760972 0.7340076 0.9485000
#> 48    yes     24 0.8268046 0.6450644 0.9292012
#> 49     no     25 0.8874238 0.7456294 0.9556299
#> 50    yes     25 0.8418238 0.6593995 0.9387753
#> 51     no     26 0.8978357 0.7566949 0.9618118
#> 52    yes     26 0.8557676 0.6734477 0.9471267
#> 53     no     27 0.9073851 0.7674209 0.9671615
#> 54    yes     27 0.8686741 0.6871904 0.9543925
#> 55     no     28 0.9161253 0.7778047 0.9717833
#> 56    yes     28 0.8805867 0.7006109 0.9606996
#> 57     no     29 0.9241097 0.7878450 0.9757706
#> 58    yes     29 0.8915537 0.7136945 0.9661638
#> 59     no     30 0.9313909 0.7975416 0.9792062
#> 60    yes     30 0.9016260 0.7264286 0.9708899
#> 61     no     31 0.9380204 0.8072769 0.9821633
#> 62    yes     31 0.9108563 0.7388026 0.9749623
#> 63     no     32 0.9440478 0.8168398 0.9847062
#> 64    yes     32 0.9192981 0.7508080 0.9784418
#> 65     no     33 0.9495206 0.8260176 0.9868912
#> 66    yes     33 0.9270045 0.7624382 0.9814460
#> 67     no     34 0.9544839 0.8345435 0.9887674
#> 68    yes     34 0.9340278 0.7736882 0.9840376
#> 69     no     35 0.9589802 0.8427316 0.9903776
#> 70    yes     35 0.9404188 0.7845552 0.9862714
#> 71     no     36 0.9630496 0.8505875 0.9917588
#> 72    yes     36 0.9462263 0.7950378 0.9881958
#> 73     no     37 0.9667293 0.8581175 0.9929430
#> 74    yes     37 0.9514970 0.8051363 0.9898526
#> 75     no     38 0.9700539 0.8653284 0.9939580
#> 76    yes     38 0.9562749 0.8148523 0.9912784
#> 77     no     39 0.9730556 0.8722279 0.9948277
#> 78    yes     39 0.9606016 0.8241889 0.9925050
#> 79     no     40 0.9757639 0.8788237 0.9955727
#> 80    yes     40 0.9645161 0.8331505 0.9935599
#> 81     no     41 0.9782061 0.8851243 0.9962107
#> 82    yes     41 0.9680546 0.8417423 0.9944668
#> 83     no     42 0.9804072 0.8911382 0.9967571
#> 84    yes     42 0.9712507 0.8499707 0.9952463
#> 85     no     43 0.9823899 0.8968742 0.9972248
#> 86    yes     43 0.9741357 0.8578430 0.9959163
#> 87     no     44 0.9841753 0.9023414 0.9976252
#> 88    yes     44 0.9767380 0.8652099 0.9964919
#> 89     no     45 0.9857822 0.9075489 0.9979679
#> 90    yes     45 0.9790842 0.8721182 0.9969865
#> 91     no     46 0.9872281 0.9125060 0.9982613
#> 92    yes     46 0.9811982 0.8787224 0.9974114
#> 93     no     47 0.9885287 0.9172217 0.9985123
#> 94    yes     47 0.9831023 0.8850307 0.9977764
#> 95     no     48 0.9896982 0.9217052 0.9987271
#> 96    yes     48 0.9848165 0.8910517 0.9980900
#> 97     no     49 0.9907496 0.9259658 0.9989109
#> 98    yes     49 0.9863593 0.8967944 0.9983593
#> 99     no     50 0.9916946 0.9300122 0.9990682
#> 100   yes     50 0.9877472 0.9022677 0.9985906
#> 101    no     51 0.9925437 0.9338535 0.9992027
#> 102   yes     51 0.9889955 0.9074809 0.9987893
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)