Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2987853 0.4250289
#> 2     yes      1 0.2814584 0.2131146 0.3504238
#> 3      no      2 0.3927672 0.3292705 0.4482993
#> 4     yes      2 0.3039568 0.2389357 0.3679316
#> 5      no      3 0.4189750 0.3615671 0.4710311
#> 6     yes      3 0.3274340 0.2668314 0.3890460
#> 7      no      4 0.4456481 0.3877187 0.4955840
#> 8     yes      4 0.3518080 0.2959099 0.4130398
#> 9      no      5 0.4726379 0.4187285 0.5232808
#> 10    yes      5 0.3769795 0.3182921 0.4385108
#> 11     no      6 0.4997887 0.4440652 0.5566055
#> 12    yes      6 0.4028326 0.3383556 0.4644336
#> 13     no      7 0.5269407 0.4688139 0.5896105
#> 14    yes      7 0.4292372 0.3614855 0.4951777
#> 15     no      8 0.5539343 0.4896659 0.6251123
#> 16    yes      8 0.4560509 0.3772776 0.5245221
#> 17     no      9 0.5806134 0.5087930 0.6593169
#> 18    yes      9 0.4831217 0.3933348 0.5517780
#> 19     no     10 0.6068295 0.5278879 0.6932940
#> 20    yes     10 0.5102919 0.4096263 0.5844619
#> 21     no     11 0.6324443 0.5452572 0.7253470
#> 22    yes     11 0.5374015 0.4261187 0.6182653
#> 23     no     12 0.6573325 0.5621268 0.7551368
#> 24    yes     12 0.5642918 0.4428501 0.6509337
#> 25     no     13 0.6813846 0.5788539 0.7820507
#> 26    yes     13 0.5908091 0.4597597 0.6820629
#> 27     no     14 0.7045073 0.5954020 0.8063862
#> 28    yes     14 0.6168084 0.4767625 0.7116445
#> 29     no     15 0.7266252 0.6117360 0.8281344
#> 30    yes     15 0.6421564 0.4938192 0.7395234
#> 31     no     16 0.7476805 0.6278230 0.8477569
#> 32    yes     16 0.6667336 0.5108902 0.7657868
#> 33     no     17 0.7676327 0.6436318 0.8655004
#> 34    yes     17 0.6904365 0.5279355 0.7909661
#> 35     no     18 0.7864578 0.6591338 0.8814627
#> 36    yes     18 0.7131789 0.5449155 0.8140957
#> 37     no     19 0.8041469 0.6743028 0.8957566
#> 38    yes     19 0.7348919 0.5617913 0.8351992
#> 39     no     20 0.8207047 0.6891153 0.9088287
#> 40    yes     20 0.7555247 0.5785249 0.8543356
#> 41     no     21 0.8361480 0.7035506 0.9206728
#> 42    yes     21 0.7750433 0.5950795 0.8715915
#> 43     no     22 0.8505034 0.7175904 0.9310950
#> 44    yes     22 0.7934296 0.6114199 0.8870732
#> 45     no     23 0.8638059 0.7312195 0.9402368
#> 46    yes     23 0.8106802 0.6275131 0.9009004
#> 47     no     24 0.8760972 0.7444253 0.9482333
#> 48    yes     24 0.8268046 0.6433278 0.9131989
#> 49     no     25 0.8874238 0.7571979 0.9552107
#> 50    yes     25 0.8418238 0.6588354 0.9240684
#> 51     no     26 0.8978357 0.7695300 0.9612860
#> 52    yes     26 0.8557676 0.6740096 0.9336755
#> 53     no     27 0.9073851 0.7814167 0.9665661
#> 54    yes     27 0.8686741 0.6888269 0.9421431
#> 55     no     28 0.9161253 0.7928556 0.9711476
#> 56    yes     28 0.8805867 0.7032664 0.9495878
#> 57     no     29 0.9241097 0.8041735 0.9751174
#> 58    yes     29 0.8915537 0.7173100 0.9559959
#> 59     no     30 0.9313909 0.8150570 0.9785530
#> 60    yes     30 0.9016260 0.7309424 0.9616198
#> 61     no     31 0.9380204 0.8254673 0.9815232
#> 62    yes     31 0.9108563 0.7441511 0.9665501
#> 63     no     32 0.9440478 0.8354102 0.9840887
#> 64    yes     32 0.9192981 0.7569261 0.9709290
#> 65     no     33 0.9495206 0.8448933 0.9863029
#> 66    yes     33 0.9270045 0.7692603 0.9748326
#> 67     no     34 0.9544839 0.8539256 0.9882127
#> 68    yes     34 0.9340278 0.7811488 0.9782842
#> 69     no     35 0.9589802 0.8625175 0.9898589
#> 70    yes     35 0.9404188 0.7925892 0.9812816
#> 71     no     36 0.9630496 0.8706806 0.9912772
#> 72    yes     36 0.9462263 0.8035812 0.9838621
#> 73     no     37 0.9667293 0.8784271 0.9924986
#> 74    yes     37 0.9514970 0.8141265 0.9860870
#> 75     no     38 0.9700539 0.8857703 0.9935501
#> 76    yes     38 0.9562749 0.8242289 0.9880387
#> 77     no     39 0.9730556 0.8927240 0.9944550
#> 78    yes     39 0.9606016 0.8338937 0.9897290
#> 79     no     40 0.9757639 0.8993022 0.9952335
#> 80    yes     40 0.9645161 0.8431275 0.9911925
#> 81     no     41 0.9782061 0.9055195 0.9959020
#> 82    yes     41 0.9680546 0.8519386 0.9924640
#> 83     no     42 0.9804072 0.9113904 0.9964764
#> 84    yes     42 0.9712507 0.8603363 0.9935479
#> 85     no     43 0.9823899 0.9169297 0.9969706
#> 86    yes     43 0.9741357 0.8683307 0.9944667
#> 87     no     44 0.9841753 0.9221521 0.9973956
#> 88    yes     44 0.9767380 0.8759331 0.9952553
#> 89     no     45 0.9857822 0.9270719 0.9977611
#> 90    yes     45 0.9790842 0.8831552 0.9959320
#> 91     no     46 0.9872281 0.9317036 0.9980755
#> 92    yes     46 0.9811982 0.8900092 0.9965125
#> 93     no     47 0.9885287 0.9360611 0.9983457
#> 94    yes     47 0.9831023 0.8965079 0.9970104
#> 95     no     48 0.9896982 0.9401581 0.9985781
#> 96    yes     48 0.9848165 0.9026642 0.9974374
#> 97     no     49 0.9907496 0.9440081 0.9987778
#> 98    yes     49 0.9863593 0.9084914 0.9978035
#> 99     no     50 0.9916946 0.9476239 0.9989495
#> 100   yes     50 0.9877472 0.9140027 0.9981175
#> 101    no     51 0.9925437 0.9510315 0.9990972
#> 102   yes     51 0.9889955 0.9192112 0.9983866
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)