Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3099718 0.4238195
#> 2     yes      1 0.2814584 0.2077055 0.3585329
#> 3      no      2 0.3927672 0.3386071 0.4458903
#> 4     yes      2 0.3039568 0.2315608 0.3715059
#> 5      no      3 0.4189750 0.3700736 0.4680016
#> 6     yes      3 0.3274340 0.2583018 0.3912485
#> 7      no      4 0.4456481 0.4043049 0.4954805
#> 8     yes      4 0.3518080 0.2901617 0.4131944
#> 9      no      5 0.4726379 0.4310126 0.5222270
#> 10    yes      5 0.3769795 0.3189540 0.4358872
#> 11     no      6 0.4997887 0.4526497 0.5519714
#> 12    yes      6 0.4028326 0.3381511 0.4669883
#> 13     no      7 0.5269407 0.4711377 0.5883375
#> 14    yes      7 0.4292372 0.3611287 0.4982350
#> 15     no      8 0.5539343 0.4852007 0.6250652
#> 16    yes      8 0.4560509 0.3877877 0.5353454
#> 17     no      9 0.5806134 0.5044641 0.6637665
#> 18    yes      9 0.4831217 0.4095588 0.5736337
#> 19     no     10 0.6068295 0.5250275 0.7017830
#> 20    yes     10 0.5102919 0.4268614 0.6106491
#> 21     no     11 0.6324443 0.5421159 0.7372002
#> 22    yes     11 0.5374015 0.4435524 0.6464126
#> 23     no     12 0.6573325 0.5585715 0.7697878
#> 24    yes     12 0.5642918 0.4603714 0.6805832
#> 25     no     13 0.6813846 0.5748994 0.7994307
#> 26    yes     13 0.5908091 0.4772804 0.7161369
#> 27     no     14 0.7045073 0.5910654 0.8261159
#> 28    yes     14 0.6168084 0.4921666 0.7492347
#> 29     no     15 0.7266252 0.6070368 0.8499145
#> 30    yes     15 0.6421564 0.5054138 0.7779973
#> 31     no     16 0.7476805 0.6227823 0.8709621
#> 32    yes     16 0.6667336 0.5188777 0.8043156
#> 33     no     17 0.7676327 0.6382728 0.8894398
#> 34    yes     17 0.6904365 0.5328416 0.8281920
#> 35     no     18 0.7864578 0.6534810 0.9055565
#> 36    yes     18 0.7131789 0.5467501 0.8494958
#> 37     no     19 0.8041469 0.6683819 0.9195348
#> 38    yes     19 0.7348919 0.5606731 0.8683292
#> 39     no     20 0.8207047 0.6829528 0.9315991
#> 40    yes     20 0.7555247 0.5771325 0.8855956
#> 41     no     21 0.8361480 0.6971734 0.9419676
#> 42    yes     21 0.7750433 0.5934231 0.9018490
#> 43     no     22 0.8505034 0.7110259 0.9508464
#> 44    yes     22 0.7934296 0.6095116 0.9160125
#> 45     no     23 0.8638059 0.7244951 0.9584259
#> 46    yes     23 0.8106802 0.6253660 0.9282947
#> 47     no     24 0.8760972 0.7375680 0.9648792
#> 48    yes     24 0.8268046 0.6409565 0.9389007
#> 49     no     25 0.8874238 0.7502344 0.9703480
#> 50    yes     25 0.8418238 0.6562553 0.9480258
#> 51     no     26 0.8978357 0.7624862 0.9749023
#> 52    yes     26 0.8557676 0.6712371 0.9558523
#> 53     no     27 0.9073851 0.7743176 0.9787725
#> 54    yes     27 0.8686741 0.6858788 0.9626321
#> 55     no     28 0.9161253 0.7857253 0.9820569
#> 56    yes     28 0.8805867 0.7001600 0.9684847
#> 57     no     29 0.9241097 0.7967075 0.9848411
#> 58    yes     29 0.8915537 0.7140629 0.9734455
#> 59     no     30 0.9313909 0.8072649 0.9871989
#> 60    yes     30 0.9016260 0.7275722 0.9776429
#> 61     no     31 0.9380204 0.8173996 0.9891940
#> 62    yes     31 0.9108563 0.7406751 0.9811893
#> 63     no     32 0.9440478 0.8271155 0.9908811
#> 64    yes     32 0.9192981 0.7533615 0.9841685
#> 65     no     33 0.9495206 0.8364177 0.9923068
#> 66    yes     33 0.9270045 0.7656238 0.9866759
#> 67     no     34 0.9544839 0.8453130 0.9935111
#> 68    yes     34 0.9340278 0.7774564 0.9887904
#> 69     no     35 0.9589802 0.8538090 0.9945279
#> 70    yes     35 0.9404188 0.7888564 0.9905723
#> 71     no     36 0.9630496 0.8619146 0.9953861
#> 72    yes     36 0.9462263 0.7998227 0.9920731
#> 73     no     37 0.9667293 0.8696393 0.9961103
#> 74    yes     37 0.9514970 0.8103563 0.9933364
#> 75     no     38 0.9700539 0.8769935 0.9967211
#> 76    yes     38 0.9562749 0.8204598 0.9943788
#> 77     no     39 0.9730556 0.8839881 0.9972363
#> 78    yes     39 0.9606016 0.8301378 0.9952535
#> 79     no     40 0.9757639 0.8906345 0.9976708
#> 80    yes     40 0.9645161 0.8393960 0.9959927
#> 81     no     41 0.9782061 0.8969445 0.9980370
#> 82    yes     41 0.9680546 0.8482417 0.9966171
#> 83     no     42 0.9804072 0.9029300 0.9983458
#> 84    yes     42 0.9712507 0.8566832 0.9971445
#> 85     no     43 0.9823899 0.9086033 0.9986061
#> 86    yes     43 0.9741357 0.8647401 0.9975899
#> 87     no     44 0.9841753 0.9139766 0.9988255
#> 88    yes     44 0.9767380 0.8724476 0.9979660
#> 89     no     45 0.9857822 0.9190620 0.9990103
#> 90    yes     45 0.9790842 0.8797768 0.9982835
#> 91     no     46 0.9872281 0.9238719 0.9991661
#> 92    yes     46 0.9811982 0.8867393 0.9985515
#> 93     no     47 0.9885287 0.9284181 0.9992974
#> 94    yes     47 0.9831023 0.8933473 0.9987777
#> 95     no     48 0.9896982 0.9327126 0.9994081
#> 96    yes     48 0.9848165 0.8996132 0.9989686
#> 97     no     49 0.9907496 0.9367670 0.9995013
#> 98    yes     49 0.9863593 0.9055497 0.9991297
#> 99     no     50 0.9916946 0.9405761 0.9995798
#> 100   yes     50 0.9877472 0.9111697 0.9992657
#> 101    no     51 0.9925437 0.9441626 0.9996460
#> 102   yes     51 0.9889955 0.9164859 0.9993804
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)