Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2910491 0.4182471
#> 2     yes      1 0.2814584 0.2132587 0.3576173
#> 3      no      2 0.3927672 0.3248151 0.4367396
#> 4     yes      2 0.3039568 0.2374648 0.3756260
#> 5      no      3 0.4189750 0.3605076 0.4599948
#> 6     yes      3 0.3274340 0.2637623 0.3952005
#> 7      no      4 0.4456481 0.3881930 0.4905030
#> 8     yes      4 0.3518080 0.2906857 0.4158121
#> 9      no      5 0.4726379 0.4125202 0.5189238
#> 10    yes      5 0.3769795 0.3178517 0.4409615
#> 11     no      6 0.4997887 0.4372826 0.5504619
#> 12    yes      6 0.4028326 0.3419587 0.4686307
#> 13     no      7 0.5269407 0.4605856 0.5821099
#> 14    yes      7 0.4292372 0.3649647 0.4984614
#> 15     no      8 0.5539343 0.4833069 0.6197818
#> 16    yes      8 0.4560509 0.3855765 0.5313430
#> 17     no      9 0.5806134 0.5057487 0.6575635
#> 18    yes      9 0.4831217 0.4066538 0.5670215
#> 19     no     10 0.6068295 0.5266041 0.6930354
#> 20    yes     10 0.5102919 0.4279304 0.6001722
#> 21     no     11 0.6324443 0.5482135 0.7262051
#> 22    yes     11 0.5374015 0.4451954 0.6387656
#> 23     no     12 0.6573325 0.5637640 0.7554564
#> 24    yes     12 0.5642918 0.4629733 0.6751388
#> 25     no     13 0.6813846 0.5786279 0.7820628
#> 26    yes     13 0.5908091 0.4811612 0.7090856
#> 27     no     14 0.7045073 0.5956953 0.8065156
#> 28    yes     14 0.6168084 0.4988559 0.7400286
#> 29     no     15 0.7266252 0.6128900 0.8288249
#> 30    yes     15 0.6421564 0.5130373 0.7687518
#> 31     no     16 0.7476805 0.6285463 0.8490427
#> 32    yes     16 0.6667336 0.5271987 0.7951781
#> 33     no     17 0.7676327 0.6434225 0.8672544
#> 34    yes     17 0.6904365 0.5413172 0.8192925
#> 35     no     18 0.7864578 0.6580253 0.8835056
#> 36    yes     18 0.7131789 0.5553701 0.8411336
#> 37     no     19 0.8041469 0.6723329 0.8979786
#> 38    yes     19 0.7348919 0.5693350 0.8607819
#> 39     no     20 0.8207047 0.6863254 0.9108342
#> 40    yes     20 0.7555247 0.5831902 0.8783504
#> 41     no     21 0.8361480 0.6999851 0.9224729
#> 42    yes     21 0.7750433 0.5971279 0.8939739
#> 43     no     22 0.8505034 0.7132965 0.9329711
#> 44    yes     22 0.7934296 0.6109512 0.9078004
#> 45     no     23 0.8638059 0.7262465 0.9421363
#> 46    yes     23 0.8106802 0.6245972 0.9199845
#> 47     no     24 0.8760972 0.7388238 0.9500835
#> 48    yes     24 0.8268046 0.6380468 0.9306807
#> 49     no     25 0.8874238 0.7510195 0.9569745
#> 50    yes     25 0.8418238 0.6512818 0.9400397
#> 51     no     26 0.8978357 0.7628268 0.9629508
#> 52    yes     26 0.8557676 0.6642856 0.9482050
#> 53     no     27 0.9073851 0.7742406 0.9681242
#> 54    yes     27 0.8686741 0.6770429 0.9553107
#> 55     no     28 0.9161253 0.7852581 0.9725955
#> 56    yes     28 0.8805867 0.6895397 0.9614809
#> 57     no     29 0.9241097 0.7958780 0.9764544
#> 58    yes     29 0.8915537 0.7029821 0.9668284
#> 59     no     30 0.9313909 0.8061007 0.9797813
#> 60    yes     30 0.9016260 0.7160897 0.9715684
#> 61     no     31 0.9380204 0.8156486 0.9827062
#> 62    yes     31 0.9108563 0.7288420 0.9756571
#> 63     no     32 0.9440478 0.8248125 0.9852143
#> 64    yes     32 0.9192981 0.7412286 0.9791706
#> 65     no     33 0.9495206 0.8336137 0.9873633
#> 66    yes     33 0.9270045 0.7532408 0.9821862
#> 67     no     34 0.9544839 0.8420573 0.9892034
#> 68    yes     34 0.9340278 0.7648720 0.9847720
#> 69     no     35 0.9589802 0.8501493 0.9907780
#> 70    yes     35 0.9404188 0.7761176 0.9869875
#> 71     no     36 0.9630496 0.8578966 0.9921248
#> 72    yes     36 0.9462263 0.7869747 0.9888843
#> 73     no     37 0.9667293 0.8653068 0.9932763
#> 74    yes     37 0.9514970 0.7974422 0.9905072
#> 75     no     38 0.9700539 0.8723880 0.9942604
#> 76    yes     38 0.9562749 0.8075207 0.9918952
#> 77     no     39 0.9730556 0.8791488 0.9951011
#> 78    yes     39 0.9606016 0.8172120 0.9930816
#> 79     no     40 0.9757639 0.8855984 0.9958192
#> 80    yes     40 0.9645161 0.8265730 0.9940954
#> 81     no     41 0.9782061 0.8917461 0.9964324
#> 82    yes     41 0.9680546 0.8355988 0.9949614
#> 83     no     42 0.9804072 0.8976016 0.9969560
#> 84    yes     42 0.9712507 0.8442427 0.9957010
#> 85     no     43 0.9823899 0.9031748 0.9974029
#> 86    yes     43 0.9741357 0.8525116 0.9963323
#> 87     no     44 0.9841753 0.9084755 0.9977844
#> 88    yes     44 0.9767380 0.8604134 0.9968713
#> 89     no     45 0.9857822 0.9135138 0.9981099
#> 90    yes     45 0.9790842 0.8679568 0.9973313
#> 91     no     46 0.9872281 0.9182997 0.9983877
#> 92    yes     46 0.9811982 0.8751509 0.9977238
#> 93     no     47 0.9885287 0.9228431 0.9986247
#> 94    yes     47 0.9831023 0.8820057 0.9980586
#> 95     no     48 0.9896982 0.9271538 0.9988269
#> 96    yes     48 0.9848165 0.8885313 0.9983443
#> 97     no     49 0.9907496 0.9312417 0.9989994
#> 98    yes     49 0.9863593 0.8947385 0.9985880
#> 99     no     50 0.9916946 0.9351162 0.9991466
#> 100   yes     50 0.9877472 0.9006381 0.9987959
#> 101    no     51 0.9925437 0.9387867 0.9992721
#> 102   yes     51 0.9889955 0.9062410 0.9989732
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)