Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3074136 0.4264829
#> 2     yes      1 0.2814584 0.1954920 0.3672133
#> 3      no      2 0.3927672 0.3401436 0.4502458
#> 4     yes      2 0.3039568 0.2195080 0.3851999
#> 5      no      3 0.4189750 0.3694824 0.4740970
#> 6     yes      3 0.3274340 0.2455924 0.4057317
#> 7      no      4 0.4456481 0.4001651 0.4997850
#> 8     yes      4 0.3518080 0.2721168 0.4275000
#> 9      no      5 0.4726379 0.4233148 0.5246407
#> 10    yes      5 0.3769795 0.2969754 0.4508624
#> 11     no      6 0.4997887 0.4471887 0.5535450
#> 12    yes      6 0.4028326 0.3231002 0.4842449
#> 13     no      7 0.5269407 0.4696532 0.5869590
#> 14    yes      7 0.4292372 0.3503813 0.5140424
#> 15     no      8 0.5539343 0.4878960 0.6247307
#> 16    yes      8 0.4560509 0.3760826 0.5459068
#> 17     no      9 0.5806134 0.5052153 0.6616456
#> 18    yes      9 0.4831217 0.3995407 0.5805505
#> 19     no     10 0.6068295 0.5228595 0.6924300
#> 20    yes     10 0.5102919 0.4235190 0.6144186
#> 21     no     11 0.6324443 0.5395516 0.7212103
#> 22    yes     11 0.5374015 0.4478955 0.6471939
#> 23     no     12 0.6573325 0.5537485 0.7482262
#> 24    yes     12 0.5642918 0.4675824 0.6807978
#> 25     no     13 0.6813846 0.5681601 0.7743526
#> 26    yes     13 0.5908091 0.4856817 0.7127058
#> 27     no     14 0.7045073 0.5836430 0.7999318
#> 28    yes     14 0.6168084 0.5038180 0.7426246
#> 29     no     15 0.7266252 0.5989642 0.8232678
#> 30    yes     15 0.6421564 0.5205386 0.7704251
#> 31     no     16 0.7476805 0.6140957 0.8444066
#> 32    yes     16 0.6667336 0.5347877 0.7960408
#> 33     no     17 0.7676327 0.6281740 0.8634325
#> 34    yes     17 0.6904365 0.5489807 0.8194615
#> 35     no     18 0.7864578 0.6418497 0.8804580
#> 36    yes     18 0.7131789 0.5630943 0.8407246
#> 37     no     19 0.8041469 0.6552982 0.8956147
#> 38    yes     19 0.7348919 0.5774059 0.8599055
#> 39     no     20 0.8207047 0.6685015 0.9090458
#> 40    yes     20 0.7555247 0.5916657 0.8771087
#> 41     no     21 0.8361480 0.6814432 0.9208994
#> 42    yes     21 0.7750433 0.6057706 0.8923719
#> 43     no     22 0.8505034 0.6941085 0.9313233
#> 44    yes     22 0.7934296 0.6202165 0.9056746
#> 45     no     23 0.8638059 0.7066295 0.9404609
#> 46    yes     23 0.8106802 0.6348674 0.9174849
#> 47     no     24 0.8760972 0.7190443 0.9484489
#> 48    yes     24 0.8268046 0.6492723 0.9279341
#> 49     no     25 0.8874238 0.7311298 0.9554150
#> 50    yes     25 0.8418238 0.6634097 0.9371507
#> 51     no     26 0.8978357 0.7428769 0.9614773
#> 52    yes     26 0.8557676 0.6772601 0.9452580
#> 53     no     27 0.9073851 0.7542785 0.9667432
#> 54    yes     27 0.8686741 0.6908059 0.9523725
#> 55     no     28 0.9161253 0.7653293 0.9713102
#> 56    yes     28 0.8805867 0.7040311 0.9586029
#> 57     no     29 0.9241097 0.7760254 0.9752656
#> 58    yes     29 0.8915537 0.7169220 0.9640490
#> 59     no     30 0.9313909 0.7863648 0.9786872
#> 60    yes     30 0.9016260 0.7294668 0.9688019
#> 61     no     31 0.9380204 0.7963465 0.9816440
#> 62    yes     31 0.9108563 0.7416555 0.9729440
#> 63     no     32 0.9440478 0.8059714 0.9841970
#> 64    yes     32 0.9192981 0.7531343 0.9766274
#> 65     no     33 0.9495206 0.8151933 0.9863995
#> 66    yes     33 0.9270045 0.7635614 0.9799608
#> 67     no     34 0.9544839 0.8237286 0.9882984
#> 68    yes     34 0.9340278 0.7736798 0.9828266
#> 69     no     35 0.9589802 0.8319484 0.9899348
#> 70    yes     35 0.9404188 0.7834872 0.9852881
#> 71     no     36 0.9630496 0.8398571 0.9913441
#> 72    yes     36 0.9462263 0.7929823 0.9874009
#> 73     no     37 0.9667293 0.8474595 0.9925575
#> 74    yes     37 0.9514970 0.8021652 0.9892131
#> 75     no     38 0.9700539 0.8547612 0.9936017
#> 76    yes     38 0.9562749 0.8110366 0.9907666
#> 77     no     39 0.9730556 0.8617682 0.9945001
#> 78    yes     39 0.9606016 0.8195984 0.9920979
#> 79     no     40 0.9757639 0.8684870 0.9952729
#> 80    yes     40 0.9645161 0.8278987 0.9932382
#> 81     no     41 0.9782061 0.8749247 0.9959375
#> 82    yes     41 0.9680546 0.8359923 0.9942147
#> 83     no     42 0.9804072 0.8810884 0.9965089
#> 84    yes     42 0.9712507 0.8437758 0.9950506
#> 85     no     43 0.9823899 0.8869858 0.9970001
#> 86    yes     43 0.9741357 0.8512541 0.9957661
#> 87     no     44 0.9841753 0.8926246 0.9974223
#> 88    yes     44 0.9767380 0.8584331 0.9963759
#> 89     no     45 0.9857822 0.8980126 0.9977852
#> 90    yes     45 0.9790842 0.8653189 0.9968943
#> 91     no     46 0.9872281 0.9031581 0.9980970
#> 92    yes     46 0.9811982 0.8719183 0.9973386
#> 93     no     47 0.9885287 0.9080689 0.9983650
#> 94    yes     47 0.9831023 0.8782384 0.9977194
#> 95     no     48 0.9896982 0.9127533 0.9985953
#> 96    yes     48 0.9848165 0.8842866 0.9980457
#> 97     no     49 0.9907496 0.9172194 0.9987931
#> 98    yes     49 0.9863593 0.8900705 0.9983254
#> 99     no     50 0.9916946 0.9214753 0.9989631
#> 100   yes     50 0.9877472 0.8955981 0.9985650
#> 101    no     51 0.9925437 0.9255288 0.9991092
#> 102   yes     51 0.9889955 0.9008774 0.9987703
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)