Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2915632 0.4262463
#> 2     yes      1 0.2814584 0.2182273 0.3625546
#> 3      no      2 0.3927672 0.3267863 0.4504323
#> 4     yes      2 0.3039568 0.2436542 0.3826558
#> 5      no      3 0.4189750 0.3624975 0.4752773
#> 6     yes      3 0.3274340 0.2695480 0.4030919
#> 7      no      4 0.4456481 0.3900492 0.4994033
#> 8     yes      4 0.3518080 0.2943850 0.4250274
#> 9      no      5 0.4726379 0.4122107 0.5296660
#> 10    yes      5 0.3769795 0.3180996 0.4463721
#> 11     no      6 0.4997887 0.4349702 0.5594403
#> 12    yes      6 0.4028326 0.3491814 0.4750187
#> 13     no      7 0.5269407 0.4608453 0.5947124
#> 14    yes      7 0.4292372 0.3748843 0.5100485
#> 15     no      8 0.5539343 0.4878698 0.6325166
#> 16    yes      8 0.4560509 0.3949407 0.5418702
#> 17     no      9 0.5806134 0.5106181 0.6687440
#> 18    yes      9 0.4831217 0.4175902 0.5726435
#> 19     no     10 0.6068295 0.5317632 0.7053327
#> 20    yes     10 0.5102919 0.4428268 0.6086718
#> 21     no     11 0.6324443 0.5498592 0.7398446
#> 22    yes     11 0.5374015 0.4600181 0.6460426
#> 23     no     12 0.6573325 0.5655952 0.7716258
#> 24    yes     12 0.5642918 0.4769776 0.6817106
#> 25     no     13 0.6813846 0.5818769 0.8005730
#> 26    yes     13 0.5908091 0.4939896 0.7153085
#> 27     no     14 0.7045073 0.5999660 0.8266768
#> 28    yes     14 0.6168084 0.5098090 0.7461351
#> 29     no     15 0.7266252 0.6174959 0.8500049
#> 30    yes     15 0.6421564 0.5239179 0.7746612
#> 31     no     16 0.7476805 0.6344213 0.8706850
#> 32    yes     16 0.6667336 0.5379892 0.8008218
#> 33     no     17 0.7676327 0.6499259 0.8889037
#> 34    yes     17 0.6904365 0.5520653 0.8265099
#> 35     no     18 0.7864578 0.6651199 0.9053131
#> 36    yes     18 0.7131789 0.5673882 0.8498207
#> 37     no     19 0.8041469 0.6799788 0.9195159
#> 38    yes     19 0.7348919 0.5834501 0.8704909
#> 39     no     20 0.8207047 0.6944811 0.9317466
#> 40    yes     20 0.7555247 0.5995002 0.8886891
#> 41     no     21 0.8361480 0.7086076 0.9422332
#> 42    yes     21 0.7750433 0.6153392 0.9046107
#> 43     no     22 0.8505034 0.7223417 0.9511043
#> 44    yes     22 0.7934296 0.6309370 0.9184639
#> 45     no     23 0.8638059 0.7356697 0.9586688
#> 46    yes     23 0.8106802 0.6462651 0.9304600
#> 47     no     24 0.8760972 0.7485803 0.9651043
#> 48    yes     24 0.8268046 0.6612977 0.9408048
#> 49     no     25 0.8874238 0.7610648 0.9705671
#> 50    yes     25 0.8418238 0.6760108 0.9496939
#> 51     no     26 0.8978357 0.7731170 0.9751955
#> 52    yes     26 0.8557676 0.6903831 0.9573087
#> 53     no     27 0.9073851 0.7847329 0.9791107
#> 54    yes     27 0.8686741 0.7043956 0.9638146
#> 55     no     28 0.9161253 0.7959107 0.9824181
#> 56    yes     28 0.8805867 0.7180319 0.9694921
#> 57     no     29 0.9241097 0.8066509 0.9852091
#> 58    yes     29 0.8915537 0.7312779 0.9743396
#> 59     no     30 0.9313909 0.8169555 0.9875620
#> 60    yes     30 0.9016260 0.7441224 0.9784083
#> 61     no     31 0.9380204 0.8268285 0.9895684
#> 62    yes     31 0.9108563 0.7565560 0.9818296
#> 63     no     32 0.9440478 0.8362755 0.9912564
#> 64    yes     32 0.9192981 0.7685723 0.9847166
#> 65     no     33 0.9495206 0.8453032 0.9926515
#> 66    yes     33 0.9270045 0.7801665 0.9871502
#> 67     no     34 0.9544839 0.8539199 0.9938183
#> 68    yes     34 0.9340278 0.7913364 0.9892000
#> 69     no     35 0.9589802 0.8619454 0.9948008
#> 70    yes     35 0.9404188 0.8020816 0.9909254
#> 71     no     36 0.9630496 0.8695097 0.9956279
#> 72    yes     36 0.9462263 0.8124035 0.9923942
#> 73     no     37 0.9667293 0.8767188 0.9963238
#> 74    yes     37 0.9514970 0.8223052 0.9936268
#> 75     no     38 0.9700539 0.8835830 0.9969094
#> 76    yes     38 0.9562749 0.8317914 0.9946606
#> 77     no     39 0.9730556 0.8901129 0.9974019
#> 78    yes     39 0.9606016 0.8408681 0.9955274
#> 79     no     40 0.9757639 0.8963196 0.9978161
#> 80    yes     40 0.9645161 0.8495426 0.9962539
#> 81     no     41 0.9782061 0.9022142 0.9981644
#> 82    yes     41 0.9680546 0.8578233 0.9968626
#> 83     no     42 0.9804072 0.9078081 0.9984572
#> 84    yes     42 0.9712507 0.8657193 0.9973727
#> 85     no     43 0.9823899 0.9131129 0.9987034
#> 86    yes     43 0.9741357 0.8732407 0.9977999
#> 87     no     44 0.9841753 0.9181322 0.9989103
#> 88    yes     44 0.9767380 0.8803981 0.9981578
#> 89     no     45 0.9857822 0.9228590 0.9990843
#> 90    yes     45 0.9790842 0.8871444 0.9984574
#> 91     no     46 0.9872281 0.9273345 0.9992305
#> 92    yes     46 0.9811982 0.8934735 0.9987084
#> 93     no     47 0.9885287 0.9315695 0.9993533
#> 94    yes     47 0.9831023 0.8994878 0.9989186
#> 95     no     48 0.9896982 0.9355748 0.9994566
#> 96    yes     48 0.9848165 0.9051983 0.9990945
#> 97     no     49 0.9907496 0.9393609 0.9995434
#> 98    yes     49 0.9863593 0.9106165 0.9992419
#> 99     no     50 0.9916946 0.9429380 0.9996163
#> 100   yes     50 0.9877472 0.9157538 0.9993652
#> 101    no     51 0.9925437 0.9463161 0.9996776
#> 102   yes     51 0.9889955 0.9206214 0.9994685
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)