Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3209072 0.4250076
#> 2     yes      1 0.2814584 0.2126660 0.3597851
#> 3      no      2 0.3927672 0.3458063 0.4396687
#> 4     yes      2 0.3039568 0.2374035 0.3812496
#> 5      no      3 0.4189750 0.3767491 0.4630535
#> 6     yes      3 0.3274340 0.2640574 0.3991605
#> 7      no      4 0.4456481 0.3996649 0.4874997
#> 8     yes      4 0.3518080 0.2907377 0.4193869
#> 9      no      5 0.4726379 0.4217631 0.5226490
#> 10    yes      5 0.3769795 0.3146977 0.4397964
#> 11     no      6 0.4997887 0.4441803 0.5578839
#> 12    yes      6 0.4028326 0.3433104 0.4628974
#> 13     no      7 0.5269407 0.4668246 0.5956303
#> 14    yes      7 0.4292372 0.3730521 0.5005138
#> 15     no      8 0.5539343 0.4874916 0.6340703
#> 16    yes      8 0.4560509 0.3962075 0.5402990
#> 17     no      9 0.5806134 0.5069157 0.6708741
#> 18    yes      9 0.4831217 0.4136071 0.5840548
#> 19     no     10 0.6068295 0.5247771 0.7056933
#> 20    yes     10 0.5102919 0.4284712 0.6257799
#> 21     no     11 0.6324443 0.5425784 0.7382649
#> 22    yes     11 0.5374015 0.4434638 0.6608775
#> 23     no     12 0.6573325 0.5603325 0.7684139
#> 24    yes     12 0.5642918 0.4613389 0.6922026
#> 25     no     13 0.6813846 0.5786781 0.7960488
#> 26    yes     13 0.5908091 0.4797394 0.7218180
#> 27     no     14 0.7045073 0.5972476 0.8211529
#> 28    yes     14 0.6168084 0.4982002 0.7521598
#> 29     no     15 0.7266252 0.6159431 0.8437730
#> 30    yes     15 0.6421564 0.5166684 0.7804600
#> 31     no     16 0.7476805 0.6316228 0.8640056
#> 32    yes     16 0.6667336 0.5354595 0.8063550
#> 33     no     17 0.7676327 0.6474890 0.8819841
#> 34    yes     17 0.6904365 0.5544230 0.8298565
#> 35     no     18 0.7864578 0.6623411 0.8978666
#> 36    yes     18 0.7131789 0.5732243 0.8510281
#> 37     no     19 0.8041469 0.6774420 0.9118251
#> 38    yes     19 0.7348919 0.5918089 0.8699741
#> 39     no     20 0.8207047 0.6921802 0.9240372
#> 40    yes     20 0.7555247 0.6101250 0.8868276
#> 41     no     21 0.8361480 0.7065355 0.9346790
#> 42    yes     21 0.7750433 0.6281239 0.9017404
#> 43     no     22 0.8505034 0.7204904 0.9439203
#> 44    yes     22 0.7934296 0.6457608 0.9148742
#> 45     no     23 0.8638059 0.7340302 0.9519214
#> 46    yes     23 0.8106802 0.6620911 0.9263938
#> 47     no     24 0.8760972 0.7469857 0.9588307
#> 48    yes     24 0.8268046 0.6774808 0.9364609
#> 49     no     25 0.8874238 0.7588003 0.9647837
#> 50    yes     25 0.8418238 0.6924909 0.9452312
#> 51     no     26 0.8978357 0.7702331 0.9699029
#> 52    yes     26 0.8557676 0.7071003 0.9528505
#> 53     no     27 0.9073851 0.7812805 0.9742977
#> 54    yes     27 0.8686741 0.7212909 0.9594825
#> 55     no     28 0.9161253 0.7919404 0.9780653
#> 56    yes     28 0.8805867 0.7350475 0.9653596
#> 57     no     29 0.9241097 0.8022123 0.9812912
#> 58    yes     29 0.8915537 0.7483581 0.9704108
#> 59     no     30 0.9313909 0.8120976 0.9840503
#> 60    yes     30 0.9016260 0.7612133 0.9747448
#> 61     no     31 0.9380204 0.8215986 0.9864082
#> 62    yes     31 0.9108563 0.7729214 0.9784582
#> 63     no     32 0.9440478 0.8307194 0.9884216
#> 64    yes     32 0.9192981 0.7838713 0.9816359
#> 65     no     33 0.9495206 0.8394649 0.9901397
#> 66    yes     33 0.9270045 0.7944351 0.9843525
#> 67     no     34 0.9544839 0.8478413 0.9916050
#> 68    yes     34 0.9340278 0.8046125 0.9866726
#> 69     no     35 0.9589802 0.8558555 0.9928541
#> 70    yes     35 0.9404188 0.8144050 0.9886528
#> 71     no     36 0.9630496 0.8635155 0.9939185
#> 72    yes     36 0.9462263 0.8238151 0.9903416
#> 73     no     37 0.9667293 0.8708297 0.9948251
#> 74    yes     37 0.9514970 0.8328470 0.9917812
#> 75     no     38 0.9700539 0.8778072 0.9955972
#> 76    yes     38 0.9562749 0.8415058 0.9930077
#> 77     no     39 0.9730556 0.8844576 0.9962545
#> 78    yes     39 0.9606016 0.8497976 0.9940523
#> 79     no     40 0.9757639 0.8907910 0.9968140
#> 80    yes     40 0.9645161 0.8577296 0.9949417
#> 81     no     41 0.9782061 0.8968175 0.9972901
#> 82    yes     41 0.9680546 0.8653096 0.9956986
#> 83     no     42 0.9804072 0.9025476 0.9976953
#> 84    yes     42 0.9712507 0.8725461 0.9963427
#> 85     no     43 0.9823899 0.9079919 0.9980400
#> 86    yes     43 0.9741357 0.8794485 0.9968906
#> 87     no     44 0.9841753 0.9131612 0.9983332
#> 88    yes     44 0.9767380 0.8860261 0.9973567
#> 89     no     45 0.9857822 0.9180732 0.9985826
#> 90    yes     45 0.9790842 0.8922891 0.9977531
#> 91     no     46 0.9872281 0.9227576 0.9987947
#> 92    yes     46 0.9811982 0.8982476 0.9980901
#> 93     no     47 0.9885287 0.9271953 0.9989751
#> 94    yes     47 0.9831023 0.9039122 0.9983767
#> 95     no     48 0.9896982 0.9313968 0.9991286
#> 96    yes     48 0.9848165 0.9092934 0.9986203
#> 97     no     49 0.9907496 0.9353726 0.9992590
#> 98    yes     49 0.9863593 0.9144018 0.9988274
#> 99     no     50 0.9916946 0.9391328 0.9993700
#> 100   yes     50 0.9877472 0.9192482 0.9990035
#> 101    no     51 0.9925437 0.9426875 0.9994643
#> 102   yes     51 0.9889955 0.9238430 0.9991531
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)