Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3070739 0.4246884
#> 2     yes      1 0.2814584 0.2162693 0.3578153
#> 3      no      2 0.3927672 0.3362416 0.4439118
#> 4     yes      2 0.3039568 0.2378032 0.3770172
#> 5      no      3 0.4189750 0.3604527 0.4629694
#> 6     yes      3 0.3274340 0.2577991 0.3922844
#> 7      no      4 0.4456481 0.3870074 0.4887056
#> 8     yes      4 0.3518080 0.2830670 0.4139784
#> 9      no      5 0.4726379 0.4123528 0.5233846
#> 10    yes      5 0.3769795 0.3097820 0.4404920
#> 11     no      6 0.4997887 0.4344797 0.5538128
#> 12    yes      6 0.4028326 0.3378337 0.4741397
#> 13     no      7 0.5269407 0.4565885 0.5863444
#> 14    yes      7 0.4292372 0.3641368 0.5080028
#> 15     no      8 0.5539343 0.4793774 0.6214461
#> 16    yes      8 0.4560509 0.3887773 0.5472429
#> 17     no      9 0.5806134 0.5007936 0.6553176
#> 18    yes      9 0.4831217 0.4107850 0.5814551
#> 19     no     10 0.6068295 0.5194301 0.6876785
#> 20    yes     10 0.5102919 0.4293880 0.6148722
#> 21     no     11 0.6324443 0.5391764 0.7191470
#> 22    yes     11 0.5374015 0.4481956 0.6472446
#> 23     no     12 0.6573325 0.5606274 0.7487483
#> 24    yes     12 0.5642918 0.4671553 0.6803783
#> 25     no     13 0.6813846 0.5818887 0.7768604
#> 26    yes     13 0.5908091 0.4862126 0.7129000
#> 27     no     14 0.7045073 0.5994591 0.8031310
#> 28    yes     14 0.6168084 0.5053120 0.7455647
#> 29     no     15 0.7266252 0.6162307 0.8269957
#> 30    yes     15 0.6421564 0.5243973 0.7762510
#> 31     no     16 0.7476805 0.6331518 0.8485028
#> 32    yes     16 0.6667336 0.5434126 0.8042051
#> 33     no     17 0.7676327 0.6497426 0.8675997
#> 34    yes     17 0.6904365 0.5623029 0.8294312
#> 35     no     18 0.7864578 0.6659686 0.8848371
#> 36    yes     18 0.7131789 0.5810144 0.8520028
#> 37     no     19 0.8041469 0.6817989 0.9005702
#> 38    yes     19 0.7348919 0.5994955 0.8720459
#> 39     no     20 0.8207047 0.6978060 0.9143619
#> 40    yes     20 0.7555247 0.6173441 0.8897239
#> 41     no     21 0.8361480 0.7131368 0.9263968
#> 42    yes     21 0.7750433 0.6340930 0.9052234
#> 43     no     22 0.8505034 0.7275924 0.9368573
#> 44    yes     22 0.7934296 0.6505760 0.9187419
#> 45     no     23 0.8638059 0.7415822 0.9459178
#> 46    yes     23 0.8106802 0.6667065 0.9304791
#> 47     no     24 0.8760972 0.7550941 0.9537424
#> 48    yes     24 0.8268046 0.6824555 0.9406294
#> 49     no     25 0.8874238 0.7681193 0.9604821
#> 50    yes     25 0.8418238 0.6977975 0.9493776
#> 51     no     26 0.8978357 0.7806519 0.9662746
#> 52    yes     26 0.8557676 0.7127100 0.9568952
#> 53     no     27 0.9073851 0.7926888 0.9712379
#> 54    yes     27 0.8686741 0.7271739 0.9633389
#> 55     no     28 0.9161253 0.8042299 0.9754869
#> 56    yes     28 0.8805867 0.7400776 0.9688502
#> 57     no     29 0.9241097 0.8152773 0.9791216
#> 58    yes     29 0.8915537 0.7523270 0.9735554
#> 59     no     30 0.9313909 0.8258354 0.9822271
#> 60    yes     30 0.9016260 0.7643439 0.9775659
#> 61     no     31 0.9380204 0.8359106 0.9848779
#> 62    yes     31 0.9108563 0.7774557 0.9809799
#> 63     no     32 0.9440478 0.8454402 0.9871384
#> 64    yes     32 0.9192981 0.7905833 0.9838826
#> 65     no     33 0.9495206 0.8543429 0.9890648
#> 66    yes     33 0.9270045 0.8031282 0.9863483
#> 67     no     34 0.9544839 0.8628162 0.9907054
#> 68    yes     34 0.9340278 0.8147019 0.9884411
#> 69     no     35 0.9589802 0.8708714 0.9921017
#> 70    yes     35 0.9404188 0.8252382 0.9902161
#> 71     no     36 0.9630496 0.8785203 0.9932898
#> 72    yes     36 0.9462263 0.8352963 0.9917073
#> 73     no     37 0.9667293 0.8857758 0.9943001
#> 74    yes     37 0.9514970 0.8448841 0.9929647
#> 75     no     38 0.9700539 0.8926510 0.9951590
#> 76    yes     38 0.9562749 0.8540110 0.9940326
#> 77     no     39 0.9730556 0.8991597 0.9958891
#> 78    yes     39 0.9606016 0.8626881 0.9949393
#> 79     no     40 0.9757639 0.9053158 0.9965094
#> 80    yes     40 0.9645161 0.8709271 0.9957088
#> 81     no     41 0.9782061 0.9111333 0.9970364
#> 82    yes     41 0.9680546 0.8787411 0.9963617
#> 83     no     42 0.9804072 0.9166264 0.9974840
#> 84    yes     42 0.9712507 0.8861437 0.9969156
#> 85     no     43 0.9823899 0.9218092 0.9978642
#> 86    yes     43 0.9741357 0.8931492 0.9973854
#> 87     no     44 0.9841753 0.9266957 0.9981870
#> 88    yes     44 0.9767380 0.8997722 0.9977838
#> 89     no     45 0.9857822 0.9312997 0.9984611
#> 90    yes     45 0.9790842 0.9060277 0.9981216
#> 91     no     46 0.9872281 0.9356347 0.9986938
#> 92    yes     46 0.9811982 0.9119309 0.9984080
#> 93     no     47 0.9885287 0.9397139 0.9988914
#> 94    yes     47 0.9831023 0.9174970 0.9986508
#> 95     no     48 0.9896982 0.9435502 0.9990591
#> 96    yes     48 0.9848165 0.9227409 0.9988566
#> 97     no     49 0.9907496 0.9471563 0.9992014
#> 98    yes     49 0.9863593 0.9276776 0.9990310
#> 99     no     50 0.9916946 0.9505440 0.9993223
#> 100   yes     50 0.9877472 0.9323220 0.9991789
#> 101    no     51 0.9925437 0.9537253 0.9994248
#> 102   yes     51 0.9889955 0.9366884 0.9993042
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)