Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3011471 0.4215386
#> 2     yes      1 0.2814584 0.2134749 0.3374510
#> 3      no      2 0.3927672 0.3333803 0.4402393
#> 4     yes      2 0.3039568 0.2414031 0.3573546
#> 5      no      3 0.4189750 0.3593946 0.4639963
#> 6     yes      3 0.3274340 0.2704906 0.3781143
#> 7      no      4 0.4456481 0.3878281 0.4932625
#> 8     yes      4 0.3518080 0.2944568 0.4044054
#> 9      no      5 0.4726379 0.4142093 0.5242981
#> 10    yes      5 0.3769795 0.3179464 0.4310097
#> 11     no      6 0.4997887 0.4400516 0.5570555
#> 12    yes      6 0.4028326 0.3414764 0.4574031
#> 13     no      7 0.5269407 0.4662232 0.5917553
#> 14    yes      7 0.4292372 0.3624226 0.4865939
#> 15     no      8 0.5539343 0.4896026 0.6249841
#> 16    yes      8 0.4560509 0.3819293 0.5198296
#> 17     no      9 0.5806134 0.5105834 0.6570466
#> 18    yes      9 0.4831217 0.4018204 0.5563107
#> 19     no     10 0.6068295 0.5302508 0.6876883
#> 20    yes     10 0.5102919 0.4207708 0.5923867
#> 21     no     11 0.6324443 0.5524678 0.7189368
#> 22    yes     11 0.5374015 0.4379047 0.6274912
#> 23     no     12 0.6573325 0.5706259 0.7478732
#> 24    yes     12 0.5642918 0.4551954 0.6612980
#> 25     no     13 0.6813846 0.5885962 0.7755799
#> 26    yes     13 0.5908091 0.4726011 0.6935315
#> 27     no     14 0.7045073 0.6063335 0.8014243
#> 28    yes     14 0.6168084 0.4900787 0.7248404
#> 29     no     15 0.7266252 0.6237949 0.8249654
#> 30    yes     15 0.6421564 0.5075847 0.7548464
#> 31     no     16 0.7476805 0.6409406 0.8462519
#> 32    yes     16 0.6667336 0.5250749 0.7825625
#> 33     no     17 0.7676327 0.6577336 0.8653727
#> 34    yes     17 0.6904365 0.5425053 0.8079426
#> 35     no     18 0.7864578 0.6741404 0.8824463
#> 36    yes     18 0.7131789 0.5598326 0.8309050
#> 37     no     19 0.8041469 0.6901312 0.8976110
#> 38    yes     19 0.7348919 0.5771808 0.8510801
#> 39     no     20 0.8207047 0.7056798 0.9110168
#> 40    yes     20 0.7555247 0.5946159 0.8690653
#> 41     no     21 0.8361480 0.7208117 0.9228183
#> 42    yes     21 0.7750433 0.6118125 0.8850723
#> 43     no     22 0.8505034 0.7354893 0.9331694
#> 44    yes     22 0.7934296 0.6287310 0.8998931
#> 45     no     23 0.8638059 0.7496630 0.9422192
#> 46    yes     23 0.8106802 0.6463528 0.9132147
#> 47     no     24 0.8760972 0.7633214 0.9501089
#> 48    yes     24 0.8268046 0.6638699 0.9248117
#> 49     no     25 0.8874238 0.7764565 0.9569704
#> 50    yes     25 0.8418238 0.6809476 0.9350853
#> 51     no     26 0.8978357 0.7887221 0.9629250
#> 52    yes     26 0.8557676 0.6975529 0.9443009
#> 53     no     27 0.9073851 0.7998839 0.9680830
#> 54    yes     27 0.8686741 0.7136573 0.9521671
#> 55     no     28 0.9161253 0.8106235 0.9725438
#> 56    yes     28 0.8805867 0.7292370 0.9588622
#> 57     no     29 0.9241097 0.8210321 0.9763962
#> 58    yes     29 0.8915537 0.7442728 0.9646024
#> 59     no     30 0.9313909 0.8309866 0.9797193
#> 60    yes     30 0.9016260 0.7587498 0.9696314
#> 61     no     31 0.9380204 0.8404937 0.9825828
#> 62    yes     31 0.9108563 0.7717945 0.9739652
#> 63     no     32 0.9440478 0.8495615 0.9850482
#> 64    yes     32 0.9192981 0.7839644 0.9776948
#> 65     no     33 0.9495206 0.8581995 0.9871685
#> 66    yes     33 0.9270045 0.7956547 0.9809005
#> 67     no     34 0.9544839 0.8664181 0.9890105
#> 68    yes     34 0.9340278 0.8068654 0.9836533
#> 69     no     35 0.9589802 0.8742288 0.9906125
#> 70    yes     35 0.9404188 0.8176540 0.9860149
#> 71     no     36 0.9630496 0.8816440 0.9919824
#> 72    yes     36 0.9462263 0.8281425 0.9880395
#> 73     no     37 0.9667293 0.8886764 0.9931537
#> 74    yes     37 0.9514970 0.8381436 0.9897741
#> 75     no     38 0.9700539 0.8953393 0.9941549
#> 76    yes     38 0.9562749 0.8475180 0.9912593
#> 77     no     39 0.9730556 0.9016464 0.9950103
#> 78    yes     39 0.9606016 0.8560615 0.9925305
#> 79     no     40 0.9757639 0.9076115 0.9957411
#> 80    yes     40 0.9645161 0.8642025 0.9936179
#> 81     no     41 0.9782061 0.9132485 0.9963652
#> 82    yes     41 0.9680546 0.8719513 0.9945479
#> 83     no     42 0.9804072 0.9185714 0.9968981
#> 84    yes     42 0.9712507 0.8793191 0.9953431
#> 85     no     43 0.9823899 0.9235940 0.9973531
#> 86    yes     43 0.9741357 0.8863178 0.9960227
#> 87     no     44 0.9841753 0.9283300 0.9977415
#> 88    yes     44 0.9767380 0.8929594 0.9966035
#> 89     no     45 0.9857822 0.9327930 0.9980730
#> 90    yes     45 0.9790842 0.8992567 0.9970997
#> 91     no     46 0.9872281 0.9369960 0.9983559
#> 92    yes     46 0.9811982 0.9052223 0.9975236
#> 93     no     47 0.9885287 0.9409521 0.9985973
#> 94    yes     47 0.9831023 0.9108692 0.9978857
#> 95     no     48 0.9896982 0.9446736 0.9988033
#> 96    yes     48 0.9848165 0.9162104 0.9981949
#> 97     no     49 0.9907496 0.9481728 0.9989791
#> 98    yes     49 0.9863593 0.9212588 0.9984590
#> 99     no     50 0.9916946 0.9514614 0.9991290
#> 100   yes     50 0.9877472 0.9260272 0.9986870
#> 101    no     51 0.9925437 0.9544716 0.9992571
#> 102   yes     51 0.9889955 0.9305282 0.9988815
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)