Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2944385 0.4157960
#> 2     yes      1 0.2814584 0.2191440 0.3395261
#> 3      no      2 0.3927672 0.3234532 0.4423727
#> 4     yes      2 0.3039568 0.2416312 0.3601828
#> 5      no      3 0.4189750 0.3539000 0.4727255
#> 6     yes      3 0.3274340 0.2656450 0.3828915
#> 7      no      4 0.4456481 0.3855848 0.4973630
#> 8     yes      4 0.3518080 0.2861516 0.4103462
#> 9      no      5 0.4726379 0.4183191 0.5207699
#> 10    yes      5 0.3769795 0.3071089 0.4409767
#> 11     no      6 0.4997887 0.4514947 0.5517731
#> 12    yes      6 0.4028326 0.3296527 0.4692628
#> 13     no      7 0.5269407 0.4732114 0.5878130
#> 14    yes      7 0.4292372 0.3531430 0.4986148
#> 15     no      8 0.5539343 0.4924908 0.6249561
#> 16    yes      8 0.4560509 0.3712818 0.5310515
#> 17     no      9 0.5806134 0.5092758 0.6606844
#> 18    yes      9 0.4831217 0.3889692 0.5659193
#> 19     no     10 0.6068295 0.5260392 0.6946718
#> 20    yes     10 0.5102919 0.4072083 0.6001475
#> 21     no     11 0.6324443 0.5427432 0.7266634
#> 22    yes     11 0.5374015 0.4271524 0.6334259
#> 23     no     12 0.6573325 0.5593506 0.7551544
#> 24    yes     12 0.5642918 0.4486429 0.6654794
#> 25     no     13 0.6813846 0.5758252 0.7828545
#> 26    yes     13 0.5908091 0.4698907 0.6960757
#> 27     no     14 0.7045073 0.5921318 0.8083561
#> 28    yes     14 0.6168084 0.4898100 0.7250307
#> 29     no     15 0.7266252 0.6082368 0.8315074
#> 30    yes     15 0.6421564 0.5097579 0.7522099
#> 31     no     16 0.7476805 0.6241081 0.8523726
#> 32    yes     16 0.6667336 0.5286849 0.7775275
#> 33     no     17 0.7676327 0.6397159 0.8710537
#> 34    yes     17 0.6904365 0.5460884 0.8016405
#> 35     no     18 0.7864578 0.6550325 0.8876725
#> 36    yes     18 0.7131789 0.5629314 0.8244616
#> 37     no     19 0.8041469 0.6700322 0.9023265
#> 38    yes     19 0.7348919 0.5796311 0.8451594
#> 39     no     20 0.8207047 0.6846922 0.9152134
#> 40    yes     20 0.7555247 0.5961510 0.8637726
#> 41     no     21 0.8361480 0.6989919 0.9265385
#> 42    yes     21 0.7750433 0.6124562 0.8806367
#> 43     no     22 0.8505034 0.7129135 0.9364558
#> 44    yes     22 0.7934296 0.6285139 0.8961512
#> 45     no     23 0.8638059 0.7264417 0.9451137
#> 46    yes     23 0.8106802 0.6442931 0.9098256
#> 47     no     24 0.8760972 0.7395638 0.9526515
#> 48    yes     24 0.8268046 0.6597654 0.9213522
#> 49     no     25 0.8874238 0.7522696 0.9591989
#> 50    yes     25 0.8418238 0.6749046 0.9314041
#> 51     no     26 0.8978357 0.7645515 0.9648742
#> 52    yes     26 0.8557676 0.6896874 0.9404315
#> 53     no     27 0.9073851 0.7764041 0.9697850
#> 54    yes     27 0.8686741 0.7040931 0.9486303
#> 55     no     28 0.9161253 0.7878242 0.9740277
#> 56    yes     28 0.8805867 0.7181037 0.9557538
#> 57     no     29 0.9241097 0.7988110 0.9776884
#> 58    yes     29 0.8915537 0.7317040 0.9619291
#> 59     no     30 0.9313909 0.8093912 0.9808432
#> 60    yes     30 0.9016260 0.7448814 0.9672720
#> 61     no     31 0.9380204 0.8195697 0.9835595
#> 62    yes     31 0.9108563 0.7576263 0.9718870
#> 63     no     32 0.9440478 0.8293184 0.9858962
#> 64    yes     32 0.9192981 0.7699314 0.9758674
#> 65     no     33 0.9495206 0.8386433 0.9879048
#> 66    yes     33 0.9270045 0.7817918 0.9792963
#> 67     no     34 0.9544839 0.8475517 0.9896304
#> 68    yes     34 0.9340278 0.7932052 0.9822468
#> 69     no     35 0.9589802 0.8560521 0.9911120
#> 70    yes     35 0.9404188 0.8041712 0.9847834
#> 71     no     36 0.9630496 0.8641538 0.9923835
#> 72    yes     36 0.9462263 0.8144650 0.9869623
#> 73     no     37 0.9667293 0.8718673 0.9934744
#> 74    yes     37 0.9514970 0.8243120 0.9888328
#> 75     no     38 0.9700539 0.8792034 0.9944099
#> 76    yes     38 0.9562749 0.8337427 0.9904375
#> 77     no     39 0.9730556 0.8861739 0.9952119
#> 78    yes     39 0.9606016 0.8427636 0.9918135
#> 79     no     40 0.9757639 0.8927907 0.9958993
#> 80    yes     40 0.9645161 0.8513819 0.9929915
#> 81     no     41 0.9782061 0.8990662 0.9964884
#> 82    yes     41 0.9680546 0.8596062 0.9940009
#> 83     no     42 0.9804072 0.9050132 0.9969931
#> 84    yes     42 0.9712507 0.8674459 0.9948657
#> 85     no     43 0.9823899 0.9106441 0.9974255
#> 86    yes     43 0.9741357 0.8749112 0.9956063
#> 87     no     44 0.9841753 0.9159720 0.9977958
#> 88    yes     44 0.9767380 0.8820128 0.9962406
#> 89     no     45 0.9857822 0.9210093 0.9981130
#> 90    yes     45 0.9790842 0.8887621 0.9967835
#> 91     no     46 0.9872281 0.9257689 0.9983846
#> 92    yes     46 0.9811982 0.8952090 0.9972483
#> 93     no     47 0.9885287 0.9302632 0.9986171
#> 94    yes     47 0.9831023 0.9013429 0.9976460
#> 95     no     48 0.9896982 0.9345043 0.9988162
#> 96    yes     48 0.9848165 0.9071547 0.9979864
#> 97     no     49 0.9907496 0.9385044 0.9989867
#> 98    yes     49 0.9863593 0.9126568 0.9982777
#> 99     no     50 0.9916946 0.9422750 0.9991331
#> 100   yes     50 0.9877472 0.9178621 0.9985268
#> 101    no     51 0.9925437 0.9458275 0.9992588
#> 102   yes     51 0.9889955 0.9227831 0.9987400
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)