Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3188526 0.4232910
#> 2     yes      1 0.2814584 0.2267852 0.3459930
#> 3      no      2 0.3927672 0.3500455 0.4417201
#> 4     yes      2 0.3039568 0.2498517 0.3619340
#> 5      no      3 0.4189750 0.3795690 0.4654615
#> 6     yes      3 0.3274340 0.2735636 0.3840373
#> 7      no      4 0.4456481 0.4046963 0.4974262
#> 8     yes      4 0.3518080 0.2965182 0.4121898
#> 9      no      5 0.4726379 0.4283702 0.5209690
#> 10    yes      5 0.3769795 0.3203096 0.4333754
#> 11     no      6 0.4997887 0.4548286 0.5515060
#> 12    yes      6 0.4028326 0.3414222 0.4604498
#> 13     no      7 0.5269407 0.4782848 0.5850769
#> 14    yes      7 0.4292372 0.3593200 0.4903162
#> 15     no      8 0.5539343 0.4996169 0.6216801
#> 16    yes      8 0.4560509 0.3776242 0.5260616
#> 17     no      9 0.5806134 0.5209793 0.6569475
#> 18    yes      9 0.4831217 0.3962892 0.5625698
#> 19     no     10 0.6068295 0.5428532 0.6905635
#> 20    yes     10 0.5102919 0.4152651 0.5982550
#> 21     no     11 0.6324443 0.5608110 0.7209318
#> 22    yes     11 0.5374015 0.4347110 0.6333863
#> 23     no     12 0.6573325 0.5782156 0.7489031
#> 24    yes     12 0.5642918 0.4546154 0.6665836
#> 25     no     13 0.6813846 0.5952878 0.7749400
#> 26    yes     13 0.5908091 0.4746657 0.6975066
#> 27     no     14 0.7045073 0.6111669 0.7989950
#> 28    yes     14 0.6168084 0.4928781 0.7264809
#> 29     no     15 0.7266252 0.6268135 0.8216245
#> 30    yes     15 0.6421564 0.5101700 0.7551345
#> 31     no     16 0.7476805 0.6421991 0.8430644
#> 32    yes     16 0.6667336 0.5274382 0.7828445
#> 33     no     17 0.7676327 0.6572971 0.8623594
#> 34    yes     17 0.6904365 0.5446415 0.8082147
#> 35     no     18 0.7864578 0.6720833 0.8796212
#> 36    yes     18 0.7131789 0.5617392 0.8312595
#> 37     no     19 0.8041469 0.6865358 0.8949819
#> 38    yes     19 0.7348919 0.5786915 0.8520419
#> 39     no     20 0.8207047 0.7006351 0.9085863
#> 40    yes     20 0.7555247 0.5954856 0.8707364
#> 41     no     21 0.8361480 0.7143640 0.9205848
#> 42    yes     21 0.7750433 0.6122249 0.8874895
#> 43     no     22 0.8505034 0.7277079 0.9311656
#> 44    yes     22 0.7934296 0.6287028 0.9023151
#> 45     no     23 0.8638059 0.7406548 0.9404410
#> 46    yes     23 0.8106802 0.6448859 0.9153737
#> 47     no     24 0.8760972 0.7531949 0.9485356
#> 48    yes     24 0.8268046 0.6607432 0.9268283
#> 49     no     25 0.8874238 0.7653207 0.9555818
#> 50    yes     25 0.8418238 0.6762468 0.9368067
#> 51     no     26 0.8978357 0.7770270 0.9617022
#> 52    yes     26 0.8557676 0.6913716 0.9454758
#> 53     no     27 0.9073851 0.7883110 0.9670082
#> 54    yes     27 0.8686741 0.7060955 0.9530154
#> 55     no     28 0.9161253 0.7984406 0.9716007
#> 56    yes     28 0.8805867 0.7203996 0.9595570
#> 57     no     29 0.9241097 0.8081170 0.9755699
#> 58    yes     29 0.8915537 0.7342681 0.9652211
#> 59     no     30 0.9313909 0.8175825 0.9789963
#> 60    yes     30 0.9016260 0.7476880 0.9701166
#> 61     no     31 0.9380204 0.8266794 0.9819432
#> 62    yes     31 0.9108563 0.7606495 0.9743413
#> 63     no     32 0.9440478 0.8354120 0.9844808
#> 64    yes     32 0.9192981 0.7731455 0.9779822
#> 65     no     33 0.9495206 0.8437858 0.9866677
#> 66    yes     33 0.9270045 0.7848692 0.9811165
#> 67     no     34 0.9544839 0.8518072 0.9885874
#> 68    yes     34 0.9340278 0.7950485 0.9838120
#> 69     no     35 0.9589802 0.8594833 0.9902333
#> 70    yes     35 0.9404188 0.8048649 0.9861281
#> 71     no     36 0.9630496 0.8668220 0.9916439
#> 72    yes     36 0.9462263 0.8143198 0.9881169
#> 73     no     37 0.9667293 0.8738317 0.9928523
#> 74    yes     37 0.9514970 0.8234158 0.9898234
#> 75     no     38 0.9700539 0.8805214 0.9938869
#> 76    yes     38 0.9562749 0.8321565 0.9912870
#> 77     no     39 0.9730556 0.8869004 0.9947726
#> 78    yes     39 0.9606016 0.8405468 0.9925417
#> 79     no     40 0.9757639 0.8929784 0.9955306
#> 80    yes     40 0.9645161 0.8485923 0.9936168
#> 81     no     41 0.9782061 0.8987652 0.9961787
#> 82    yes     41 0.9680546 0.8562995 0.9945378
#> 83     no     42 0.9804072 0.9042710 0.9967328
#> 84    yes     42 0.9712507 0.8638034 0.9953266
#> 85     no     43 0.9823899 0.9095058 0.9972067
#> 86    yes     43 0.9741357 0.8709911 0.9960019
#> 87     no     44 0.9841753 0.9144799 0.9976121
#> 88    yes     44 0.9767380 0.8778512 0.9965799
#> 89     no     45 0.9857822 0.9192033 0.9979587
#> 90    yes     45 0.9790842 0.8843932 0.9970746
#> 91     no     46 0.9872281 0.9236861 0.9982551
#> 92    yes     46 0.9811982 0.8906266 0.9974979
#> 93     no     47 0.9885287 0.9279383 0.9985086
#> 94    yes     47 0.9831023 0.8965615 0.9978601
#> 95     no     48 0.9896982 0.9319697 0.9987235
#> 96    yes     48 0.9848165 0.9022079 0.9981696
#> 97     no     49 0.9907496 0.9357898 0.9989072
#> 98    yes     49 0.9863593 0.9075762 0.9984340
#> 99     no     50 0.9916946 0.9394083 0.9990644
#> 100   yes     50 0.9877472 0.9126767 0.9986603
#> 101    no     51 0.9925437 0.9428341 0.9991991
#> 102   yes     51 0.9889955 0.9175196 0.9988540
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)