Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3100603 0.4272277
#> 2     yes      1 0.2814584 0.2119735 0.3452984
#> 3      no      2 0.3927672 0.3398446 0.4469535
#> 4     yes      2 0.3039568 0.2370822 0.3646270
#> 5      no      3 0.4189750 0.3667611 0.4712549
#> 6     yes      3 0.3274340 0.2639073 0.3922865
#> 7      no      4 0.4456481 0.3928405 0.4974265
#> 8     yes      4 0.3518080 0.2877353 0.4222192
#> 9      no      5 0.4726379 0.4231173 0.5272649
#> 10    yes      5 0.3769795 0.3093421 0.4505148
#> 11     no      6 0.4997887 0.4517197 0.5576709
#> 12    yes      6 0.4028326 0.3325879 0.4794764
#> 13     no      7 0.5269407 0.4767991 0.5962165
#> 14    yes      7 0.4292372 0.3534391 0.5136530
#> 15     no      8 0.5539343 0.4984795 0.6359613
#> 16    yes      8 0.4560509 0.3731062 0.5509244
#> 17     no      9 0.5806134 0.5174552 0.6722124
#> 18    yes      9 0.4831217 0.3969045 0.5862845
#> 19     no     10 0.6068295 0.5363755 0.7048543
#> 20    yes     10 0.5102919 0.4187970 0.6196447
#> 21     no     11 0.6324443 0.5551819 0.7355062
#> 22    yes     11 0.5374015 0.4361628 0.6516431
#> 23     no     12 0.6573325 0.5738172 0.7640223
#> 24    yes     12 0.5642918 0.4536877 0.6823274
#> 25     no     13 0.6813846 0.5922269 0.7903231
#> 26    yes     13 0.5908091 0.4713272 0.7115025
#> 27     no     14 0.7045073 0.6103591 0.8143887
#> 28    yes     14 0.6168084 0.4890355 0.7403893
#> 29     no     15 0.7266252 0.6268148 0.8362499
#> 30    yes     15 0.6421564 0.5067664 0.7685431
#> 31     no     16 0.7476805 0.6424073 0.8559787
#> 32    yes     16 0.6667336 0.5254327 0.7944590
#> 33     no     17 0.7676327 0.6567732 0.8736782
#> 34    yes     17 0.6904365 0.5446570 0.8178905
#> 35     no     18 0.7864578 0.6703390 0.8894737
#> 36    yes     18 0.7131789 0.5637476 0.8391863
#> 37     no     19 0.8041469 0.6836276 0.9035039
#> 38    yes     19 0.7348919 0.5824419 0.8587097
#> 39     no     20 0.8207047 0.6966236 0.9159143
#> 40    yes     20 0.7555247 0.5994274 0.8771295
#> 41     no     21 0.8361480 0.7093136 0.9268517
#> 42    yes     21 0.7750433 0.6161799 0.8934459
#> 43     no     22 0.8505034 0.7216855 0.9364601
#> 44    yes     22 0.7934296 0.6326634 0.9078231
#> 45     no     23 0.8638059 0.7337292 0.9452522
#> 46    yes     23 0.8106802 0.6488439 0.9204331
#> 47     no     24 0.8760972 0.7454362 0.9529550
#> 48    yes     24 0.8268046 0.6646904 0.9314481
#> 49     no     25 0.8874238 0.7567999 0.9596209
#> 50    yes     25 0.8418238 0.6801749 0.9410358
#> 51     no     26 0.8978357 0.7678149 0.9653770
#> 52    yes     26 0.8557676 0.6952720 0.9493552
#> 53     no     27 0.9073851 0.7784777 0.9703384
#> 54    yes     27 0.8686741 0.7099600 0.9565549
#> 55     no     28 0.9161253 0.7887862 0.9746078
#> 56    yes     28 0.8805867 0.7242199 0.9627710
#> 57     no     29 0.9241097 0.7987396 0.9782766
#> 58    yes     29 0.8915537 0.7380363 0.9681146
#> 59     no     30 0.9313909 0.8083385 0.9814255
#> 60    yes     30 0.9016260 0.7513965 0.9727081
#> 61     no     31 0.9380204 0.8175846 0.9841255
#> 62    yes     31 0.9108563 0.7643765 0.9766557
#> 63     no     32 0.9440478 0.8264810 0.9864386
#> 64    yes     32 0.9192981 0.7770100 0.9800439
#> 65     no     33 0.9495206 0.8350316 0.9884186
#> 66    yes     33 0.9270045 0.7882359 0.9829489
#> 67     no     34 0.9544839 0.8432411 0.9901124
#> 68    yes     34 0.9340278 0.7986128 0.9854372
#> 69     no     35 0.9589802 0.8511154 0.9915607
#> 70    yes     35 0.9404188 0.8086041 0.9875669
#> 71     no     36 0.9630496 0.8586609 0.9927983
#> 72    yes     36 0.9462263 0.8182801 0.9893885
#> 73     no     37 0.9667293 0.8658846 0.9938556
#> 74    yes     37 0.9514970 0.8276050 0.9909456
#> 75     no     38 0.9700539 0.8727941 0.9947585
#> 76    yes     38 0.9562749 0.8365464 0.9922759
#> 77     no     39 0.9730556 0.8793975 0.9955293
#> 78    yes     39 0.9606016 0.8451102 0.9934121
#> 79     no     40 0.9757639 0.8857034 0.9961872
#> 80    yes     40 0.9645161 0.8533032 0.9943820
#> 81     no     41 0.9782061 0.8917204 0.9967486
#> 82    yes     41 0.9680546 0.8611335 0.9952098
#> 83     no     42 0.9804072 0.8974576 0.9972275
#> 84    yes     42 0.9712507 0.8686094 0.9959162
#> 85     no     43 0.9823899 0.9029241 0.9976361
#> 86    yes     43 0.9741357 0.8757403 0.9965187
#> 87     no     44 0.9841753 0.9081292 0.9979846
#> 88    yes     44 0.9767380 0.8825358 0.9970326
#> 89     no     45 0.9857822 0.9130824 0.9982818
#> 90    yes     45 0.9790842 0.8890062 0.9974708
#> 91     no     46 0.9872281 0.9177928 0.9985352
#> 92    yes     46 0.9811982 0.8951619 0.9978444
#> 93     no     47 0.9885287 0.9224157 0.9987513
#> 94    yes     47 0.9831023 0.9010136 0.9981629
#> 95     no     48 0.9896982 0.9268284 0.9989355
#> 96    yes     48 0.9848165 0.9065721 0.9984344
#> 97     no     49 0.9907496 0.9310088 0.9990926
#> 98    yes     49 0.9863593 0.9118485 0.9986659
#> 99     no     50 0.9916946 0.9349671 0.9992265
#> 100   yes     50 0.9877472 0.9168537 0.9988631
#> 101    no     51 0.9925437 0.9387132 0.9993407
#> 102   yes     51 0.9889955 0.9215986 0.9990312
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)