Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3033959 0.4188489
#> 2     yes      1 0.2814584 0.2227314 0.3471034
#> 3      no      2 0.3927672 0.3339176 0.4413700
#> 4     yes      2 0.3039568 0.2488664 0.3671766
#> 5      no      3 0.4189750 0.3644009 0.4694087
#> 6     yes      3 0.3274340 0.2732815 0.3911581
#> 7      no      4 0.4456481 0.3952890 0.5001726
#> 8     yes      4 0.3518080 0.2946307 0.4173591
#> 9      no      5 0.4726379 0.4249617 0.5322813
#> 10    yes      5 0.3769795 0.3173875 0.4444746
#> 11     no      6 0.4997887 0.4464412 0.5654525
#> 12    yes      6 0.4028326 0.3384070 0.4763726
#> 13     no      7 0.5269407 0.4673319 0.5953350
#> 14    yes      7 0.4292372 0.3578909 0.5110204
#> 15     no      8 0.5539343 0.4802742 0.6256261
#> 16    yes      8 0.4560509 0.3784745 0.5473775
#> 17     no      9 0.5806134 0.4936949 0.6569372
#> 18    yes      9 0.4831217 0.3974783 0.5811778
#> 19     no     10 0.6068295 0.5071168 0.6865780
#> 20    yes     10 0.5102919 0.4133011 0.6142908
#> 21     no     11 0.6324443 0.5188794 0.7147555
#> 22    yes     11 0.5374015 0.4273609 0.6482002
#> 23     no     12 0.6573325 0.5301843 0.7413491
#> 24    yes     12 0.5642918 0.4420621 0.6806785
#> 25     no     13 0.6813846 0.5414430 0.7665297
#> 26    yes     13 0.5908091 0.4568678 0.7114926
#> 27     no     14 0.7045073 0.5536655 0.7918644
#> 28    yes     14 0.6168084 0.4717452 0.7404663
#> 29     no     15 0.7266252 0.5658350 0.8151125
#> 30    yes     15 0.6421564 0.4864792 0.7673136
#> 31     no     16 0.7476805 0.5774885 0.8362995
#> 32    yes     16 0.6667336 0.5009138 0.7943906
#> 33     no     17 0.7676327 0.5887464 0.8554875
#> 34    yes     17 0.6904365 0.5153253 0.8192374
#> 35     no     18 0.7864578 0.5999021 0.8727546
#> 36    yes     18 0.7131789 0.5282749 0.8416640
#> 37     no     19 0.8041469 0.6109448 0.8881062
#> 38    yes     19 0.7348919 0.5398179 0.8617614
#> 39     no     20 0.8207047 0.6218642 0.9018122
#> 40    yes     20 0.7555247 0.5513085 0.8796569
#> 41     no     21 0.8361480 0.6326505 0.9140003
#> 42    yes     21 0.7750433 0.5627344 0.8951689
#> 43     no     22 0.8505034 0.6432944 0.9248000
#> 44    yes     22 0.7934296 0.5740835 0.9085761
#> 45     no     23 0.8638059 0.6537874 0.9343395
#> 46    yes     23 0.8106802 0.5853440 0.9204200
#> 47     no     24 0.8760972 0.6641213 0.9427425
#> 48    yes     24 0.8268046 0.5965046 0.9308453
#> 49     no     25 0.8874238 0.6742888 0.9501263
#> 50    yes     25 0.8418238 0.6075544 0.9399931
#> 51     no     26 0.8978357 0.6842832 0.9566007
#> 52    yes     26 0.8557676 0.6184830 0.9479978
#> 53     no     27 0.9073851 0.6940984 0.9625475
#> 54    yes     27 0.8686741 0.6292808 0.9549852
#> 55     no     28 0.9161253 0.7037289 0.9677272
#> 56    yes     28 0.8805867 0.6399383 0.9610719
#> 57     no     29 0.9241097 0.7131699 0.9722111
#> 58    yes     29 0.8915537 0.6504470 0.9663641
#> 59     no     30 0.9313909 0.7224172 0.9760874
#> 60    yes     30 0.9016260 0.6607989 0.9709582
#> 61     no     31 0.9380204 0.7314672 0.9794344
#> 62    yes     31 0.9108563 0.6709865 0.9749408
#> 63     no     32 0.9440478 0.7403170 0.9823213
#> 64    yes     32 0.9192981 0.6810031 0.9783892
#> 65     no     33 0.9495206 0.7489642 0.9848092
#> 66    yes     33 0.9270045 0.6908425 0.9813719
#> 67     no     34 0.9544839 0.7574069 0.9869516
#> 68    yes     34 0.9340278 0.7004993 0.9839495
#> 69     no     35 0.9589802 0.7656440 0.9887952
#> 70    yes     35 0.9404188 0.7099685 0.9861753
#> 71     no     36 0.9630496 0.7736745 0.9903808
#> 72    yes     36 0.9462263 0.7192459 0.9880961
#> 73     no     37 0.9667293 0.7814983 0.9917439
#> 74    yes     37 0.9514970 0.7283278 0.9897526
#> 75     no     38 0.9700539 0.7891155 0.9929152
#> 76    yes     38 0.9562749 0.7372114 0.9911806
#> 77     no     39 0.9730556 0.7965268 0.9939213
#> 78    yes     39 0.9606016 0.7458940 0.9924110
#> 79     no     40 0.9757639 0.8037331 0.9947852
#> 80    yes     40 0.9645161 0.7543737 0.9934708
#> 81     no     41 0.9782061 0.8107359 0.9955269
#> 82    yes     41 0.9680546 0.7626493 0.9943834
#> 83     no     42 0.9804072 0.8175369 0.9961635
#> 84    yes     42 0.9712507 0.7707199 0.9951690
#> 85     no     43 0.9823899 0.8241381 0.9967098
#> 86    yes     43 0.9741357 0.7785851 0.9958451
#> 87     no     44 0.9841753 0.8305420 0.9971785
#> 88    yes     44 0.9767380 0.7862449 0.9964269
#> 89     no     45 0.9857822 0.8367512 0.9975792
#> 90    yes     45 0.9790842 0.7937000 0.9969275
#> 91     no     46 0.9872281 0.8427685 0.9979231
#> 92    yes     46 0.9811982 0.8009512 0.9973581
#> 93     no     47 0.9885287 0.8485971 0.9982182
#> 94    yes     47 0.9831023 0.8079999 0.9977284
#> 95     no     48 0.9896982 0.8542401 0.9984714
#> 96    yes     48 0.9848165 0.8148476 0.9980469
#> 97     no     49 0.9907496 0.8597011 0.9986886
#> 98    yes     49 0.9863593 0.8214965 0.9983208
#> 99     no     50 0.9916946 0.8649835 0.9988751
#> 100   yes     50 0.9877472 0.8279487 0.9985564
#> 101    no     51 0.9925437 0.8700912 0.9990350
#> 102   yes     51 0.9889955 0.8342067 0.9987589
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)