Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3049644 0.4266749
#> 2     yes      1 0.2814584 0.2181302 0.3499862
#> 3      no      2 0.3927672 0.3366707 0.4444183
#> 4     yes      2 0.3039568 0.2398282 0.3647957
#> 5      no      3 0.4189750 0.3706120 0.4706871
#> 6     yes      3 0.3274340 0.2647458 0.3872793
#> 7      no      4 0.4456481 0.4009176 0.4981158
#> 8     yes      4 0.3518080 0.2882004 0.4121149
#> 9      no      5 0.4726379 0.4251229 0.5297660
#> 10    yes      5 0.3769795 0.3084663 0.4390213
#> 11     no      6 0.4997887 0.4416016 0.5590694
#> 12    yes      6 0.4028326 0.3282177 0.4733149
#> 13     no      7 0.5269407 0.4637485 0.5879597
#> 14    yes      7 0.4292372 0.3486188 0.5056235
#> 15     no      8 0.5539343 0.4797971 0.6195225
#> 16    yes      8 0.4560509 0.3672873 0.5370323
#> 17     no      9 0.5806134 0.4933469 0.6508718
#> 18    yes      9 0.4831217 0.3832570 0.5700165
#> 19     no     10 0.6068295 0.5065820 0.6842863
#> 20    yes     10 0.5102919 0.3994876 0.6029258
#> 21     no     11 0.6324443 0.5198076 0.7181353
#> 22    yes     11 0.5374015 0.4157302 0.6347547
#> 23     no     12 0.6573325 0.5330053 0.7489093
#> 24    yes     12 0.5642918 0.4285204 0.6672064
#> 25     no     13 0.6813846 0.5461566 0.7761060
#> 26    yes     13 0.5908091 0.4414078 0.7009185
#> 27     no     14 0.7045073 0.5597895 0.8011360
#> 28    yes     14 0.6168084 0.4543754 0.7324188
#> 29     no     15 0.7266252 0.5734641 0.8239979
#> 30    yes     15 0.6421564 0.4683646 0.7622208
#> 31     no     16 0.7476805 0.5870273 0.8448848
#> 32    yes     16 0.6667336 0.4839109 0.7901769
#> 33     no     17 0.7676327 0.6004599 0.8647540
#> 34    yes     17 0.6904365 0.4994929 0.8156280
#> 35     no     18 0.7864578 0.6137431 0.8823687
#> 36    yes     18 0.7131789 0.5158248 0.8386100
#> 37     no     19 0.8041469 0.6268591 0.8979513
#> 38    yes     19 0.7348919 0.5322528 0.8592102
#> 39     no     20 0.8207047 0.6397912 0.9116759
#> 40    yes     20 0.7555247 0.5486035 0.8775544
#> 41     no     21 0.8361480 0.6525234 0.9237109
#> 42    yes     21 0.7750433 0.5648382 0.8937947
#> 43     no     22 0.8505034 0.6650411 0.9342240
#> 44    yes     22 0.7934296 0.5803380 0.9080984
#> 45     no     23 0.8638059 0.6773308 0.9433768
#> 46    yes     23 0.8106802 0.5935225 0.9206398
#> 47     no     24 0.8760972 0.6894526 0.9513220
#> 48    yes     24 0.8268046 0.6065709 0.9315928
#> 49     no     25 0.8874238 0.7013263 0.9582015
#> 50    yes     25 0.8418238 0.6194661 0.9410072
#> 51     no     26 0.8978357 0.7129348 0.9641451
#> 52    yes     26 0.8557676 0.6321919 0.9490552
#> 53     no     27 0.9073851 0.7242693 0.9692704
#> 54    yes     27 0.8686741 0.6447332 0.9560566
#> 55     no     28 0.9161253 0.7353222 0.9736830
#> 56    yes     28 0.8805867 0.6570758 0.9621343
#> 57     no     29 0.9241097 0.7460871 0.9774764
#> 58    yes     29 0.8915537 0.6692065 0.9674836
#> 59     no     30 0.9313909 0.7565589 0.9807338
#> 60    yes     30 0.9016260 0.6811133 0.9721829
#> 61     no     31 0.9380204 0.7667336 0.9835279
#> 62    yes     31 0.9108563 0.6927851 0.9762200
#> 63     no     32 0.9440478 0.7766083 0.9859225
#> 64    yes     32 0.9192981 0.7042122 0.9796836
#> 65     no     33 0.9495206 0.7861598 0.9879732
#> 66    yes     33 0.9270045 0.7153858 0.9826518
#> 67     no     34 0.9544839 0.7951916 0.9897282
#> 68    yes     34 0.9340278 0.7264919 0.9851930
#> 69     no     35 0.9589802 0.8039371 0.9912293
#> 70    yes     35 0.9404188 0.7374481 0.9873668
#> 71     no     36 0.9630496 0.8123973 0.9925127
#> 72    yes     36 0.9462263 0.7481169 0.9892250
#> 73     no     37 0.9667293 0.8205739 0.9936094
#> 74    yes     37 0.9514970 0.7584934 0.9908124
#> 75     no     38 0.9700539 0.8284693 0.9945463
#> 76    yes     38 0.9562749 0.7685741 0.9921678
#> 77     no     39 0.9730556 0.8360867 0.9953465
#> 78    yes     39 0.9606016 0.7783564 0.9933246
#> 79     no     40 0.9757639 0.8434298 0.9960297
#> 80    yes     40 0.9645161 0.7878388 0.9943115
#> 81     no     41 0.9782061 0.8505028 0.9966130
#> 82    yes     41 0.9680546 0.7970209 0.9951532
#> 83     no     42 0.9804072 0.8573102 0.9971107
#> 84    yes     42 0.9712507 0.8059030 0.9958709
#> 85     no     43 0.9823899 0.8638574 0.9975355
#> 86    yes     43 0.9741357 0.8144864 0.9964826
#> 87     no     44 0.9841753 0.8701496 0.9978980
#> 88    yes     44 0.9767380 0.8227734 0.9970040
#> 89     no     45 0.9857822 0.8761927 0.9982072
#> 90    yes     45 0.9790842 0.8307668 0.9974483
#> 91     no     46 0.9872281 0.8819927 0.9984710
#> 92    yes     46 0.9811982 0.8384701 0.9978269
#> 93     no     47 0.9885287 0.8875558 0.9986960
#> 94    yes     47 0.9831023 0.8458876 0.9981494
#> 95     no     48 0.9896982 0.8928886 0.9988880
#> 96    yes     48 0.9848165 0.8530238 0.9984241
#> 97     no     49 0.9907496 0.8979975 0.9990517
#> 98    yes     49 0.9863593 0.8598842 0.9986580
#> 99     no     50 0.9916946 0.9028892 0.9991913
#> 100   yes     50 0.9877472 0.8664742 0.9988573
#> 101    no     51 0.9925437 0.9075705 0.9993104
#> 102   yes     51 0.9889955 0.8727999 0.9990270
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)