Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3064724 0.4191639
#> 2     yes      1 0.2814584 0.2206554 0.3700316
#> 3      no      2 0.3927672 0.3363166 0.4391637
#> 4     yes      2 0.3039568 0.2461056 0.3881010
#> 5      no      3 0.4189750 0.3683263 0.4642627
#> 6     yes      3 0.3274340 0.2702835 0.4090279
#> 7      no      4 0.4456481 0.3984834 0.4937148
#> 8     yes      4 0.3518080 0.2931181 0.4328558
#> 9      no      5 0.4726379 0.4281886 0.5266688
#> 10    yes      5 0.3769795 0.3176820 0.4596073
#> 11     no      6 0.4997887 0.4491071 0.5559714
#> 12    yes      6 0.4028326 0.3452455 0.4866065
#> 13     no      7 0.5269407 0.4733387 0.5950333
#> 14    yes      7 0.4292372 0.3674215 0.5136908
#> 15     no      8 0.5539343 0.4957845 0.6269268
#> 16    yes      8 0.4560509 0.3857026 0.5424684
#> 17     no      9 0.5806134 0.5148907 0.6605065
#> 18    yes      9 0.4831217 0.4057230 0.5739052
#> 19     no     10 0.6068295 0.5325751 0.6938630
#> 20    yes     10 0.5102919 0.4259515 0.6094395
#> 21     no     11 0.6324443 0.5502571 0.7252719
#> 22    yes     11 0.5374015 0.4442366 0.6486671
#> 23     no     12 0.6573325 0.5677107 0.7555074
#> 24    yes     12 0.5642918 0.4627622 0.6846754
#> 25     no     13 0.6813846 0.5846411 0.7840496
#> 26    yes     13 0.5908091 0.4820829 0.7179517
#> 27     no     14 0.7045073 0.6013618 0.8100468
#> 28    yes     14 0.6168084 0.5011283 0.7490043
#> 29     no     15 0.7266252 0.6161201 0.8335776
#> 30    yes     15 0.6421564 0.5189603 0.7779691
#> 31     no     16 0.7476805 0.6300486 0.8547152
#> 32    yes     16 0.6667336 0.5367457 0.8046550
#> 33     no     17 0.7676327 0.6437624 0.8735742
#> 34    yes     17 0.6904365 0.5541587 0.8288393
#> 35     no     18 0.7864578 0.6572430 0.8902984
#> 36    yes     18 0.7131789 0.5712595 0.8505853
#> 37     no     19 0.8041469 0.6704732 0.9050497
#> 38    yes     19 0.7348919 0.5881939 0.8700014
#> 39     no     20 0.8207047 0.6839931 0.9179994
#> 40    yes     20 0.7555247 0.6049236 0.8872284
#> 41     no     21 0.8361480 0.6973482 0.9293201
#> 42    yes     21 0.7750433 0.6214125 0.9024280
#> 43     no     22 0.8505034 0.7103791 0.9391809
#> 44    yes     22 0.7934296 0.6376263 0.9157641
#> 45     no     23 0.8638059 0.7230729 0.9477427
#> 46    yes     23 0.8106802 0.6535180 0.9273738
#> 47     no     24 0.8760972 0.7354186 0.9551564
#> 48    yes     24 0.8268046 0.6679482 0.9374920
#> 49     no     25 0.8874238 0.7474071 0.9615605
#> 50    yes     25 0.8418238 0.6820665 0.9462818
#> 51     no     26 0.8978357 0.7590313 0.9670813
#> 52    yes     26 0.8557676 0.6959292 0.9538959
#> 53     no     27 0.9073851 0.7702856 0.9718321
#> 54    yes     27 0.8686741 0.7098133 0.9604755
#> 55     no     28 0.9161253 0.7811665 0.9759140
#> 56    yes     28 0.8805867 0.7233092 0.9661491
#> 57     no     29 0.9241097 0.7916720 0.9794168
#> 58    yes     29 0.8915537 0.7364036 0.9710325
#> 59     no     30 0.9313909 0.8018015 0.9824192
#> 60    yes     30 0.9016260 0.7487175 0.9752293
#> 61     no     31 0.9380204 0.8115561 0.9849902
#> 62    yes     31 0.9108563 0.7602737 0.9788311
#> 63     no     32 0.9440478 0.8209381 0.9871900
#> 64    yes     32 0.9192981 0.7714577 0.9819187
#> 65     no     33 0.9495206 0.8299512 0.9890708
#> 66    yes     33 0.9270045 0.7822661 0.9845629
#> 67     no     34 0.9544839 0.8386000 0.9906780
#> 68    yes     34 0.9340278 0.7926975 0.9868254
#> 69     no     35 0.9589802 0.8468901 0.9920507
#> 70    yes     35 0.9404188 0.8027519 0.9887601
#> 71     no     36 0.9630496 0.8548283 0.9932226
#> 72    yes     36 0.9462263 0.8124308 0.9904133
#> 73     no     37 0.9667293 0.8624218 0.9942227
#> 74    yes     37 0.9514970 0.8217370 0.9918252
#> 75     no     38 0.9700539 0.8696787 0.9950759
#> 76    yes     38 0.9562749 0.8306747 0.9930306
#> 77     no     39 0.9730556 0.8766075 0.9958035
#> 78    yes     39 0.9606016 0.8392487 0.9940593
#> 79     no     40 0.9757639 0.8832174 0.9964240
#> 80    yes     40 0.9645161 0.8474654 0.9949369
#> 81     no     41 0.9782061 0.8895177 0.9969530
#> 82    yes     41 0.9680546 0.8553315 0.9956853
#> 83     no     42 0.9804072 0.8955183 0.9974040
#> 84    yes     42 0.9712507 0.8628548 0.9963235
#> 85     no     43 0.9823899 0.9012291 0.9977883
#> 86    yes     43 0.9741357 0.8700435 0.9968676
#> 87     no     44 0.9841753 0.9066601 0.9981158
#> 88    yes     44 0.9767380 0.8772135 0.9973313
#> 89     no     45 0.9857822 0.9118216 0.9983949
#> 90    yes     45 0.9790842 0.8840962 0.9977265
#> 91     no     46 0.9872281 0.9167238 0.9986327
#> 92    yes     46 0.9811982 0.8906412 0.9980636
#> 93     no     47 0.9885287 0.9213768 0.9988352
#> 94    yes     47 0.9831023 0.8968597 0.9983520
#> 95     no     48 0.9896982 0.9257908 0.9990078
#> 96    yes     48 0.9848165 0.9026478 0.9985975
#> 97     no     49 0.9907496 0.9299756 0.9991549
#> 98    yes     49 0.9863593 0.9080134 0.9988065
#> 99     no     50 0.9916946 0.9339412 0.9992801
#> 100   yes     50 0.9877472 0.9131120 0.9989844
#> 101    no     51 0.9925437 0.9376971 0.9993868
#> 102   yes     51 0.9889955 0.9179537 0.9991357
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)