Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3059894 0.4144533
#> 2     yes      1 0.2814584 0.2031904 0.3506575
#> 3      no      2 0.3927672 0.3346066 0.4366836
#> 4     yes      2 0.3039568 0.2269830 0.3740866
#> 5      no      3 0.4189750 0.3676155 0.4617723
#> 6     yes      3 0.3274340 0.2498597 0.3964947
#> 7      no      4 0.4456481 0.4006317 0.4924558
#> 8     yes      4 0.3518080 0.2728734 0.4134650
#> 9      no      5 0.4726379 0.4318053 0.5231358
#> 10    yes      5 0.3769795 0.3024054 0.4294337
#> 11     no      6 0.4997887 0.4558965 0.5600305
#> 12    yes      6 0.4028326 0.3313325 0.4632771
#> 13     no      7 0.5269407 0.4780616 0.5966549
#> 14    yes      7 0.4292372 0.3554181 0.4991920
#> 15     no      8 0.5539343 0.4982706 0.6336053
#> 16    yes      8 0.4560509 0.3763368 0.5412260
#> 17     no      9 0.5806134 0.5184815 0.6688045
#> 18    yes      9 0.4831217 0.3924111 0.5753397
#> 19     no     10 0.6068295 0.5355628 0.7062354
#> 20    yes     10 0.5102919 0.4108914 0.6156094
#> 21     no     11 0.6324443 0.5513504 0.7408746
#> 22    yes     11 0.5374015 0.4344898 0.6543521
#> 23     no     12 0.6573325 0.5670364 0.7727328
#> 24    yes     12 0.5642918 0.4585919 0.6910716
#> 25     no     13 0.6813846 0.5825889 0.8014142
#> 26    yes     13 0.5908091 0.4828822 0.7252589
#> 27     no     14 0.7045073 0.5980451 0.8266962
#> 28    yes     14 0.6168084 0.4979574 0.7569932
#> 29     no     15 0.7266252 0.6137245 0.8493632
#> 30    yes     15 0.6421564 0.5122262 0.7861429
#> 31     no     16 0.7476805 0.6291676 0.8695325
#> 32    yes     16 0.6667336 0.5275388 0.8126613
#> 33     no     17 0.7676327 0.6443459 0.8873584
#> 34    yes     17 0.6904365 0.5432104 0.8365751
#> 35     no     18 0.7864578 0.6592334 0.9030198
#> 36    yes     18 0.7131789 0.5587743 0.8579699
#> 37     no     19 0.8041469 0.6738061 0.9167074
#> 38    yes     19 0.7348919 0.5741902 0.8769757
#> 39     no     20 0.8207047 0.6880428 0.9286154
#> 40    yes     20 0.7555247 0.5894201 0.8937533
#> 41     no     21 0.8361480 0.7019245 0.9389339
#> 42    yes     21 0.7750433 0.6044279 0.9084817
#> 43     no     22 0.8505034 0.7154348 0.9478444
#> 44    yes     22 0.7934296 0.6191801 0.9213481
#> 45     no     23 0.8638059 0.7285601 0.9555160
#> 46    yes     23 0.8106802 0.6336988 0.9325400
#> 47     no     24 0.8760972 0.7412889 0.9621040
#> 48    yes     24 0.8268046 0.6479773 0.9422392
#> 49     no     25 0.8874238 0.7533564 0.9677492
#> 50    yes     25 0.8418238 0.6619858 0.9506178
#> 51     no     26 0.8978357 0.7648587 0.9725772
#> 52    yes     26 0.8557676 0.6757058 0.9578354
#> 53     no     27 0.9073851 0.7759781 0.9766997
#> 54    yes     27 0.8686741 0.6891206 0.9640380
#> 55     no     28 0.9161253 0.7867123 0.9802149
#> 56    yes     28 0.8805867 0.7022154 0.9693574
#> 57     no     29 0.9241097 0.7970608 0.9832089
#> 58    yes     29 0.8915537 0.7149774 0.9739113
#> 59     no     30 0.9313909 0.8070246 0.9857563
#> 60    yes     30 0.9016260 0.7273957 0.9779859
#> 61     no     31 0.9380204 0.8166061 0.9879219
#> 62    yes     31 0.9108563 0.7394614 0.9814857
#> 63     no     32 0.9440478 0.8258090 0.9897616
#> 64    yes     32 0.9192981 0.7511675 0.9844369
#> 65     no     33 0.9495206 0.8346384 0.9913235
#> 66    yes     33 0.9270045 0.7625085 0.9869230
#> 67     no     34 0.9544839 0.8431002 0.9926489
#> 68    yes     34 0.9340278 0.7734809 0.9890068
#> 69     no     35 0.9589802 0.8512012 0.9937730
#> 70    yes     35 0.9404188 0.7840827 0.9907462
#> 71     no     36 0.9630496 0.8589493 0.9947261
#> 72    yes     36 0.9462263 0.7943134 0.9922120
#> 73     no     37 0.9667293 0.8663528 0.9955340
#> 74    yes     37 0.9514970 0.8041740 0.9934466
#> 75     no     38 0.9700539 0.8734208 0.9962185
#> 76    yes     38 0.9562749 0.8136667 0.9944862
#> 77     no     39 0.9730556 0.8801625 0.9967985
#> 78    yes     39 0.9606016 0.8227950 0.9953612
#> 79     no     40 0.9757639 0.8865880 0.9972897
#> 80    yes     40 0.9645161 0.8315633 0.9960938
#> 81     no     41 0.9782061 0.8927073 0.9977057
#> 82    yes     41 0.9680546 0.8399772 0.9966955
#> 83     no     42 0.9804072 0.8985308 0.9980580
#> 84    yes     42 0.9712507 0.8480429 0.9972047
#> 85     no     43 0.9823899 0.9040690 0.9983563
#> 86    yes     43 0.9741357 0.8557675 0.9976357
#> 87     no     44 0.9841753 0.9093323 0.9986088
#> 88    yes     44 0.9767380 0.8631588 0.9980003
#> 89     no     45 0.9857822 0.9143313 0.9988225
#> 90    yes     45 0.9790842 0.8702251 0.9983088
#> 91     no     46 0.9872281 0.9190765 0.9990035
#> 92    yes     46 0.9811982 0.8769749 0.9985697
#> 93     no     47 0.9885287 0.9235783 0.9991566
#> 94    yes     47 0.9831023 0.8834175 0.9987905
#> 95     no     48 0.9896982 0.9278470 0.9992863
#> 96    yes     48 0.9848165 0.8895623 0.9989772
#> 97     no     49 0.9907496 0.9318925 0.9993960
#> 98    yes     49 0.9863593 0.8954187 0.9991351
#> 99     no     50 0.9916946 0.9357249 0.9994888
#> 100   yes     50 0.9877472 0.9009965 0.9992686
#> 101    no     51 0.9925437 0.9393536 0.9995674
#> 102   yes     51 0.9889955 0.9063057 0.9993816
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)