Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.2889696 0.4264679
#> 2     yes      1 0.2814584 0.2091797 0.3555612
#> 3      no      2 0.3927672 0.3247977 0.4481968
#> 4     yes      2 0.3039568 0.2305168 0.3791863
#> 5      no      3 0.4189750 0.3616201 0.4676392
#> 6     yes      3 0.3274340 0.2538922 0.4007414
#> 7      no      4 0.4456481 0.3975736 0.4884279
#> 8     yes      4 0.3518080 0.2797465 0.4212970
#> 9      no      5 0.4726379 0.4165771 0.5225893
#> 10    yes      5 0.3769795 0.3083297 0.4454863
#> 11     no      6 0.4997887 0.4384637 0.5612610
#> 12    yes      6 0.4028326 0.3358825 0.4787205
#> 13     no      7 0.5269407 0.4613918 0.5969560
#> 14    yes      7 0.4292372 0.3609325 0.5177589
#> 15     no      8 0.5539343 0.4850284 0.6374733
#> 16    yes      8 0.4560509 0.3812248 0.5581589
#> 17     no      9 0.5806134 0.5023232 0.6761207
#> 18    yes      9 0.4831217 0.4019676 0.5972358
#> 19     no     10 0.6068295 0.5194835 0.7124998
#> 20    yes     10 0.5102919 0.4237393 0.6339112
#> 21     no     11 0.6324443 0.5366152 0.7463201
#> 22    yes     11 0.5374015 0.4463409 0.6697288
#> 23     no     12 0.6573325 0.5533298 0.7773998
#> 24    yes     12 0.5642918 0.4663334 0.7024953
#> 25     no     13 0.6813846 0.5685473 0.8056589
#> 26    yes     13 0.5908091 0.4818137 0.7352804
#> 27     no     14 0.7045073 0.5839456 0.8311061
#> 28    yes     14 0.6168084 0.4973335 0.7662918
#> 29     no     15 0.7266252 0.5991828 0.8538228
#> 30    yes     15 0.6421564 0.5128619 0.7946719
#> 31     no     16 0.7476805 0.6142300 0.8739450
#> 32    yes     16 0.6667336 0.5283673 0.8204011
#> 33     no     17 0.7676327 0.6290597 0.8916467
#> 34    yes     17 0.6904365 0.5438187 0.8435293
#> 35     no     18 0.7864578 0.6438591 0.9071251
#> 36    yes     18 0.7131789 0.5594736 0.8651512
#> 37     no     19 0.8041469 0.6587056 0.9205880
#> 38    yes     19 0.7348919 0.5752317 0.8844159
#> 39     no     20 0.8207047 0.6732406 0.9322441
#> 40    yes     20 0.7555247 0.5908349 0.9012437
#> 41     no     21 0.8361480 0.6874423 0.9422957
#> 42    yes     21 0.7750433 0.6047565 0.9157139
#> 43     no     22 0.8505034 0.7012918 0.9509339
#> 44    yes     22 0.7934296 0.6185019 0.9281422
#> 45     no     23 0.8638059 0.7147722 0.9583355
#> 46    yes     23 0.8106802 0.6320592 0.9388605
#> 47     no     24 0.8760972 0.7278698 0.9646617
#> 48    yes     24 0.8268046 0.6454100 0.9480695
#> 49     no     25 0.8874238 0.7405728 0.9700569
#> 50    yes     25 0.8418238 0.6585370 0.9559563
#> 51     no     26 0.8978357 0.7528724 0.9746497
#> 52    yes     26 0.8557676 0.6714244 0.9626922
#> 53     no     27 0.9073851 0.7647618 0.9785534
#> 54    yes     27 0.8686741 0.6840577 0.9684318
#> 55     no     28 0.9161253 0.7762368 0.9818669
#> 56    yes     28 0.8805867 0.6964238 0.9733125
#> 57     no     29 0.9241097 0.7872952 0.9846762
#> 58    yes     29 0.8915537 0.7085110 0.9774560
#> 59     no     30 0.9313909 0.7972418 0.9870559
#> 60    yes     30 0.9016260 0.7203090 0.9809684
#> 61     no     31 0.9380204 0.8064008 0.9890700
#> 62    yes     31 0.9108563 0.7318092 0.9839424
#> 63     no     32 0.9440478 0.8152407 0.9907735
#> 64    yes     32 0.9192981 0.7430040 0.9864579
#> 65     no     33 0.9495206 0.8237641 0.9922135
#> 66    yes     33 0.9270045 0.7538876 0.9885837
#> 67     no     34 0.9544839 0.8319742 0.9934301
#> 68    yes     34 0.9340278 0.7644553 0.9903789
#> 69     no     35 0.9589802 0.8398753 0.9944609
#> 70    yes     35 0.9404188 0.7747037 0.9918940
#> 71     no     36 0.9630496 0.8474720 0.9953514
#> 72    yes     36 0.9462263 0.7846307 0.9931721
#> 73     no     37 0.9667293 0.8547697 0.9960994
#> 74    yes     37 0.9514970 0.7942355 0.9942497
#> 75     no     38 0.9700539 0.8617745 0.9967274
#> 76    yes     38 0.9562749 0.8035181 0.9951539
#> 77     no     39 0.9730556 0.8684926 0.9972546
#> 78    yes     39 0.9606016 0.8124799 0.9959154
#> 79     no     40 0.9757639 0.8749308 0.9976970
#> 80    yes     40 0.9645161 0.8211229 0.9965576
#> 81     no     41 0.9782061 0.8810964 0.9980683
#> 82    yes     41 0.9680546 0.8294503 0.9970991
#> 83     no     42 0.9804072 0.8869966 0.9983799
#> 84    yes     42 0.9712507 0.8374658 0.9975556
#> 85     no     43 0.9823899 0.8926392 0.9986412
#> 86    yes     43 0.9741357 0.8451741 0.9979403
#> 87     no     44 0.9841753 0.8980319 0.9988605
#> 88    yes     44 0.9767380 0.8525804 0.9982646
#> 89     no     45 0.9857822 0.9031825 0.9990444
#> 90    yes     45 0.9790842 0.8596904 0.9985379
#> 91     no     46 0.9872281 0.9080992 0.9991986
#> 92    yes     46 0.9811982 0.8665104 0.9987681
#> 93     no     47 0.9885287 0.9127898 0.9993280
#> 94    yes     47 0.9831023 0.8730472 0.9989622
#> 95     no     48 0.9896982 0.9172625 0.9994365
#> 96    yes     48 0.9848165 0.8793078 0.9991257
#> 97     no     49 0.9907496 0.9215251 0.9995275
#> 98    yes     49 0.9863593 0.8852996 0.9992634
#> 99     no     50 0.9916946 0.9255856 0.9996038
#> 100   yes     50 0.9877472 0.8910302 0.9993804
#> 101    no     51 0.9925437 0.9294517 0.9996678
#> 102   yes     51 0.9889955 0.8965074 0.9994818
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)