Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3059981 0.4346657
#> 2     yes      1 0.2814584 0.2208062 0.3502744
#> 3      no      2 0.3927672 0.3338305 0.4503949
#> 4     yes      2 0.3039568 0.2419739 0.3674317
#> 5      no      3 0.4189750 0.3630079 0.4660090
#> 6     yes      3 0.3274340 0.2627762 0.3895830
#> 7      no      4 0.4456481 0.3917252 0.4864051
#> 8     yes      4 0.3518080 0.2848759 0.4125560
#> 9      no      5 0.4726379 0.4171821 0.5164894
#> 10    yes      5 0.3769795 0.3086999 0.4392523
#> 11     no      6 0.4997887 0.4414180 0.5480998
#> 12    yes      6 0.4028326 0.3317677 0.4670092
#> 13     no      7 0.5269407 0.4685560 0.5793231
#> 14    yes      7 0.4292372 0.3525113 0.5003362
#> 15     no      8 0.5539343 0.4928010 0.6105308
#> 16    yes      8 0.4560509 0.3738301 0.5359638
#> 17     no      9 0.5806134 0.5142988 0.6434310
#> 18    yes      9 0.4831217 0.3956535 0.5721022
#> 19     no     10 0.6068295 0.5360040 0.6754177
#> 20    yes     10 0.5102919 0.4179030 0.6074816
#> 21     no     11 0.6324443 0.5565326 0.7058450
#> 22    yes     11 0.5374015 0.4407360 0.6417605
#> 23     no     12 0.6573325 0.5766567 0.7358273
#> 24    yes     12 0.5642918 0.4613868 0.6746411
#> 25     no     13 0.6813846 0.5926957 0.7645128
#> 26    yes     13 0.5908091 0.4809534 0.7075421
#> 27     no     14 0.7045073 0.6086912 0.7879795
#> 28    yes     14 0.6168084 0.5005819 0.7392497
#> 29     no     15 0.7266252 0.6226293 0.8112647
#> 30    yes     15 0.6421564 0.5178671 0.7686458
#> 31     no     16 0.7476805 0.6351811 0.8333062
#> 32    yes     16 0.6667336 0.5331582 0.7956450
#> 33     no     17 0.7676327 0.6475489 0.8532392
#> 34    yes     17 0.6904365 0.5475946 0.8202305
#> 35     no     18 0.7864578 0.6597192 0.8711570
#> 36    yes     18 0.7131789 0.5627697 0.8424433
#> 37     no     19 0.8041469 0.6716792 0.8871763
#> 38    yes     19 0.7348919 0.5778280 0.8623711
#> 39     no     20 0.8207047 0.6834173 0.9014117
#> 40    yes     20 0.7555247 0.5930015 0.8801359
#> 41     no     21 0.8361480 0.6952745 0.9140033
#> 42    yes     21 0.7750433 0.6080444 0.8958833
#> 43     no     22 0.8505034 0.7073770 0.9251202
#> 44    yes     22 0.7934296 0.6228833 0.9096314
#> 45     no     23 0.8638059 0.7191934 0.9349021
#> 46    yes     23 0.8106802 0.6374932 0.9214543
#> 47     no     24 0.8760972 0.7307148 0.9434841
#> 48    yes     24 0.8268046 0.6518509 0.9318462
#> 49     no     25 0.8874238 0.7419336 0.9509939
#> 50    yes     25 0.8418238 0.6659352 0.9409514
#> 51     no     26 0.8978357 0.7528434 0.9576157
#> 52    yes     26 0.8557676 0.6797266 0.9489069
#> 53     no     27 0.9073851 0.7634395 0.9633988
#> 54    yes     27 0.8686741 0.6931834 0.9558404
#> 55     no     28 0.9161253 0.7737182 0.9684188
#> 56    yes     28 0.8805867 0.7053646 0.9618690
#> 57     no     29 0.9241097 0.7836772 0.9727698
#> 58    yes     29 0.8915537 0.7172592 0.9671029
#> 59     no     30 0.9313909 0.7933153 0.9765359
#> 60    yes     30 0.9016260 0.7288581 0.9716396
#> 61     no     31 0.9380204 0.8026324 0.9797919
#> 62    yes     31 0.9108563 0.7401535 0.9755664
#> 63     no     32 0.9440478 0.8116295 0.9826041
#> 64    yes     32 0.9192981 0.7511388 0.9789613
#> 65     no     33 0.9495206 0.8203085 0.9850310
#> 66    yes     33 0.9270045 0.7618091 0.9818933
#> 67     no     34 0.9544839 0.8286723 0.9871237
#> 68    yes     34 0.9340278 0.7721606 0.9844231
#> 69     no     35 0.9589802 0.8367245 0.9889272
#> 70    yes     35 0.9404188 0.7821907 0.9866043
#> 71     no     36 0.9630496 0.8444695 0.9904805
#> 72    yes     36 0.9462263 0.7918983 0.9884837
#> 73     no     37 0.9667293 0.8519123 0.9918177
#> 74    yes     37 0.9514970 0.8012829 0.9901020
#> 75     no     38 0.9700539 0.8590585 0.9929684
#> 76    yes     38 0.9562749 0.8103457 0.9914949
#> 77     no     39 0.9730556 0.8659144 0.9939582
#> 78    yes     39 0.9606016 0.8190884 0.9926932
#> 79     no     40 0.9757639 0.8724863 0.9948095
#> 80    yes     40 0.9645161 0.8275138 0.9937237
#> 81     no     41 0.9782061 0.8787814 0.9955413
#> 82    yes     41 0.9680546 0.8356256 0.9946097
#> 83     no     42 0.9804072 0.8848068 0.9961704
#> 84    yes     42 0.9712507 0.8434279 0.9953712
#> 85     no     43 0.9823899 0.8905701 0.9967110
#> 86    yes     43 0.9741357 0.8509260 0.9960256
#> 87     no     44 0.9841753 0.8960790 0.9971755
#> 88    yes     44 0.9767380 0.8581253 0.9965877
#> 89     no     45 0.9857822 0.9013414 0.9975745
#> 90    yes     45 0.9790842 0.8650321 0.9970706
#> 91     no     46 0.9872281 0.9063652 0.9979173
#> 92    yes     46 0.9811982 0.8716529 0.9974853
#> 93     no     47 0.9885287 0.9111585 0.9982118
#> 94    yes     47 0.9831023 0.8779947 0.9978415
#> 95     no     48 0.9896982 0.9157293 0.9984646
#> 96    yes     48 0.9848165 0.8840648 0.9981473
#> 97     no     49 0.9907496 0.9200855 0.9986818
#> 98    yes     49 0.9863593 0.8898707 0.9984098
#> 99     no     50 0.9916946 0.9242353 0.9988683
#> 100   yes     50 0.9877472 0.8954203 0.9986352
#> 101    no     51 0.9925437 0.9281864 0.9990284
#> 102   yes     51 0.9889955 0.9007214 0.9988287
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)