Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3074786 0.4329116
#> 2     yes      1 0.2814584 0.2159142 0.3603686
#> 3      no      2 0.3927672 0.3326739 0.4526487
#> 4     yes      2 0.3039568 0.2395823 0.3736657
#> 5      no      3 0.4189750 0.3624281 0.4757282
#> 6     yes      3 0.3274340 0.2638855 0.3865959
#> 7      no      4 0.4456481 0.3936165 0.4983376
#> 8     yes      4 0.3518080 0.2886376 0.4081486
#> 9      no      5 0.4726379 0.4245044 0.5274577
#> 10    yes      5 0.3769795 0.3152846 0.4333418
#> 11     no      6 0.4997887 0.4494178 0.5503386
#> 12    yes      6 0.4028326 0.3397541 0.4637393
#> 13     no      7 0.5269407 0.4704506 0.5775566
#> 14    yes      7 0.4292372 0.3640775 0.4952681
#> 15     no      8 0.5539343 0.4864430 0.6133198
#> 16    yes      8 0.4560509 0.3880016 0.5293923
#> 17     no      9 0.5806134 0.4976786 0.6501972
#> 18    yes      9 0.4831217 0.4063280 0.5678039
#> 19     no     10 0.6068295 0.5148665 0.6861351
#> 20    yes     10 0.5102919 0.4240182 0.6001311
#> 21     no     11 0.6324443 0.5321638 0.7196437
#> 22    yes     11 0.5374015 0.4391206 0.6320195
#> 23     no     12 0.6573325 0.5507134 0.7495539
#> 24    yes     12 0.5642918 0.4521788 0.6646739
#> 25     no     13 0.6813846 0.5678085 0.7763249
#> 26    yes     13 0.5908091 0.4655164 0.6958249
#> 27     no     14 0.7045073 0.5812985 0.8009571
#> 28    yes     14 0.6168084 0.4798118 0.7251871
#> 29     no     15 0.7266252 0.5946640 0.8232886
#> 30    yes     15 0.6421564 0.4941362 0.7531200
#> 31     no     16 0.7476805 0.6078863 0.8436004
#> 32    yes     16 0.6667336 0.5095620 0.7796134
#> 33     no     17 0.7676327 0.6209474 0.8619654
#> 34    yes     17 0.6904365 0.5256939 0.8039169
#> 35     no     18 0.7864578 0.6338306 0.8784811
#> 36    yes     18 0.7131789 0.5417531 0.8261374
#> 37     no     19 0.8041469 0.6465198 0.8937130
#> 38    yes     19 0.7348919 0.5574782 0.8463205
#> 39     no     20 0.8207047 0.6590003 0.9072791
#> 40    yes     20 0.7555247 0.5720925 0.8645441
#> 41     no     21 0.8361480 0.6712584 0.9192700
#> 42    yes     21 0.7750433 0.5865649 0.8809102
#> 43     no     22 0.8505034 0.6832817 0.9298299
#> 44    yes     22 0.7934296 0.6008709 0.8955372
#> 45     no     23 0.8638059 0.6950589 0.9390999
#> 46    yes     23 0.8106802 0.6149874 0.9085537
#> 47     no     24 0.8760972 0.7065801 0.9472146
#> 48    yes     24 0.8268046 0.6288927 0.9200924
#> 49     no     25 0.8874238 0.7178365 0.9543007
#> 50    yes     25 0.8418238 0.6425665 0.9302865
#> 51     no     26 0.8978357 0.7288206 0.9604750
#> 52    yes     26 0.8557676 0.6559903 0.9392655
#> 53     no     27 0.9073851 0.7395260 0.9658450
#> 54    yes     27 0.8686741 0.6691472 0.9471534
#> 55     no     28 0.9161253 0.7499475 0.9705077
#> 56    yes     28 0.8805867 0.6820224 0.9540666
#> 57     no     29 0.9241097 0.7600812 0.9745507
#> 58    yes     29 0.8915537 0.6946023 0.9601134
#> 59     no     30 0.9313909 0.7699242 0.9780519
#> 60    yes     30 0.9016260 0.7068757 0.9653927
#> 61     no     31 0.9380204 0.7794745 0.9810807
#> 62    yes     31 0.9108563 0.7188327 0.9699950
#> 63     no     32 0.9440478 0.7887314 0.9836985
#> 64    yes     32 0.9192981 0.7304653 0.9740015
#> 65     no     33 0.9495206 0.7976948 0.9859592
#> 66    yes     33 0.9270045 0.7417673 0.9774853
#> 67     no     34 0.9544839 0.8063658 0.9879102
#> 68    yes     34 0.9340278 0.7527339 0.9805115
#> 69     no     35 0.9589802 0.8147461 0.9895930
#> 70    yes     35 0.9404188 0.7633619 0.9831379
#> 71     no     36 0.9630496 0.8228382 0.9910437
#> 72    yes     36 0.9462263 0.7736495 0.9854273
#> 73     no     37 0.9667293 0.8306453 0.9922937
#> 74    yes     37 0.9514970 0.7835965 0.9874767
#> 75     no     38 0.9700539 0.8381712 0.9933704
#> 76    yes     38 0.9562749 0.7932035 0.9892410
#> 77     no     39 0.9730556 0.8454202 0.9942975
#> 78    yes     39 0.9606016 0.8023800 0.9907591
#> 79     no     40 0.9757639 0.8523973 0.9950987
#> 80    yes     40 0.9645161 0.8107448 0.9920646
#> 81     no     41 0.9782061 0.8591077 0.9957989
#> 82    yes     41 0.9680546 0.8188294 0.9931869
#> 83     no     42 0.9804072 0.8655571 0.9963994
#> 84    yes     42 0.9712507 0.8266366 0.9941513
#> 85     no     43 0.9823899 0.8717515 0.9969143
#> 86    yes     43 0.9741357 0.8341700 0.9949799
#> 87     no     44 0.9841753 0.8776971 0.9973558
#> 88    yes     44 0.9767380 0.8414337 0.9956916
#> 89     no     45 0.9857822 0.8834005 0.9977342
#> 90    yes     45 0.9790842 0.8484323 0.9963027
#> 91     no     46 0.9872281 0.8888682 0.9980586
#> 92    yes     46 0.9811982 0.8551706 0.9968274
#> 93     no     47 0.9885287 0.8941071 0.9983366
#> 94    yes     47 0.9831023 0.8616540 0.9972779
#> 95     no     48 0.9896982 0.8991241 0.9985749
#> 96    yes     48 0.9848165 0.8678882 0.9976645
#> 97     no     49 0.9907496 0.9039262 0.9987791
#> 98    yes     49 0.9863593 0.8738788 0.9979963
#> 99     no     50 0.9916946 0.9085202 0.9989540
#> 100   yes     50 0.9877472 0.8796321 0.9982810
#> 101    no     51 0.9925437 0.9129131 0.9991039
#> 102   yes     51 0.9889955 0.8851543 0.9985253
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)