Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3185485 0.4336729
#> 2     yes      1 0.2814584 0.2138648 0.3653162
#> 3      no      2 0.3927672 0.3466876 0.4508077
#> 4     yes      2 0.3039568 0.2357409 0.3798954
#> 5      no      3 0.4189750 0.3742159 0.4756532
#> 6     yes      3 0.3274340 0.2601033 0.3964790
#> 7      no      4 0.4456481 0.3973734 0.5038539
#> 8     yes      4 0.3518080 0.2880827 0.4198471
#> 9      no      5 0.4726379 0.4210469 0.5341066
#> 10    yes      5 0.3769795 0.3127603 0.4442721
#> 11     no      6 0.4997887 0.4420743 0.5661041
#> 12    yes      6 0.4028326 0.3383223 0.4678195
#> 13     no      7 0.5269407 0.4566768 0.5993576
#> 14    yes      7 0.4292372 0.3578309 0.4901079
#> 15     no      8 0.5539343 0.4712810 0.6350472
#> 16    yes      8 0.4560509 0.3787671 0.5180828
#> 17     no      9 0.5806134 0.4863065 0.6752163
#> 18    yes      9 0.4831217 0.4015275 0.5553543
#> 19     no     10 0.6068295 0.5029348 0.7124792
#> 20    yes     10 0.5102919 0.4268705 0.5923189
#> 21     no     11 0.6324443 0.5201345 0.7462402
#> 22    yes     11 0.5374015 0.4504239 0.6284982
#> 23     no     12 0.6573325 0.5372867 0.7772133
#> 24    yes     12 0.5642918 0.4736934 0.6677230
#> 25     no     13 0.6813846 0.5544079 0.8036998
#> 26    yes     13 0.5908091 0.4944020 0.7046266
#> 27     no     14 0.7045073 0.5714932 0.8289383
#> 28    yes     14 0.6168084 0.5125230 0.7389638
#> 29     no     15 0.7266252 0.5884095 0.8515276
#> 30    yes     15 0.6421564 0.5289102 0.7700461
#> 31     no     16 0.7476805 0.6051190 0.8715903
#> 32    yes     16 0.6667336 0.5432705 0.7974732
#> 33     no     17 0.7676327 0.6215859 0.8892876
#> 34    yes     17 0.6904365 0.5575575 0.8223536
#> 35     no     18 0.7864578 0.6377766 0.9047647
#> 36    yes     18 0.7131789 0.5717477 0.8447450
#> 37     no     19 0.8041469 0.6536598 0.9180861
#> 38    yes     19 0.7348919 0.5858184 0.8647540
#> 39     no     20 0.8207047 0.6688570 0.9296822
#> 40    yes     20 0.7555247 0.5995479 0.8825216
#> 41     no     21 0.8361480 0.6815116 0.9397383
#> 42    yes     21 0.7750433 0.6131365 0.8982114
#> 43     no     22 0.8505034 0.6939490 0.9484304
#> 44    yes     22 0.7934296 0.6271280 0.9119992
#> 45     no     23 0.8638059 0.7065069 0.9559227
#> 46    yes     23 0.8106802 0.6409066 0.9240641
#> 47     no     24 0.8760972 0.7187523 0.9623655
#> 48    yes     24 0.8268046 0.6544516 0.9345825
#> 49     no     25 0.8874238 0.7306746 0.9678944
#> 50    yes     25 0.8418238 0.6683237 0.9437234
#> 51     no     26 0.8978357 0.7422654 0.9726310
#> 52    yes     26 0.8557676 0.6819169 0.9516454
#> 53     no     27 0.9073851 0.7535176 0.9766830
#> 54    yes     27 0.8686741 0.6951973 0.9584949
#> 55     no     28 0.9161253 0.7644260 0.9801450
#> 56    yes     28 0.8805867 0.7081494 0.9644051
#> 57     no     29 0.9241097 0.7749870 0.9831000
#> 58    yes     29 0.8915537 0.7207600 0.9694960
#> 59     no     30 0.9313909 0.7851985 0.9856200
#> 60    yes     30 0.9016260 0.7322250 0.9738748
#> 61     no     31 0.9380204 0.7950600 0.9877676
#> 62    yes     31 0.9108563 0.7432066 0.9776363
#> 63     no     32 0.9440478 0.8045720 0.9895966
#> 64    yes     32 0.9192981 0.7538840 0.9808644
#> 65     no     33 0.9495206 0.8137366 0.9911537
#> 66    yes     33 0.9270045 0.7642529 0.9836320
#> 67     no     34 0.9544839 0.8225568 0.9924786
#> 68    yes     34 0.9340278 0.7743104 0.9860033
#> 69     no     35 0.9589802 0.8310368 0.9936057
#> 70    yes     35 0.9404188 0.7840546 0.9880336
#> 71     no     36 0.9630496 0.8391815 0.9945642
#> 72    yes     36 0.9462263 0.7934849 0.9897712
#> 73     no     37 0.9667293 0.8469969 0.9953793
#> 74    yes     37 0.9514970 0.8026017 0.9912576
#> 75     no     38 0.9700539 0.8544896 0.9960721
#> 76    yes     38 0.9562749 0.8114064 0.9925287
#> 77     no     39 0.9730556 0.8616666 0.9966611
#> 78    yes     39 0.9606016 0.8199012 0.9936154
#> 79     no     40 0.9757639 0.8685357 0.9971618
#> 80    yes     40 0.9645161 0.8280895 0.9945443
#> 81     no     41 0.9782061 0.8751051 0.9975873
#> 82    yes     41 0.9680546 0.8359749 0.9953380
#> 83     no     42 0.9804072 0.8813832 0.9979489
#> 84    yes     42 0.9712507 0.8435621 0.9960163
#> 85     no     43 0.9823899 0.8873788 0.9982562
#> 86    yes     43 0.9741357 0.8508562 0.9965959
#> 87     no     44 0.9841753 0.8931009 0.9985175
#> 88    yes     44 0.9767380 0.8578630 0.9970910
#> 89     no     45 0.9857822 0.8985586 0.9987395
#> 90    yes     45 0.9790842 0.8645886 0.9975140
#> 91     no     46 0.9872281 0.9037611 0.9989282
#> 92    yes     46 0.9811982 0.8710396 0.9978754
#> 93     no     47 0.9885287 0.9087175 0.9990885
#> 94    yes     47 0.9831023 0.8772228 0.9981841
#> 95     no     48 0.9896982 0.9134371 0.9992249
#> 96    yes     48 0.9848165 0.8831455 0.9984479
#> 97     no     49 0.9907496 0.9179291 0.9993408
#> 98    yes     49 0.9863593 0.8888149 0.9986733
#> 99     no     50 0.9916946 0.9222023 0.9994393
#> 100   yes     50 0.9877472 0.8942387 0.9988658
#> 101    no     51 0.9925437 0.9262658 0.9995230
#> 102   yes     51 0.9889955 0.8994244 0.9990321
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)