Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3104787 0.4260765
#> 2     yes      1 0.2814584 0.2226581 0.3690519
#> 3      no      2 0.3927672 0.3389048 0.4446670
#> 4     yes      2 0.3039568 0.2458138 0.3906912
#> 5      no      3 0.4189750 0.3653465 0.4668478
#> 6     yes      3 0.3274340 0.2708674 0.4107501
#> 7      no      4 0.4456481 0.3947553 0.5017316
#> 8     yes      4 0.3518080 0.2951004 0.4321330
#> 9      no      5 0.4726379 0.4217434 0.5358303
#> 10    yes      5 0.3769795 0.3147501 0.4537719
#> 11     no      6 0.4997887 0.4462666 0.5647597
#> 12    yes      6 0.4028326 0.3336106 0.4755468
#> 13     no      7 0.5269407 0.4685608 0.5931700
#> 14    yes      7 0.4292372 0.3515206 0.5097896
#> 15     no      8 0.5539343 0.4924611 0.6261926
#> 16    yes      8 0.4560509 0.3685872 0.5458093
#> 17     no      9 0.5806134 0.5123485 0.6607038
#> 18    yes      9 0.4831217 0.3859894 0.5760543
#> 19     no     10 0.6068295 0.5323755 0.6952598
#> 20    yes     10 0.5102919 0.4039514 0.6053182
#> 21     no     11 0.6324443 0.5521737 0.7282526
#> 22    yes     11 0.5374015 0.4224458 0.6348818
#> 23     no     12 0.6573325 0.5701301 0.7589254
#> 24    yes     12 0.5642918 0.4411602 0.6663299
#> 25     no     13 0.6813846 0.5879029 0.7871578
#> 26    yes     13 0.5908091 0.4600435 0.6982603
#> 27     no     14 0.7045073 0.6054482 0.8129059
#> 28    yes     14 0.6168084 0.4790422 0.7284731
#> 29     no     15 0.7266252 0.6227247 0.8361920
#> 30    yes     15 0.6421564 0.4981019 0.7571237
#> 31     no     16 0.7476805 0.6396611 0.8570919
#> 32    yes     16 0.6667336 0.5171670 0.7840903
#> 33     no     17 0.7676327 0.6555016 0.8757228
#> 34    yes     17 0.6904365 0.5361824 0.8092083
#> 35     no     18 0.7864578 0.6710041 0.8922303
#> 36    yes     18 0.7131789 0.5550932 0.8320168
#> 37     no     19 0.8041469 0.6861431 0.9067781
#> 38    yes     19 0.7348919 0.5738457 0.8525826
#> 39     no     20 0.8207047 0.7008961 0.9195382
#> 40    yes     20 0.7555247 0.5924098 0.8710094
#> 41     no     21 0.8361480 0.7152436 0.9306840
#> 42    yes     21 0.7750433 0.6107390 0.8874265
#> 43     no     22 0.8505034 0.7291689 0.9403847
#> 44    yes     22 0.7934296 0.6287605 0.9019802
#> 45     no     23 0.8638059 0.7426583 0.9488012
#> 46    yes     23 0.8106802 0.6464305 0.9148248
#> 47     no     24 0.8760972 0.7557010 0.9560837
#> 48    yes     24 0.8268046 0.6637093 0.9261172
#> 49     no     25 0.8874238 0.7682891 0.9623702
#> 50    yes     25 0.8418238 0.6805609 0.9360113
#> 51     no     26 0.8978357 0.7804170 0.9677847
#> 52    yes     26 0.8557676 0.6969535 0.9448834
#> 53     no     27 0.9073851 0.7920821 0.9723984
#> 54    yes     27 0.8686741 0.7128598 0.9526584
#> 55     no     28 0.9161253 0.8032837 0.9763667
#> 56    yes     28 0.8805867 0.7282566 0.9593832
#> 57     no     29 0.9241097 0.8138444 0.9797756
#> 58    yes     29 0.8915537 0.7420853 0.9651871
#> 59     no     30 0.9313909 0.8239423 0.9827009
#> 60    yes     30 0.9016260 0.7554914 0.9701869
#> 61     no     31 0.9380204 0.8336044 0.9852089
#> 62    yes     31 0.9108563 0.7685164 0.9744872
#> 63     no     32 0.9440478 0.8428373 0.9873190
#> 64    yes     32 0.9192981 0.7810466 0.9781809
#> 65     no     33 0.9495206 0.8516491 0.9891220
#> 66    yes     33 0.9270045 0.7930791 0.9813497
#> 67     no     34 0.9544839 0.8600488 0.9906711
#> 68    yes     34 0.9340278 0.8046135 0.9840656
#> 69     no     35 0.9589802 0.8680465 0.9920014
#> 70    yes     35 0.9404188 0.8155511 0.9863912
#> 71     no     36 0.9630496 0.8756802 0.9931432
#> 72    yes     36 0.9462263 0.8254061 0.9883813
#> 73     no     37 0.9667293 0.8829921 0.9941230
#> 74    yes     37 0.9514970 0.8348412 0.9900831
#> 75     no     38 0.9700539 0.8899277 0.9949635
#> 76    yes     38 0.9562749 0.8438628 0.9915376
#> 77     no     39 0.9730556 0.8965000 0.9956843
#> 78    yes     39 0.9606016 0.8524788 0.9927802
#> 79     no     40 0.9757639 0.9027225 0.9963024
#> 80    yes     40 0.9645161 0.8606979 0.9938414
#> 81     no     41 0.9782061 0.9086206 0.9968322
#> 82    yes     41 0.9680546 0.8685296 0.9947473
#> 83     no     42 0.9804072 0.9142332 0.9972863
#> 84    yes     42 0.9712507 0.8759845 0.9955205
#> 85     no     43 0.9823899 0.9195313 0.9976754
#> 86    yes     43 0.9741357 0.8830737 0.9961803
#> 87     no     44 0.9841753 0.9245287 0.9980089
#> 88    yes     44 0.9767380 0.8898085 0.9967431
#> 89     no     45 0.9857822 0.9293063 0.9982945
#> 90    yes     45 0.9790842 0.8962011 0.9972232
#> 91     no     46 0.9872281 0.9338168 0.9985393
#> 92    yes     46 0.9811982 0.9022635 0.9976327
#> 93     no     47 0.9885287 0.9380585 0.9987490
#> 94    yes     47 0.9831023 0.9081167 0.9979818
#> 95     no     48 0.9896982 0.9420453 0.9989286
#> 96    yes     48 0.9848165 0.9137302 0.9982795
#> 97     no     49 0.9907496 0.9457902 0.9990824
#> 98    yes     49 0.9863593 0.9190310 0.9985334
#> 99     no     50 0.9916946 0.9493061 0.9992142
#> 100   yes     50 0.9877472 0.9240328 0.9987490
#> 101    no     51 0.9925437 0.9526054 0.9993271
#> 102   yes     51 0.9889955 0.9287494 0.9989320
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)