Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3062377 0.4353159
#> 2     yes      1 0.2814584 0.2159151 0.3485928
#> 3      no      2 0.3927672 0.3424692 0.4519065
#> 4     yes      2 0.3039568 0.2381146 0.3705549
#> 5      no      3 0.4189750 0.3781058 0.4761109
#> 6     yes      3 0.3274340 0.2633431 0.3924491
#> 7      no      4 0.4456481 0.4097034 0.4981581
#> 8     yes      4 0.3518080 0.2913081 0.4148011
#> 9      no      5 0.4726379 0.4380913 0.5194773
#> 10    yes      5 0.3769795 0.3201084 0.4401541
#> 11     no      6 0.4997887 0.4622530 0.5535293
#> 12    yes      6 0.4028326 0.3443213 0.4678204
#> 13     no      7 0.5269407 0.4871658 0.5915480
#> 14    yes      7 0.4292372 0.3647231 0.5091374
#> 15     no      8 0.5539343 0.5071018 0.6339905
#> 16    yes      8 0.4560509 0.3873806 0.5486405
#> 17     no      9 0.5806134 0.5265706 0.6720023
#> 18    yes      9 0.4831217 0.4044014 0.5861232
#> 19     no     10 0.6068295 0.5416307 0.7041587
#> 20    yes     10 0.5102919 0.4196448 0.6233535
#> 21     no     11 0.6324443 0.5563756 0.7342422
#> 22    yes     11 0.5374015 0.4381608 0.6580577
#> 23     no     12 0.6573325 0.5726937 0.7637915
#> 24    yes     12 0.5642918 0.4569194 0.6912247
#> 25     no     13 0.6813846 0.5897831 0.7940544
#> 26    yes     13 0.5908091 0.4758001 0.7286822
#> 27     no     14 0.7045073 0.6066578 0.8213600
#> 28    yes     14 0.6168084 0.4947496 0.7631533
#> 29     no     15 0.7266252 0.6232811 0.8457461
#> 30    yes     15 0.6421564 0.5137133 0.7944808
#> 31     no     16 0.7476805 0.6396185 0.8673374
#> 32    yes     16 0.6667336 0.5326367 0.8226283
#> 33     no     17 0.7676327 0.6556382 0.8862712
#> 34    yes     17 0.6904365 0.5514658 0.8476605
#> 35     no     18 0.7864578 0.6713111 0.9033121
#> 36    yes     18 0.7131789 0.5701476 0.8697200
#> 37     no     19 0.8041469 0.6866111 0.9180964
#> 38    yes     19 0.7348919 0.5886307 0.8890039
#> 39     no     20 0.8207047 0.7015149 0.9307933
#> 40    yes     20 0.7555247 0.6068662 0.9057431
#> 41     no     21 0.8361480 0.7160025 0.9416469
#> 42    yes     21 0.7750433 0.6248076 0.9201848
#> 43     no     22 0.8505034 0.7300569 0.9508882
#> 44    yes     22 0.7934296 0.6424117 0.9325788
#> 45     no     23 0.8638059 0.7436643 0.9587301
#> 46    yes     23 0.8106802 0.6596388 0.9431672
#> 47     no     24 0.8760972 0.7568137 0.9653654
#> 48    yes     24 0.8268046 0.6764531 0.9521781
#> 49     no     25 0.8874238 0.7694972 0.9709662
#> 50    yes     25 0.8418238 0.6928226 0.9598213
#> 51     no     26 0.8978357 0.7817095 0.9756841
#> 52    yes     26 0.8557676 0.7081201 0.9662863
#> 53     no     27 0.9073851 0.7934480 0.9796515
#> 54    yes     27 0.8686741 0.7222277 0.9717416
#> 55     no     28 0.9161253 0.8047126 0.9829828
#> 56    yes     28 0.8805867 0.7359081 0.9763359
#> 57     no     29 0.9241097 0.8155053 0.9857766
#> 58    yes     29 0.8915537 0.7491488 0.9801985
#> 59     no     30 0.9313909 0.8258303 0.9881173
#> 60    yes     30 0.9016260 0.7619401 0.9834413
#> 61     no     31 0.9380204 0.8356935 0.9900767
#> 62    yes     31 0.9108563 0.7742750 0.9861606
#> 63     no     32 0.9440478 0.8451027 0.9917157
#> 64    yes     32 0.9192981 0.7861492 0.9884385
#> 65     no     33 0.9495206 0.8540028 0.9930859
#> 66    yes     33 0.9270045 0.7975610 0.9903452
#> 67     no     34 0.9544839 0.8624418 0.9942308
#> 68    yes     34 0.9340278 0.8085109 0.9919400
#> 69     no     35 0.9589802 0.8704671 0.9951870
#> 70    yes     35 0.9404188 0.8190014 0.9932732
#> 71     no     36 0.9630496 0.8780904 0.9959853
#> 72    yes     36 0.9462263 0.8290371 0.9943871
#> 73     no     37 0.9667293 0.8853241 0.9966517
#> 74    yes     37 0.9514970 0.8387449 0.9953174
#> 75     no     38 0.9700539 0.8921814 0.9972078
#> 76    yes     38 0.9562749 0.8480384 0.9960942
#> 77     no     39 0.9730556 0.8986755 0.9976718
#> 78    yes     39 0.9606016 0.8568860 0.9967425
#> 79     no     40 0.9757639 0.9048202 0.9980588
#> 80    yes     40 0.9645161 0.8652985 0.9972835
#> 81     no     41 0.9782061 0.9106294 0.9983815
#> 82    yes     41 0.9680546 0.8732878 0.9977348
#> 83     no     42 0.9804072 0.9161169 0.9986507
#> 84    yes     42 0.9712507 0.8808666 0.9981113
#> 85     no     43 0.9823899 0.9212966 0.9988752
#> 86    yes     43 0.9741357 0.8880486 0.9984253
#> 87     no     44 0.9841753 0.9261822 0.9990623
#> 88    yes     44 0.9767380 0.8948475 0.9986872
#> 89     no     45 0.9857822 0.9307873 0.9992184
#> 90    yes     45 0.9790842 0.9010268 0.9989056
#> 91     no     46 0.9872281 0.9351253 0.9993485
#> 92    yes     46 0.9811982 0.9068660 0.9990877
#> 93     no     47 0.9885287 0.9392092 0.9994569
#> 94    yes     47 0.9831023 0.9123939 0.9992395
#> 95     no     48 0.9896982 0.9430516 0.9995473
#> 96    yes     48 0.9848165 0.9176233 0.9993661
#> 97     no     49 0.9907496 0.9466650 0.9996227
#> 98    yes     49 0.9863593 0.9225668 0.9994716
#> 99     no     50 0.9916946 0.9500531 0.9996855
#> 100   yes     50 0.9877472 0.9272369 0.9995595
#> 101    no     51 0.9925437 0.9531362 0.9997379
#> 102   yes     51 0.9889955 0.9316460 0.9996329
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)