Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3058667 0.4270242
#> 2     yes      1 0.2814584 0.2093322 0.3616414
#> 3      no      2 0.3927672 0.3317764 0.4492256
#> 4     yes      2 0.3039568 0.2287151 0.3758291
#> 5      no      3 0.4189750 0.3608091 0.4730050
#> 6     yes      3 0.3274340 0.2505825 0.3921540
#> 7      no      4 0.4456481 0.3897797 0.4994403
#> 8     yes      4 0.3518080 0.2738175 0.4154135
#> 9      no      5 0.4726379 0.4143974 0.5287157
#> 10    yes      5 0.3769795 0.2983659 0.4440903
#> 11     no      6 0.4997887 0.4347138 0.5642619
#> 12    yes      6 0.4028326 0.3246672 0.4789686
#> 13     no      7 0.5269407 0.4617677 0.5973579
#> 14    yes      7 0.4292372 0.3496472 0.5066639
#> 15     no      8 0.5539343 0.4873063 0.6282610
#> 16    yes      8 0.4560509 0.3726196 0.5400130
#> 17     no      9 0.5806134 0.5094023 0.6577764
#> 18    yes      9 0.4831217 0.3954549 0.5723687
#> 19     no     10 0.6068295 0.5301986 0.6861114
#> 20    yes     10 0.5102919 0.4188509 0.6091977
#> 21     no     11 0.6324443 0.5508888 0.7152031
#> 22    yes     11 0.5374015 0.4426168 0.6468184
#> 23     no     12 0.6573325 0.5699966 0.7446176
#> 24    yes     12 0.5642918 0.4606807 0.6822903
#> 25     no     13 0.6813846 0.5880375 0.7719575
#> 26    yes     13 0.5908091 0.4773145 0.7155048
#> 27     no     14 0.7045073 0.6058454 0.7968275
#> 28    yes     14 0.6168084 0.4977120 0.7465211
#> 29     no     15 0.7266252 0.6230410 0.8195457
#> 30    yes     15 0.6421564 0.5214576 0.7752003
#> 31     no     16 0.7476805 0.6394854 0.8409560
#> 32    yes     16 0.6667336 0.5431968 0.8014784
#> 33     no     17 0.7676327 0.6559219 0.8618338
#> 34    yes     17 0.6904365 0.5615912 0.8253577
#> 35     no     18 0.7864578 0.6715390 0.8803582
#> 36    yes     18 0.7131789 0.5798210 0.8468949
#> 37     no     19 0.8041469 0.6863033 0.8966937
#> 38    yes     19 0.7348919 0.5978380 0.8656444
#> 39     no     20 0.8207047 0.7006986 0.9108910
#> 40    yes     20 0.7555247 0.6155964 0.8823238
#> 41     no     21 0.8361480 0.7147066 0.9232842
#> 42    yes     21 0.7750433 0.6330531 0.8971760
#> 43     no     22 0.8505034 0.7283120 0.9340758
#> 44    yes     22 0.7934296 0.6501681 0.9103420
#> 45     no     23 0.8638059 0.7415018 0.9434399
#> 46    yes     23 0.8106802 0.6669049 0.9219670
#> 47     no     24 0.8760972 0.7542659 0.9515408
#> 48    yes     24 0.8268046 0.6832305 0.9324477
#> 49     no     25 0.8874238 0.7665967 0.9585306
#> 50    yes     25 0.8418238 0.6991160 0.9416671
#> 51     no     26 0.8978357 0.7784889 0.9645480
#> 52    yes     26 0.8557676 0.7145362 0.9496965
#> 53     no     27 0.9073851 0.7899397 0.9697185
#> 54    yes     27 0.8686741 0.7296574 0.9566720
#> 55     no     28 0.9161253 0.8009485 0.9741539
#> 56    yes     28 0.8805867 0.7443223 0.9627184
#> 57     no     29 0.9241097 0.8115165 0.9779533
#> 58    yes     29 0.8915537 0.7584526 0.9679496
#> 59     no     30 0.9313909 0.8216470 0.9812041
#> 60    yes     30 0.9016260 0.7720387 0.9724679
#> 61     no     31 0.9380204 0.8313447 0.9839827
#> 62    yes     31 0.9108563 0.7850743 0.9763649
#> 63     no     32 0.9440478 0.8406160 0.9863556
#> 64    yes     32 0.9192981 0.7975572 0.9797219
#> 65     no     33 0.9495206 0.8494686 0.9883806
#> 66    yes     33 0.9270045 0.8082036 0.9826106
#> 67     no     34 0.9544839 0.8579114 0.9900809
#> 68    yes     34 0.9340278 0.8183747 0.9850942
#> 69     no     35 0.9589802 0.8659541 0.9915179
#> 70    yes     35 0.9404188 0.8281207 0.9872277
#> 71     no     36 0.9630496 0.8736075 0.9927483
#> 72    yes     36 0.9462263 0.8374471 0.9890592
#> 73     no     37 0.9667293 0.8808829 0.9938014
#> 74    yes     37 0.9514970 0.8463609 0.9906306
#> 75     no     38 0.9700539 0.8877922 0.9947023
#> 76    yes     38 0.9562749 0.8548699 0.9919782
#> 77     no     39 0.9730556 0.8943478 0.9954729
#> 78    yes     39 0.9606016 0.8629834 0.9931333
#> 79     no     40 0.9757639 0.9005624 0.9961318
#> 80    yes     40 0.9645161 0.8707113 0.9941231
#> 81     no     41 0.9782061 0.9064488 0.9966952
#> 82    yes     41 0.9680546 0.8780642 0.9949709
#> 83     no     42 0.9804072 0.9120199 0.9971767
#> 84    yes     42 0.9712507 0.8851157 0.9956969
#> 85     no     43 0.9823899 0.9172887 0.9975883
#> 86    yes     43 0.9741357 0.8918267 0.9963186
#> 87     no     44 0.9841753 0.9222681 0.9979399
#> 88    yes     44 0.9767380 0.8981902 0.9968507
#> 89     no     45 0.9857822 0.9269709 0.9982404
#> 90    yes     45 0.9790842 0.9042188 0.9973061
#> 91     no     46 0.9872281 0.9314097 0.9984971
#> 92    yes     46 0.9811982 0.9099256 0.9976958
#> 93     no     47 0.9885287 0.9355969 0.9987165
#> 94    yes     47 0.9831023 0.9153236 0.9980292
#> 95     no     48 0.9896982 0.9395444 0.9989038
#> 96    yes     48 0.9848165 0.9204259 0.9983145
#> 97     no     49 0.9907496 0.9432641 0.9990638
#> 98    yes     49 0.9863593 0.9252452 0.9985585
#> 99     no     50 0.9916946 0.9467674 0.9992005
#> 100   yes     50 0.9877472 0.9297943 0.9987673
#> 101    no     51 0.9925437 0.9500654 0.9993172
#> 102   yes     51 0.9889955 0.9340858 0.9989458
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)