Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3139940 0.4224937
#> 2     yes      1 0.2814584 0.2140394 0.3493007
#> 3      no      2 0.3927672 0.3422128 0.4451918
#> 4     yes      2 0.3039568 0.2338594 0.3662938
#> 5      no      3 0.4189750 0.3698300 0.4712443
#> 6     yes      3 0.3274340 0.2593957 0.3857989
#> 7      no      4 0.4456481 0.3924592 0.4977669
#> 8     yes      4 0.3518080 0.2866442 0.4145146
#> 9      no      5 0.4726379 0.4159695 0.5281897
#> 10    yes      5 0.3769795 0.3137748 0.4387419
#> 11     no      6 0.4997887 0.4390119 0.5639332
#> 12    yes      6 0.4028326 0.3414710 0.4635574
#> 13     no      7 0.5269407 0.4572154 0.6005814
#> 14    yes      7 0.4292372 0.3666022 0.4926017
#> 15     no      8 0.5539343 0.4765211 0.6361393
#> 16    yes      8 0.4560509 0.3932278 0.5218515
#> 17     no      9 0.5806134 0.4924838 0.6702686
#> 18    yes      9 0.4831217 0.4140601 0.5528011
#> 19     no     10 0.6068295 0.5065930 0.7024280
#> 20    yes     10 0.5102919 0.4350908 0.5896923
#> 21     no     11 0.6324443 0.5223888 0.7348135
#> 22    yes     11 0.5374015 0.4566638 0.6253771
#> 23     no     12 0.6573325 0.5381406 0.7663268
#> 24    yes     12 0.5642918 0.4783999 0.6665641
#> 25     no     13 0.6813846 0.5543336 0.7950124
#> 26    yes     13 0.5908091 0.4996322 0.7063412
#> 27     no     14 0.7045073 0.5705006 0.8209847
#> 28    yes     14 0.6168084 0.5162630 0.7425437
#> 29     no     15 0.7266252 0.5865133 0.8442958
#> 30    yes     15 0.6421564 0.5353774 0.7733268
#> 31     no     16 0.7476805 0.6023381 0.8650559
#> 32    yes     16 0.6667336 0.5512644 0.8004735
#> 33     no     17 0.7676327 0.6179431 0.8834167
#> 34    yes     17 0.6904365 0.5681646 0.8249463
#> 35     no     18 0.7864578 0.6332983 0.8995570
#> 36    yes     18 0.7131789 0.5849548 0.8469337
#> 37     no     19 0.8041469 0.6483758 0.9136701
#> 38    yes     19 0.7348919 0.6015417 0.8665487
#> 39     no     20 0.8207047 0.6631502 0.9259539
#> 40    yes     20 0.7555247 0.6178879 0.8839400
#> 41     no     21 0.8361480 0.6775985 0.9366029
#> 42    yes     21 0.7750433 0.6339583 0.8992781
#> 43     no     22 0.8505034 0.6917002 0.9458030
#> 44    yes     22 0.7934296 0.6497207 0.9127443
#> 45     no     23 0.8638059 0.7054377 0.9537280
#> 46    yes     23 0.8106802 0.6651457 0.9243398
#> 47     no     24 0.8760972 0.7187956 0.9605375
#> 48    yes     24 0.8268046 0.6802069 0.9348473
#> 49     no     25 0.8874238 0.7317614 0.9663760
#> 50    yes     25 0.8418238 0.6948813 0.9442320
#> 51     no     26 0.8978357 0.7443251 0.9713728
#> 52    yes     26 0.8557676 0.7091489 0.9523282
#> 53     no     27 0.9073851 0.7564791 0.9756429
#> 54    yes     27 0.8686741 0.7229932 0.9592950
#> 55     no     28 0.9161253 0.7682181 0.9792871
#> 56    yes     28 0.8805867 0.7364007 0.9652768
#> 57     no     29 0.9241097 0.7795392 0.9823938
#> 58    yes     29 0.8915537 0.7493609 0.9704035
#> 59     no     30 0.9313909 0.7904413 0.9850400
#> 60    yes     30 0.9016260 0.7618667 0.9747904
#> 61     no     31 0.9380204 0.8009255 0.9872921
#> 62    yes     31 0.9108563 0.7739133 0.9785391
#> 63     no     32 0.9440478 0.8109945 0.9892077
#> 64    yes     32 0.9192981 0.7854552 0.9817390
#> 65     no     33 0.9495206 0.8206526 0.9908361
#> 66    yes     33 0.9270045 0.7959250 0.9844678
#> 67     no     34 0.9544839 0.8299055 0.9922200
#> 68    yes     34 0.9340278 0.8057305 0.9867929
#> 69     no     35 0.9589802 0.8387603 0.9933955
#> 70    yes     35 0.9404188 0.8151678 0.9887729
#> 71     no     36 0.9630496 0.8472248 0.9943938
#> 72    yes     36 0.9462263 0.8242405 0.9904580
#> 73     no     37 0.9667293 0.8553083 0.9952414
#> 74    yes     37 0.9514970 0.8329534 0.9918914
#> 75     no     38 0.9700539 0.8630203 0.9959610
#> 76    yes     38 0.9562749 0.8413120 0.9931104
#> 77     no     39 0.9730556 0.8703714 0.9965718
#> 78    yes     39 0.9606016 0.8493229 0.9941466
#> 79     no     40 0.9757639 0.8773724 0.9970901
#> 80    yes     40 0.9645161 0.8569932 0.9950273
#> 81     no     41 0.9782061 0.8840348 0.9975301
#> 82    yes     41 0.9680546 0.8643308 0.9957756
#> 83     no     42 0.9804072 0.8903700 0.9979034
#> 84    yes     42 0.9712507 0.8713442 0.9964114
#> 85     no     43 0.9823899 0.8963899 0.9982203
#> 86    yes     43 0.9741357 0.8780421 0.9969515
#> 87     no     44 0.9841753 0.9021063 0.9984891
#> 88    yes     44 0.9767380 0.8844338 0.9974103
#> 89     no     45 0.9857822 0.9074375 0.9987173
#> 90    yes     45 0.9790842 0.8905287 0.9978000
#> 91     no     46 0.9872281 0.9123600 0.9989110
#> 92    yes     46 0.9811982 0.8963366 0.9981310
#> 93     no     47 0.9885287 0.9170845 0.9990753
#> 94    yes     47 0.9831023 0.9018671 0.9984122
#> 95     no     48 0.9896982 0.9215740 0.9992148
#> 96    yes     48 0.9848165 0.9071303 0.9986510
#> 97     no     49 0.9907496 0.9258381 0.9993332
#> 98    yes     49 0.9863593 0.9121361 0.9988538
#> 99     no     50 0.9916946 0.9298861 0.9994337
#> 100   yes     50 0.9877472 0.9168942 0.9990260
#> 101    no     51 0.9925437 0.9337272 0.9995190
#> 102   yes     51 0.9889955 0.9214145 0.9991724
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)