Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3097003 0.4408566
#> 2     yes      1 0.2814584 0.2206396 0.3464326
#> 3      no      2 0.3927672 0.3360473 0.4574517
#> 4     yes      2 0.3039568 0.2432983 0.3662364
#> 5      no      3 0.4189750 0.3615717 0.4825431
#> 6     yes      3 0.3274340 0.2683643 0.3861668
#> 7      no      4 0.4456481 0.3937292 0.5004807
#> 8     yes      4 0.3518080 0.2907871 0.4101891
#> 9      no      5 0.4726379 0.4206856 0.5223799
#> 10    yes      5 0.3769795 0.3115575 0.4405404
#> 11     no      6 0.4997887 0.4451255 0.5531514
#> 12    yes      6 0.4028326 0.3357718 0.4683805
#> 13     no      7 0.5269407 0.4687049 0.5888860
#> 14    yes      7 0.4292372 0.3600093 0.5010197
#> 15     no      8 0.5539343 0.4924241 0.6236990
#> 16    yes      8 0.4560509 0.3803229 0.5385976
#> 17     no      9 0.5806134 0.5147828 0.6572678
#> 18    yes      9 0.4831217 0.3988049 0.5757337
#> 19     no     10 0.6068295 0.5373625 0.6897304
#> 20    yes     10 0.5102919 0.4158343 0.6117615
#> 21     no     11 0.6324443 0.5602224 0.7220648
#> 22    yes     11 0.5374015 0.4332724 0.6462934
#> 23     no     12 0.6573325 0.5803196 0.7522395
#> 24    yes     12 0.5642918 0.4508778 0.6793637
#> 25     no     13 0.6813846 0.5986752 0.7801354
#> 26    yes     13 0.5908091 0.4686945 0.7107264
#> 27     no     14 0.7045073 0.6153310 0.8057014
#> 28    yes     14 0.6168084 0.4887612 0.7401951
#> 29     no     15 0.7266252 0.6333052 0.8289462
#> 30    yes     15 0.6421564 0.5088641 0.7676441
#> 31     no     16 0.7476805 0.6502013 0.8499281
#> 32    yes     16 0.6667336 0.5272932 0.7940315
#> 33     no     17 0.7676327 0.6662165 0.8687437
#> 34    yes     17 0.6904365 0.5436200 0.8188780
#> 35     no     18 0.7864578 0.6818568 0.8855180
#> 36    yes     18 0.7131789 0.5598543 0.8413275
#> 37     no     19 0.8041469 0.6965050 0.9003944
#> 38    yes     19 0.7348919 0.5763232 0.8614642
#> 39     no     20 0.8207047 0.7101179 0.9135265
#> 40    yes     20 0.7555247 0.5924866 0.8794092
#> 41     no     21 0.8361480 0.7233620 0.9250713
#> 42    yes     21 0.7750433 0.6066332 0.8953090
#> 43     no     22 0.8505034 0.7363439 0.9351842
#> 44    yes     22 0.7934296 0.6206014 0.9093247
#> 45     no     23 0.8638059 0.7489755 0.9440148
#> 46    yes     23 0.8106802 0.6343704 0.9216243
#> 47     no     24 0.8760972 0.7611971 0.9517044
#> 48    yes     24 0.8268046 0.6479209 0.9323584
#> 49     no     25 0.8874238 0.7730029 0.9583844
#> 50    yes     25 0.8418238 0.6612348 0.9415969
#> 51     no     26 0.8978357 0.7843890 0.9641753
#> 52    yes     26 0.8557676 0.6742957 0.9496384
#> 53     no     27 0.9073851 0.7953539 0.9693671
#> 54    yes     27 0.8686741 0.6870884 0.9566206
#> 55     no     28 0.9161253 0.8058979 0.9738508
#> 56    yes     28 0.8805867 0.6995996 0.9626702
#> 57     no     29 0.9241097 0.8160229 0.9776426
#> 58    yes     29 0.8915537 0.7118172 0.9679020
#> 59     no     30 0.9313909 0.8257326 0.9808669
#> 60    yes     30 0.9016260 0.7237309 0.9724195
#> 61     no     31 0.9380204 0.8350320 0.9836328
#> 62    yes     31 0.9108563 0.7353321 0.9763093
#> 63     no     32 0.9440478 0.8435068 0.9860034
#> 64    yes     32 0.9192981 0.7466133 0.9795706
#> 65     no     33 0.9495206 0.8513551 0.9880339
#> 66    yes     33 0.9270045 0.7575690 0.9823909
#> 67     no     34 0.9544839 0.8588756 0.9897612
#> 68    yes     34 0.9340278 0.7681947 0.9848277
#> 69     no     35 0.9589802 0.8660755 0.9912282
#> 70    yes     35 0.9404188 0.7784876 0.9869484
#> 71     no     36 0.9630496 0.8729623 0.9924861
#> 72    yes     36 0.9462263 0.7884462 0.9888160
#> 73     no     37 0.9667293 0.8795443 0.9935642
#> 74    yes     37 0.9514970 0.7980700 0.9904190
#> 75     no     38 0.9700539 0.8858297 0.9944880
#> 76    yes     38 0.9562749 0.8073599 0.9917941
#> 77     no     39 0.9730556 0.8918274 0.9952709
#> 78    yes     39 0.9606016 0.8163177 0.9929733
#> 79     no     40 0.9757639 0.8975464 0.9959345
#> 80    yes     40 0.9645161 0.8249464 0.9939841
#> 81     no     41 0.9782061 0.9029959 0.9965052
#> 82    yes     41 0.9680546 0.8332497 0.9948502
#> 83     no     42 0.9804072 0.9081850 0.9969960
#> 84    yes     42 0.9712507 0.8412321 0.9955922
#> 85     no     43 0.9823899 0.9131232 0.9974180
#> 86    yes     43 0.9741357 0.8491215 0.9962276
#> 87     no     44 0.9841753 0.9178197 0.9977808
#> 88    yes     44 0.9767380 0.8567583 0.9967718
#> 89     no     45 0.9857822 0.9222838 0.9980927
#> 90    yes     45 0.9790842 0.8640701 0.9972377
#> 91     no     46 0.9872281 0.9265247 0.9983607
#> 92    yes     46 0.9811982 0.8710645 0.9976365
#> 93     no     47 0.9885287 0.9305515 0.9985940
#> 94    yes     47 0.9831023 0.8777497 0.9979778
#> 95     no     48 0.9896982 0.9343731 0.9987966
#> 96    yes     48 0.9848165 0.8841341 0.9982699
#> 97     no     49 0.9907496 0.9379984 0.9989701
#> 98    yes     49 0.9863593 0.8902265 0.9985199
#> 99     no     50 0.9916946 0.9414358 0.9991186
#> 100   yes     50 0.9877472 0.8960361 0.9987338
#> 101    no     51 0.9925437 0.9446938 0.9992457
#> 102   yes     51 0.9889955 0.9015720 0.9989168
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)