Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3040029 0.4275546
#> 2     yes      1 0.2814584 0.2222909 0.3535699
#> 3      no      2 0.3927672 0.3336686 0.4509035
#> 4     yes      2 0.3039568 0.2443584 0.3767256
#> 5      no      3 0.4189750 0.3643538 0.4746549
#> 6     yes      3 0.3274340 0.2681694 0.4002474
#> 7      no      4 0.4456481 0.3915554 0.5007079
#> 8     yes      4 0.3518080 0.2919522 0.4242385
#> 9      no      5 0.4726379 0.4208198 0.5266566
#> 10    yes      5 0.3769795 0.3126154 0.4528697
#> 11     no      6 0.4997887 0.4442734 0.5556628
#> 12    yes      6 0.4028326 0.3331168 0.4782456
#> 13     no      7 0.5269407 0.4615768 0.5883913
#> 14    yes      7 0.4292372 0.3546520 0.5090622
#> 15     no      8 0.5539343 0.4815491 0.6251895
#> 16    yes      8 0.4560509 0.3767951 0.5443876
#> 17     no      9 0.5806134 0.5006907 0.6651927
#> 18    yes      9 0.4831217 0.3994657 0.5824513
#> 19     no     10 0.6068295 0.5168855 0.7021760
#> 20    yes     10 0.5102919 0.4225742 0.6222224
#> 21     no     11 0.6324443 0.5330438 0.7359012
#> 22    yes     11 0.5374015 0.4460238 0.6636517
#> 23     no     12 0.6573325 0.5491320 0.7670592
#> 24    yes     12 0.5642918 0.4687711 0.7026917
#> 25     no     13 0.6813846 0.5651167 0.7955492
#> 26    yes     13 0.5908091 0.4901252 0.7389626
#> 27     no     14 0.7045073 0.5809658 0.8213550
#> 28    yes     14 0.6168084 0.5117594 0.7722211
#> 29     no     15 0.7266252 0.5966481 0.8445307
#> 30    yes     15 0.6421564 0.5329235 0.8015225
#> 31     no     16 0.7476805 0.6121337 0.8651856
#> 32    yes     16 0.6667336 0.5515276 0.8275815
#> 33     no     17 0.7676327 0.6273945 0.8838237
#> 34    yes     17 0.6904365 0.5686877 0.8508550
#> 35     no     18 0.7864578 0.6424040 0.9003864
#> 36    yes     18 0.7131789 0.5869057 0.8714741
#> 37     no     19 0.8041469 0.6571375 0.9148167
#> 38    yes     19 0.7348919 0.6054949 0.8896121
#> 39     no     20 0.8207047 0.6715726 0.9273265
#> 40    yes     20 0.7555247 0.6237746 0.9054676
#> 41     no     21 0.8361480 0.6856889 0.9381243
#> 42    yes     21 0.7750433 0.6396977 0.9192522
#> 43     no     22 0.8505034 0.6994682 0.9474094
#> 44    yes     22 0.7934296 0.6540669 0.9311794
#> 45     no     23 0.8638059 0.7128946 0.9553680
#> 46    yes     23 0.8106802 0.6681607 0.9414568
#> 47     no     24 0.8760972 0.7259545 0.9621707
#> 48    yes     24 0.8268046 0.6819597 0.9502813
#> 49     no     25 0.8874238 0.7386365 0.9679717
#> 50    yes     25 0.8418238 0.6954468 0.9578351
#> 51     no     26 0.8978357 0.7509313 0.9729083
#> 52    yes     26 0.8557676 0.7086066 0.9642842
#> 53     no     27 0.9073851 0.7628318 0.9771022
#> 54    yes     27 0.8686741 0.7214256 0.9697780
#> 55     no     28 0.9161253 0.7743331 0.9806598
#> 56    yes     28 0.8805867 0.7338928 0.9744491
#> 57     no     29 0.9241097 0.7854320 0.9836741
#> 58    yes     29 0.8915537 0.7459985 0.9784143
#> 59     no     30 0.9313909 0.7961273 0.9862251
#> 60    yes     30 0.9016260 0.7577355 0.9817755
#> 61     no     31 0.9380204 0.8064196 0.9883823
#> 62    yes     31 0.9108563 0.7690981 0.9846215
#> 63     no     32 0.9440478 0.8163108 0.9902050
#> 64    yes     32 0.9192981 0.7800825 0.9870289
#> 65     no     33 0.9495206 0.8258045 0.9917442
#> 66    yes     33 0.9270045 0.7906867 0.9890636
#> 67     no     34 0.9544839 0.8349057 0.9930432
#> 68    yes     34 0.9340278 0.8009101 0.9907821
#> 69     no     35 0.9589802 0.8436202 0.9941390
#> 70    yes     35 0.9404188 0.8107537 0.9922326
#> 71     no     36 0.9630496 0.8519553 0.9950630
#> 72    yes     36 0.9462263 0.8202198 0.9934564
#> 73     no     37 0.9667293 0.8599189 0.9958420
#> 74    yes     37 0.9514970 0.8293122 0.9944884
#> 75     no     38 0.9700539 0.8675198 0.9964985
#> 76    yes     38 0.9562749 0.8380354 0.9953584
#> 77     no     39 0.9730556 0.8747552 0.9970516
#> 78    yes     39 0.9606016 0.8463955 0.9960916
#> 79     no     40 0.9757639 0.8815004 0.9975176
#> 80    yes     40 0.9645161 0.8543989 0.9967093
#> 81     no     41 0.9782061 0.8879288 0.9979101
#> 82    yes     41 0.9680546 0.8620533 0.9972297
#> 83     no     42 0.9804072 0.8940503 0.9982406
#> 84    yes     42 0.9712507 0.8693668 0.9976679
#> 85     no     43 0.9823899 0.8998752 0.9985189
#> 86    yes     43 0.9741357 0.8763481 0.9980370
#> 87     no     44 0.9841753 0.9054137 0.9987533
#> 88    yes     44 0.9767380 0.8830064 0.9983477
#> 89     no     45 0.9857822 0.9106763 0.9989506
#> 90    yes     45 0.9790842 0.8893515 0.9986093
#> 91     no     46 0.9872281 0.9156733 0.9991167
#> 92    yes     46 0.9811982 0.8953931 0.9988296
#> 93     no     47 0.9885287 0.9204152 0.9992565
#> 94    yes     47 0.9831023 0.9011415 0.9990149
#> 95     no     48 0.9896982 0.9249124 0.9993742
#> 96    yes     48 0.9848165 0.9066069 0.9991710
#> 97     no     49 0.9907496 0.9291749 0.9994733
#> 98    yes     49 0.9863593 0.9117996 0.9993023
#> 99     no     50 0.9916946 0.9332130 0.9995567
#> 100   yes     50 0.9877472 0.9167301 0.9994129
#> 101    no     51 0.9925437 0.9370364 0.9996269
#> 102   yes     51 0.9889955 0.9214087 0.9995059
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)