Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3129314 0.4279791
#> 2     yes      1 0.2814584 0.2141649 0.3588850
#> 3      no      2 0.3927672 0.3461839 0.4472139
#> 4     yes      2 0.3039568 0.2404218 0.3771635
#> 5      no      3 0.4189750 0.3761403 0.4679120
#> 6     yes      3 0.3274340 0.2643309 0.3958037
#> 7      no      4 0.4456481 0.4037139 0.4974492
#> 8     yes      4 0.3518080 0.2864873 0.4170231
#> 9      no      5 0.4726379 0.4282853 0.5265287
#> 10    yes      5 0.3769795 0.3098027 0.4416086
#> 11     no      6 0.4997887 0.4471325 0.5549240
#> 12    yes      6 0.4028326 0.3376409 0.4677375
#> 13     no      7 0.5269407 0.4648719 0.5829621
#> 14    yes      7 0.4292372 0.3661592 0.4975059
#> 15     no      8 0.5539343 0.4819814 0.6130263
#> 16    yes      8 0.4560509 0.3875277 0.5312291
#> 17     no      9 0.5806134 0.4991331 0.6475634
#> 18    yes      9 0.4831217 0.4044148 0.5673818
#> 19     no     10 0.6068295 0.5162867 0.6783172
#> 20    yes     10 0.5102919 0.4215420 0.6009891
#> 21     no     11 0.6324443 0.5308749 0.7074945
#> 22    yes     11 0.5374015 0.4424152 0.6336713
#> 23     no     12 0.6573325 0.5438423 0.7366117
#> 24    yes     12 0.5642918 0.4584855 0.6651679
#> 25     no     13 0.6813846 0.5545833 0.7637897
#> 26    yes     13 0.5908091 0.4747721 0.6952587
#> 27     no     14 0.7045073 0.5652721 0.7899603
#> 28    yes     14 0.6168084 0.4923362 0.7239018
#> 29     no     15 0.7266252 0.5758989 0.8149740
#> 30    yes     15 0.6421564 0.5099135 0.7528409
#> 31     no     16 0.7476805 0.5864539 0.8375477
#> 32    yes     16 0.6667336 0.5274507 0.7792364
#> 33     no     17 0.7676327 0.5969277 0.8578356
#> 34    yes     17 0.6904365 0.5417972 0.8039193
#> 35     no     18 0.7864578 0.6073111 0.8759646
#> 36    yes     18 0.7131789 0.5522146 0.8278298
#> 37     no     19 0.8041469 0.6175954 0.8920721
#> 38    yes     19 0.7348919 0.5625891 0.8493710
#> 39     no     20 0.8207047 0.6277721 0.9063111
#> 40    yes     20 0.7555247 0.5729115 0.8686448
#> 41     no     21 0.8361480 0.6378332 0.9188420
#> 42    yes     21 0.7750433 0.5831727 0.8857842
#> 43     no     22 0.8505034 0.6477711 0.9298263
#> 44    yes     22 0.7934296 0.5933639 0.9009422
#> 45     no     23 0.8638059 0.6575786 0.9394216
#> 46    yes     23 0.8106802 0.6034765 0.9142833
#> 47     no     24 0.8760972 0.6672491 0.9477783
#> 48    yes     24 0.8268046 0.6135022 0.9259753
#> 49     no     25 0.8874238 0.6767763 0.9550371
#> 50    yes     25 0.8418238 0.6234330 0.9361838
#> 51     no     26 0.8978357 0.6863498 0.9613279
#> 52    yes     26 0.8557676 0.6332612 0.9450679
#> 53     no     27 0.9073851 0.6966597 0.9667690
#> 54    yes     27 0.8686741 0.6429794 0.9527777
#> 55     no     28 0.9161253 0.7067762 0.9714671
#> 56    yes     28 0.8805867 0.6525808 0.9594518
#> 57     no     29 0.9241097 0.7166932 0.9755177
#> 58    yes     29 0.8915537 0.6620586 0.9652170
#> 59     no     30 0.9313909 0.7264052 0.9790055
#> 60    yes     30 0.9016260 0.6714068 0.9701880
#> 61     no     31 0.9380204 0.7359075 0.9820056
#> 62    yes     31 0.9108563 0.6806196 0.9744673
#> 63     no     32 0.9440478 0.7451962 0.9846079
#> 64    yes     32 0.9192981 0.6899316 0.9781462
#> 65     no     33 0.9495206 0.7542677 0.9868422
#> 66    yes     33 0.9270045 0.6991343 0.9813052
#> 67     no     34 0.9544839 0.7631195 0.9887560
#> 68    yes     34 0.9340278 0.7081771 0.9840150
#> 69     no     35 0.9589802 0.7717494 0.9903941
#> 70    yes     35 0.9404188 0.7170555 0.9863375
#> 71     no     36 0.9630496 0.7801559 0.9917955
#> 72    yes     36 0.9462263 0.7257657 0.9883265
#> 73     no     37 0.9667293 0.7883381 0.9929939
#> 74    yes     37 0.9514970 0.7343042 0.9900289
#> 75     no     38 0.9700539 0.7962957 0.9940183
#> 76    yes     38 0.9562749 0.7426680 0.9914852
#> 77     no     39 0.9730556 0.8040289 0.9948937
#> 78    yes     39 0.9606016 0.7516410 0.9927302
#> 79     no     40 0.9757639 0.8115383 0.9956416
#> 80    yes     40 0.9645161 0.7605610 0.9937932
#> 81     no     41 0.9782061 0.8188252 0.9962803
#> 82    yes     41 0.9680546 0.7692591 0.9947015
#> 83     no     42 0.9804072 0.8258910 0.9968257
#> 84    yes     42 0.9712507 0.7777338 0.9954775
#> 85     no     43 0.9823899 0.8327378 0.9972914
#> 86    yes     43 0.9741357 0.7859839 0.9961403
#> 87     no     44 0.9841753 0.8393680 0.9976889
#> 88    yes     44 0.9767380 0.7940089 0.9967063
#> 89     no     45 0.9857822 0.8457842 0.9980282
#> 90    yes     45 0.9790842 0.8018091 0.9971895
#> 91     no     46 0.9872281 0.8519896 0.9983177
#> 92    yes     46 0.9811982 0.8093848 0.9976020
#> 93     no     47 0.9885287 0.8579874 0.9985648
#> 94    yes     47 0.9831023 0.8167373 0.9979541
#> 95     no     48 0.9896982 0.8637811 0.9987757
#> 96    yes     48 0.9848165 0.8238679 0.9982546
#> 97     no     49 0.9907496 0.8693747 0.9989556
#> 98    yes     49 0.9863593 0.8307786 0.9985110
#> 99     no     50 0.9916946 0.8747720 0.9991091
#> 100   yes     50 0.9877472 0.8374717 0.9987298
#> 101    no     51 0.9925437 0.8799772 0.9992400
#> 102   yes     51 0.9889955 0.8439498 0.9989164
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)