Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3184858 0.4323962
#> 2     yes      1 0.2814584 0.2336091 0.3633599
#> 3      no      2 0.3927672 0.3503601 0.4555646
#> 4     yes      2 0.3039568 0.2551077 0.3829181
#> 5      no      3 0.4189750 0.3794161 0.4789269
#> 6     yes      3 0.3274340 0.2755719 0.4004703
#> 7      no      4 0.4456481 0.4016643 0.5044606
#> 8     yes      4 0.3518080 0.2958737 0.4191901
#> 9      no      5 0.4726379 0.4227815 0.5303328
#> 10    yes      5 0.3769795 0.3181654 0.4426632
#> 11     no      6 0.4997887 0.4449927 0.5648560
#> 12    yes      6 0.4028326 0.3395737 0.4705878
#> 13     no      7 0.5269407 0.4654971 0.5975123
#> 14    yes      7 0.4292372 0.3597775 0.5018334
#> 15     no      8 0.5539343 0.4824561 0.6293263
#> 16    yes      8 0.4560509 0.3795440 0.5356826
#> 17     no      9 0.5806134 0.4994560 0.6600565
#> 18    yes      9 0.4831217 0.4003831 0.5692021
#> 19     no     10 0.6068295 0.5164574 0.6894964
#> 20    yes     10 0.5102919 0.4204053 0.6048845
#> 21     no     11 0.6324443 0.5334208 0.7174807
#> 22    yes     11 0.5374015 0.4387358 0.6414585
#> 23     no     12 0.6573325 0.5503716 0.7470299
#> 24    yes     12 0.5642918 0.4554398 0.6764007
#> 25     no     13 0.6813846 0.5673032 0.7765214
#> 26    yes     13 0.5908091 0.4722459 0.7116153
#> 27     no     14 0.7045073 0.5840788 0.8043260
#> 28    yes     14 0.6168084 0.4902080 0.7448431
#> 29     no     15 0.7266252 0.6006613 0.8294284
#> 30    yes     15 0.6421564 0.5087939 0.7754476
#> 31     no     16 0.7476805 0.6161052 0.8519005
#> 32    yes     16 0.6667336 0.5278869 0.8033471
#> 33     no     17 0.7676327 0.6312998 0.8718667
#> 34    yes     17 0.6904365 0.5468947 0.8285434
#> 35     no     18 0.7864578 0.6462383 0.8894877
#> 36    yes     18 0.7131789 0.5634146 0.8511064
#> 37     no     19 0.8041469 0.6608967 0.9049472
#> 38    yes     19 0.7348919 0.5798515 0.8711589
#> 39     no     20 0.8207047 0.6752529 0.9184404
#> 40    yes     20 0.7555247 0.5961138 0.8888606
#> 41     no     21 0.8361480 0.6887294 0.9301641
#> 42    yes     21 0.7750433 0.6121680 0.9043946
#> 43     no     22 0.8505034 0.7022084 0.9403105
#> 44    yes     22 0.7934296 0.6279823 0.9179556
#> 45     no     23 0.8638059 0.7161265 0.9490621
#> 46    yes     23 0.8106802 0.6435270 0.9297407
#> 47     no     24 0.8760972 0.7293346 0.9565887
#> 48    yes     24 0.8268046 0.6587744 0.9399420
#> 49     no     25 0.8874238 0.7419769 0.9630455
#> 50    yes     25 0.8418238 0.6738131 0.9487424
#> 51     no     26 0.8978357 0.7542277 0.9685726
#> 52    yes     26 0.8557676 0.6886069 0.9563121
#> 53     no     27 0.9073851 0.7660800 0.9732953
#> 54    yes     27 0.8686741 0.7030242 0.9628067
#> 55     no     28 0.9161253 0.7773890 0.9773244
#> 56    yes     28 0.8805867 0.7170466 0.9683669
#> 57     no     29 0.9241097 0.7880085 0.9807571
#> 58    yes     29 0.8915537 0.7306590 0.9731185
#> 59     no     30 0.9313909 0.7982527 0.9836784
#> 60    yes     30 0.9016260 0.7432194 0.9771725
#> 61     no     31 0.9380204 0.8081223 0.9861621
#> 62    yes     31 0.9108563 0.7549560 0.9806269
#> 63     no     32 0.9440478 0.8176509 0.9882721
#> 64    yes     32 0.9192981 0.7663248 0.9835669
#> 65     no     33 0.9495206 0.8268250 0.9900634
#> 66    yes     33 0.9270045 0.7773218 0.9860667
#> 67     no     34 0.9544839 0.8356302 0.9915832
#> 68    yes     34 0.9340278 0.7876875 0.9881905
#> 69     no     35 0.9589802 0.8440719 0.9928720
#> 70    yes     35 0.9404188 0.7974631 0.9899937
#> 71     no     36 0.9630496 0.8521567 0.9939645
#> 72    yes     36 0.9462263 0.8068978 0.9915237
#> 73     no     37 0.9667293 0.8598917 0.9948904
#> 74    yes     37 0.9514970 0.8159996 0.9928212
#> 75     no     38 0.9700539 0.8672849 0.9956747
#> 76    yes     38 0.9562749 0.8249083 0.9939213
#> 77     no     39 0.9730556 0.8743448 0.9963390
#> 78    yes     39 0.9606016 0.8334723 0.9948535
#> 79     no     40 0.9757639 0.8810805 0.9969015
#> 80    yes     40 0.9645161 0.8416963 0.9956432
#> 81     no     41 0.9782061 0.8875015 0.9973777
#> 82    yes     41 0.9680546 0.8496162 0.9963122
#> 83     no     42 0.9804072 0.8936174 0.9977809
#> 84    yes     42 0.9712507 0.8576384 0.9968786
#> 85     no     43 0.9823899 0.8994384 0.9981221
#> 86    yes     43 0.9741357 0.8653006 0.9973583
#> 87     no     44 0.9841753 0.9049747 0.9984109
#> 88    yes     44 0.9767380 0.8726117 0.9977643
#> 89     no     45 0.9857822 0.9102364 0.9986554
#> 90    yes     45 0.9790842 0.8794791 0.9981080
#> 91     no     46 0.9872281 0.9152340 0.9988622
#> 92    yes     46 0.9811982 0.8859801 0.9983989
#> 93     no     47 0.9885287 0.9199776 0.9990372
#> 94    yes     47 0.9831023 0.8921734 0.9986452
#> 95     no     48 0.9896982 0.9244776 0.9991854
#> 96    yes     48 0.9848165 0.8980687 0.9988535
#> 97     no     49 0.9907496 0.9287440 0.9993107
#> 98    yes     49 0.9863593 0.9036764 0.9990299
#> 99     no     50 0.9916946 0.9327868 0.9994167
#> 100   yes     50 0.9877472 0.9090067 0.9991791
#> 101    no     51 0.9925437 0.9366158 0.9995065
#> 102   yes     51 0.9889955 0.9140699 0.9993053
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)