Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3025795 0.4338553
#> 2     yes      1 0.2814584 0.2057580 0.3686142
#> 3      no      2 0.3927672 0.3390539 0.4523015
#> 4     yes      2 0.3039568 0.2319293 0.3911413
#> 5      no      3 0.4189750 0.3700260 0.4699975
#> 6     yes      3 0.3274340 0.2557436 0.4135156
#> 7      no      4 0.4456481 0.4020975 0.5005309
#> 8     yes      4 0.3518080 0.2811092 0.4344020
#> 9      no      5 0.4726379 0.4291052 0.5319496
#> 10    yes      5 0.3769795 0.3043495 0.4603866
#> 11     no      6 0.4997887 0.4531831 0.5631165
#> 12    yes      6 0.4028326 0.3278773 0.4888740
#> 13     no      7 0.5269407 0.4732678 0.5937922
#> 14    yes      7 0.4292372 0.3497684 0.5145095
#> 15     no      8 0.5539343 0.4947723 0.6244657
#> 16    yes      8 0.4560509 0.3749327 0.5455924
#> 17     no      9 0.5806134 0.5127162 0.6575681
#> 18    yes      9 0.4831217 0.3997685 0.5772571
#> 19     no     10 0.6068295 0.5308283 0.6943239
#> 20    yes     10 0.5102919 0.4215346 0.6080319
#> 21     no     11 0.6324443 0.5494517 0.7306643
#> 22    yes     11 0.5374015 0.4391922 0.6482900
#> 23     no     12 0.6573325 0.5678495 0.7654935
#> 24    yes     12 0.5642918 0.4553711 0.6865402
#> 25     no     13 0.6813846 0.5847704 0.7974457
#> 26    yes     13 0.5908091 0.4718825 0.7217139
#> 27     no     14 0.7045073 0.6014941 0.8260313
#> 28    yes     14 0.6168084 0.4887934 0.7536203
#> 29     no     15 0.7266252 0.6179844 0.8513326
#> 30    yes     15 0.6421564 0.5057286 0.7837020
#> 31     no     16 0.7476805 0.6342072 0.8724903
#> 32    yes     16 0.6667336 0.5226476 0.8114217
#> 33     no     17 0.7676327 0.6501309 0.8904789
#> 34    yes     17 0.6904365 0.5395099 0.8358609
#> 35     no     18 0.7864578 0.6657264 0.9059290
#> 36    yes     18 0.7131789 0.5562757 0.8576678
#> 37     no     19 0.8041469 0.6811093 0.9193870
#> 38    yes     19 0.7348919 0.5730753 0.8769848
#> 39     no     20 0.8207047 0.6955879 0.9310570
#> 40    yes     20 0.7555247 0.5910623 0.8939871
#> 41     no     21 0.8361480 0.7086043 0.9411375
#> 42    yes     21 0.7750433 0.6088076 0.9088687
#> 43     no     22 0.8505034 0.7214469 0.9501676
#> 44    yes     22 0.7934296 0.6262687 0.9218309
#> 45     no     23 0.8638059 0.7339341 0.9579120
#> 46    yes     23 0.8106802 0.6428875 0.9330742
#> 47     no     24 0.8760972 0.7460565 0.9644967
#> 48    yes     24 0.8268046 0.6579861 0.9427912
#> 49     no     25 0.8874238 0.7578065 0.9700826
#> 50    yes     25 0.8418238 0.6713851 0.9511634
#> 51     no     26 0.8978357 0.7691787 0.9748118
#> 52    yes     26 0.8557676 0.6845074 0.9587101
#> 53     no     27 0.9073851 0.7801693 0.9788092
#> 54    yes     27 0.8686741 0.6973362 0.9651518
#> 55     no     28 0.9161253 0.7907763 0.9821832
#> 56    yes     28 0.8805867 0.7098571 0.9706189
#> 57     no     29 0.9241097 0.8009994 0.9850278
#> 58    yes     29 0.8915537 0.7222179 0.9752498
#> 59     no     30 0.9313909 0.8108397 0.9874238
#> 60    yes     30 0.9016260 0.7346579 0.9791662
#> 61     no     31 0.9380204 0.8203000 0.9894401
#> 62    yes     31 0.9108563 0.7467242 0.9824736
#> 63     no     32 0.9440478 0.8293840 0.9911358
#> 64    yes     32 0.9192981 0.7584092 0.9852637
#> 65     no     33 0.9495206 0.8380968 0.9925610
#> 66    yes     33 0.9270045 0.7697072 0.9876149
#> 67     no     34 0.9544839 0.8464444 0.9937584
#> 68    yes     34 0.9340278 0.7806150 0.9895948
#> 69     no     35 0.9589802 0.8544338 0.9947639
#> 70    yes     35 0.9404188 0.7911309 0.9912608
#> 71     no     36 0.9630496 0.8620728 0.9956080
#> 72    yes     36 0.9462263 0.8012553 0.9926618
#> 73     no     37 0.9667293 0.8693697 0.9963165
#> 74    yes     37 0.9514970 0.8109902 0.9938396
#> 75     no     38 0.9700539 0.8763337 0.9969109
#> 76    yes     38 0.9562749 0.8203390 0.9948291
#> 77     no     39 0.9730556 0.8829741 0.9974096
#> 78    yes     39 0.9606016 0.8293065 0.9956604
#> 79     no     40 0.9757639 0.8893008 0.9978279
#> 80    yes     40 0.9645161 0.8374267 0.9963584
#> 81     no     41 0.9782061 0.8953240 0.9981787
#> 82    yes     41 0.9680546 0.8451870 0.9969444
#> 83     no     42 0.9804072 0.9010539 0.9984729
#> 84    yes     42 0.9712507 0.8526342 0.9974363
#> 85     no     43 0.9823899 0.9065010 0.9987197
#> 86    yes     43 0.9741357 0.8597751 0.9978491
#> 87     no     44 0.9841753 0.9116758 0.9989265
#> 88    yes     44 0.9767380 0.8666167 0.9981956
#> 89     no     45 0.9857822 0.9165889 0.9991000
#> 90    yes     45 0.9790842 0.8731666 0.9984862
#> 91     no     46 0.9872281 0.9212506 0.9992454
#> 92    yes     46 0.9811982 0.8794328 0.9987301
#> 93     no     47 0.9885287 0.9256714 0.9993673
#> 94    yes     47 0.9831023 0.8854235 0.9989347
#> 95     no     48 0.9896982 0.9298614 0.9994695
#> 96    yes     48 0.9848165 0.8913497 0.9991064
#> 97     no     49 0.9907496 0.9338308 0.9995552
#> 98    yes     49 0.9863593 0.8971608 0.9992504
#> 99     no     50 0.9916946 0.9375893 0.9996271
#> 100   yes     50 0.9877472 0.9026936 0.9993712
#> 101    no     51 0.9925437 0.9411466 0.9996873
#> 102   yes     51 0.9889955 0.9079580 0.9994725
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)