Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3054969 0.4219587
#> 2     yes      1 0.2814584 0.2156396 0.3487163
#> 3      no      2 0.3927672 0.3352643 0.4425945
#> 4     yes      2 0.3039568 0.2395563 0.3684390
#> 5      no      3 0.4189750 0.3632813 0.4633734
#> 6     yes      3 0.3274340 0.2649509 0.3933649
#> 7      no      4 0.4456481 0.3908597 0.4895755
#> 8     yes      4 0.3518080 0.2920040 0.4175562
#> 9      no      5 0.4726379 0.4183048 0.5221421
#> 10    yes      5 0.3769795 0.3161620 0.4406419
#> 11     no      6 0.4997887 0.4393098 0.5572296
#> 12    yes      6 0.4028326 0.3413554 0.4661672
#> 13     no      7 0.5269407 0.4572797 0.5934423
#> 14    yes      7 0.4292372 0.3639679 0.4977085
#> 15     no      8 0.5539343 0.4754090 0.6236235
#> 16    yes      8 0.4560509 0.3832618 0.5314876
#> 17     no      9 0.5806134 0.4957343 0.6576418
#> 18    yes      9 0.4831217 0.4037522 0.5658745
#> 19     no     10 0.6068295 0.5160747 0.6927812
#> 20    yes     10 0.5102919 0.4245882 0.6010670
#> 21     no     11 0.6324443 0.5363625 0.7269112
#> 22    yes     11 0.5374015 0.4489507 0.6374243
#> 23     no     12 0.6573325 0.5547026 0.7576670
#> 24    yes     12 0.5642918 0.4720408 0.6722747
#> 25     no     13 0.6813846 0.5723276 0.7859759
#> 26    yes     13 0.5908091 0.4904100 0.7069542
#> 27     no     14 0.7045073 0.5880326 0.8115277
#> 28    yes     14 0.6168084 0.5085732 0.7396829
#> 29     no     15 0.7266252 0.6035605 0.8346592
#> 30    yes     15 0.6421564 0.5267131 0.7684345
#> 31     no     16 0.7476805 0.6188825 0.8554529
#> 32    yes     16 0.6667336 0.5447812 0.7948913
#> 33     no     17 0.7676327 0.6339715 0.8740216
#> 34    yes     17 0.6904365 0.5619025 0.8195972
#> 35     no     18 0.7864578 0.6488022 0.8905060
#> 36    yes     18 0.7131789 0.5776783 0.8421772
#> 37     no     19 0.8041469 0.6633510 0.9050638
#> 38    yes     19 0.7348919 0.5932976 0.8621892
#> 39     no     20 0.8207047 0.6775967 0.9178609
#> 40    yes     20 0.7555247 0.6087308 0.8800215
#> 41     no     21 0.8361480 0.6915199 0.9290651
#> 42    yes     21 0.7750433 0.6239497 0.8958215
#> 43     no     22 0.8505034 0.7051035 0.9388400
#> 44    yes     22 0.7934296 0.6389279 0.9100382
#> 45     no     23 0.8638059 0.7183326 0.9473419
#> 46    yes     23 0.8106802 0.6536406 0.9227578
#> 47     no     24 0.8760972 0.7311945 0.9548390
#> 48    yes     24 0.8268046 0.6680651 0.9338098
#> 49     no     25 0.8874238 0.7436786 0.9614908
#> 50    yes     25 0.8418238 0.6821807 0.9433774
#> 51     no     26 0.8978357 0.7557765 0.9671966
#> 52    yes     26 0.8557676 0.6959797 0.9516338
#> 53     no     27 0.9073851 0.7674819 0.9720816
#> 54    yes     27 0.8686741 0.7095491 0.9587389
#> 55     no     28 0.9161253 0.7787905 0.9762571
#> 56    yes     28 0.8805867 0.7227536 0.9648389
#> 57     no     29 0.9241097 0.7897216 0.9798211
#> 58    yes     29 0.8915537 0.7358133 0.9700652
#> 59     no     30 0.9313909 0.8002585 0.9828596
#> 60    yes     30 0.9016260 0.7489622 0.9745352
#> 61     no     31 0.9380204 0.8103942 0.9854473
#> 62    yes     31 0.9108563 0.7616684 0.9783526
#> 63     no     32 0.9440478 0.8201310 0.9876493
#> 64    yes     32 0.9192981 0.7739571 0.9816085
#> 65     no     33 0.9495206 0.8294730 0.9895217
#> 66    yes     33 0.9270045 0.7858540 0.9843826
#> 67     no     34 0.9544839 0.8384253 0.9911128
#> 68    yes     34 0.9340278 0.7972884 0.9867438
#> 69     no     35 0.9589802 0.8469943 0.9924641
#> 70    yes     35 0.9404188 0.8082607 0.9887522
#> 71     no     36 0.9630496 0.8551873 0.9936113
#> 72    yes     36 0.9462263 0.8187736 0.9904592
#> 73     no     37 0.9667293 0.8630125 0.9945848
#> 74    yes     37 0.9514970 0.8288317 0.9919093
#> 75     no     38 0.9700539 0.8704789 0.9954107
#> 76    yes     38 0.9562749 0.8384412 0.9931405
#> 77     no     39 0.9730556 0.8775959 0.9961111
#> 78    yes     39 0.9606016 0.8476099 0.9941854
#> 79     no     40 0.9757639 0.8843738 0.9967049
#> 80    yes     40 0.9645161 0.8563469 0.9950720
#> 81     no     41 0.9782061 0.8909201 0.9972084
#> 82    yes     41 0.9680546 0.8646624 0.9958239
#> 83     no     42 0.9804072 0.8971687 0.9976351
#> 84    yes     42 0.9712507 0.8725678 0.9964615
#> 85     no     43 0.9823899 0.9030982 0.9979967
#> 86    yes     43 0.9741357 0.8800750 0.9970021
#> 87     no     44 0.9841753 0.9087205 0.9983031
#> 88    yes     44 0.9767380 0.8871966 0.9974603
#> 89     no     45 0.9857822 0.9140477 0.9985627
#> 90    yes     45 0.9790842 0.8938372 0.9978486
#> 91     no     46 0.9872281 0.9190916 0.9987826
#> 92    yes     46 0.9811982 0.8998715 0.9981777
#> 93     no     47 0.9885287 0.9238641 0.9989690
#> 94    yes     47 0.9831023 0.9055989 0.9984565
#> 95     no     48 0.9896982 0.9283770 0.9991268
#> 96    yes     48 0.9848165 0.9110310 0.9986927
#> 97     no     49 0.9907496 0.9326418 0.9992605
#> 98    yes     49 0.9863593 0.9161794 0.9988928
#> 99     no     50 0.9916946 0.9366748 0.9993737
#> 100   yes     50 0.9877472 0.9210556 0.9990623
#> 101    no     51 0.9925437 0.9404899 0.9994696
#> 102   yes     51 0.9889955 0.9256711 0.9992058
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)