Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3150686 0.4297486
#> 2     yes      1 0.2814584 0.2216806 0.3620317
#> 3      no      2 0.3927672 0.3429244 0.4484993
#> 4     yes      2 0.3039568 0.2438729 0.3766215
#> 5      no      3 0.4189750 0.3688964 0.4728269
#> 6     yes      3 0.3274340 0.2645075 0.3986847
#> 7      no      4 0.4456481 0.3959240 0.5010808
#> 8     yes      4 0.3518080 0.2882147 0.4215503
#> 9      no      5 0.4726379 0.4207177 0.5337220
#> 10    yes      5 0.3769795 0.3139423 0.4447673
#> 11     no      6 0.4997887 0.4414431 0.5654049
#> 12    yes      6 0.4028326 0.3389937 0.4698887
#> 13     no      7 0.5269407 0.4562959 0.5965592
#> 14    yes      7 0.4292372 0.3640053 0.4982451
#> 15     no      8 0.5539343 0.4743654 0.6269486
#> 16    yes      8 0.4560509 0.3840313 0.5308153
#> 17     no      9 0.5806134 0.4939850 0.6563606
#> 18    yes      9 0.4831217 0.4008922 0.5641714
#> 19     no     10 0.6068295 0.5102295 0.6861393
#> 20    yes     10 0.5102919 0.4164707 0.5972895
#> 21     no     11 0.6324443 0.5260811 0.7167753
#> 22    yes     11 0.5374015 0.4339533 0.6318175
#> 23     no     12 0.6573325 0.5418719 0.7455277
#> 24    yes     12 0.5642918 0.4516204 0.6650339
#> 25     no     13 0.6813846 0.5575685 0.7722837
#> 26    yes     13 0.5908091 0.4702619 0.6987295
#> 27     no     14 0.7045073 0.5731385 0.7979289
#> 28    yes     14 0.6168084 0.4889673 0.7311377
#> 29     no     15 0.7266252 0.5885507 0.8219892
#> 30    yes     15 0.6421564 0.5013363 0.7612514
#> 31     no     16 0.7476805 0.6037751 0.8437449
#> 32    yes     16 0.6667336 0.5137054 0.7889662
#> 33     no     17 0.7676327 0.6187835 0.8633338
#> 34    yes     17 0.6904365 0.5260587 0.8142491
#> 35     no     18 0.7864578 0.6335497 0.8810385
#> 36    yes     18 0.7131789 0.5383807 0.8371287
#> 37     no     19 0.8041469 0.6460379 0.8967239
#> 38    yes     19 0.7348919 0.5506558 0.8574861
#> 39     no     20 0.8207047 0.6581680 0.9105511
#> 40    yes     20 0.7555247 0.5631184 0.8755531
#> 41     no     21 0.8361480 0.6700937 0.9226866
#> 42    yes     21 0.7750433 0.5756159 0.8916186
#> 43     no     22 0.8505034 0.6818037 0.9332963
#> 44    yes     22 0.7934296 0.5880158 0.9058326
#> 45     no     23 0.8638059 0.6932875 0.9425407
#> 46    yes     23 0.8106802 0.6003027 0.9183526
#> 47     no     24 0.8760972 0.7045358 0.9505717
#> 48    yes     24 0.8268046 0.6124620 0.9293372
#> 49     no     25 0.8874238 0.7155402 0.9575309
#> 50    yes     25 0.8418238 0.6244796 0.9389344
#> 51     no     26 0.8978357 0.7262934 0.9635478
#> 52    yes     26 0.8557676 0.6363422 0.9472586
#> 53     no     27 0.9073851 0.7367891 0.9687401
#> 54    yes     27 0.8686741 0.6480373 0.9545025
#> 55     no     28 0.9161253 0.7470221 0.9732133
#> 56    yes     28 0.8805867 0.6595531 0.9607922
#> 57     no     29 0.9241097 0.7569880 0.9770616
#> 58    yes     29 0.8915537 0.6708789 0.9662428
#> 59     no     30 0.9313909 0.7666836 0.9803681
#> 60    yes     30 0.9016260 0.6820046 0.9709581
#> 61     no     31 0.9380204 0.7761065 0.9832063
#> 62    yes     31 0.9108563 0.6929210 0.9750315
#> 63     no     32 0.9440478 0.7852551 0.9856401
#> 64    yes     32 0.9192981 0.7036202 0.9785459
#> 65     no     33 0.9495206 0.7941289 0.9877256
#> 66    yes     33 0.9270045 0.7140947 0.9815747
#> 67     no     34 0.9544839 0.8027279 0.9895114
#> 68    yes     34 0.9340278 0.7243384 0.9841826
#> 69     no     35 0.9589802 0.8110530 0.9910398
#> 70    yes     35 0.9404188 0.7343456 0.9864264
#> 71     no     36 0.9630496 0.8191059 0.9923472
#> 72    yes     36 0.9462263 0.7444961 0.9883554
#> 73     no     37 0.9667293 0.8268887 0.9934651
#> 74    yes     37 0.9514970 0.7546615 0.9900130
#> 75     no     38 0.9700539 0.8344042 0.9944206
#> 76    yes     38 0.9562749 0.7645495 0.9914365
#> 77     no     39 0.9730556 0.8416559 0.9952370
#> 78    yes     39 0.9606016 0.7741572 0.9926585
#> 79     no     40 0.9757639 0.8486477 0.9959345
#> 80    yes     40 0.9645161 0.7834831 0.9937072
#> 81     no     41 0.9782061 0.8553837 0.9965302
#> 82    yes     41 0.9680546 0.7925263 0.9946068
#> 83     no     42 0.9804072 0.8618687 0.9970389
#> 84    yes     42 0.9712507 0.8012867 0.9953784
#> 85     no     43 0.9823899 0.8681077 0.9974732
#> 86    yes     43 0.9741357 0.8097654 0.9960399
#> 87     no     44 0.9841753 0.8741061 0.9978439
#> 88    yes     44 0.9767380 0.8179637 0.9966071
#> 89     no     45 0.9857822 0.8798695 0.9981603
#> 90    yes     45 0.9790842 0.8258840 0.9970932
#> 91     no     46 0.9872281 0.8854036 0.9984304
#> 92    yes     46 0.9811982 0.8335293 0.9975098
#> 93     no     47 0.9885287 0.8907145 0.9986609
#> 94    yes     47 0.9831023 0.8409030 0.9978668
#> 95     no     48 0.9896982 0.8958083 0.9988575
#> 96    yes     48 0.9848165 0.8480090 0.9981727
#> 97     no     49 0.9907496 0.9006911 0.9990254
#> 98    yes     49 0.9863593 0.8548520 0.9984348
#> 99     no     50 0.9916946 0.9053694 0.9991685
#> 100   yes     50 0.9877472 0.8614367 0.9986593
#> 101    no     51 0.9925437 0.9098493 0.9992907
#> 102   yes     51 0.9889955 0.8677685 0.9988516
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)