Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3066821 0.4124110
#> 2     yes      1 0.2814584 0.2188427 0.3513303
#> 3      no      2 0.3927672 0.3365255 0.4368886
#> 4     yes      2 0.3039568 0.2416430 0.3716964
#> 5      no      3 0.4189750 0.3660610 0.4630067
#> 6     yes      3 0.3274340 0.2663828 0.3962854
#> 7      no      4 0.4456481 0.3902120 0.4897943
#> 8     yes      4 0.3518080 0.2900493 0.4167923
#> 9      no      5 0.4726379 0.4107077 0.5214612
#> 10    yes      5 0.3769795 0.3151601 0.4447429
#> 11     no      6 0.4997887 0.4293620 0.5528985
#> 12    yes      6 0.4028326 0.3389419 0.4707786
#> 13     no      7 0.5269407 0.4487336 0.5849319
#> 14    yes      7 0.4292372 0.3631056 0.5017167
#> 15     no      8 0.5539343 0.4682697 0.6180223
#> 16    yes      8 0.4560509 0.3882171 0.5365339
#> 17     no      9 0.5806134 0.4828872 0.6495461
#> 18    yes      9 0.4831217 0.4125476 0.5737244
#> 19     no     10 0.6068295 0.4981273 0.6818283
#> 20    yes     10 0.5102919 0.4348182 0.6128469
#> 21     no     11 0.6324443 0.5165590 0.7130535
#> 22    yes     11 0.5374015 0.4548656 0.6504695
#> 23     no     12 0.6573325 0.5351223 0.7423723
#> 24    yes     12 0.5642918 0.4711161 0.6861896
#> 25     no     13 0.6813846 0.5535881 0.7696633
#> 26    yes     13 0.5908091 0.4874299 0.7198301
#> 27     no     14 0.7045073 0.5719060 0.7948623
#> 28    yes     14 0.6168084 0.5037722 0.7511703
#> 29     no     15 0.7266252 0.5900274 0.8186897
#> 30    yes     15 0.6421564 0.5201076 0.7800739
#> 31     no     16 0.7476805 0.6079057 0.8409189
#> 32    yes     16 0.6667336 0.5364010 0.8064830
#> 33     no     17 0.7676327 0.6254970 0.8606094
#> 34    yes     17 0.6904365 0.5526176 0.8304084
#> 35     no     18 0.7864578 0.6420536 0.8782092
#> 36    yes     18 0.7131789 0.5687232 0.8517715
#> 37     no     19 0.8041469 0.6571034 0.8938544
#> 38    yes     19 0.7348919 0.5846846 0.8703876
#> 39     no     20 0.8207047 0.6718442 0.9079193
#> 40    yes     20 0.7555247 0.6005906 0.8869760
#> 41     no     21 0.8361480 0.6862540 0.9204641
#> 42    yes     21 0.7750433 0.6164045 0.9016809
#> 43     no     22 0.8505034 0.7003132 0.9314275
#> 44    yes     22 0.7934296 0.6319749 0.9146566
#> 45     no     23 0.8638059 0.7140046 0.9409754
#> 46    yes     23 0.8106802 0.6454185 0.9260599
#> 47     no     24 0.8760972 0.7273136 0.9492653
#> 48    yes     24 0.8268046 0.6579720 0.9360601
#> 49     no     25 0.8874238 0.7402279 0.9564439
#> 50    yes     25 0.8418238 0.6706571 0.9448342
#> 51     no     26 0.8978357 0.7527378 0.9626460
#> 52    yes     26 0.8557676 0.6830953 0.9524649
#> 53     no     27 0.9073851 0.7649164 0.9679939
#> 54    yes     27 0.8686741 0.6952736 0.9590624
#> 55     no     28 0.9161253 0.7767451 0.9725974
#> 56    yes     28 0.8805867 0.7071806 0.9648310
#> 57     no     29 0.9241097 0.7881424 0.9765543
#> 58    yes     29 0.8915537 0.7188061 0.9698449
#> 59     no     30 0.9313909 0.7991071 0.9799513
#> 60    yes     30 0.9016260 0.7301415 0.9741631
#> 61     no     31 0.9380204 0.8096403 0.9828644
#> 62    yes     31 0.9108563 0.7422349 0.9778770
#> 63     no     32 0.9440478 0.8197446 0.9853602
#> 64    yes     32 0.9192981 0.7543462 0.9810674
#> 65     no     33 0.9495206 0.8294242 0.9874969
#> 66    yes     33 0.9270045 0.7660680 0.9838053
#> 67     no     34 0.9544839 0.8386852 0.9893250
#> 68    yes     34 0.9340278 0.7773957 0.9861529
#> 69     no     35 0.9589802 0.8475345 0.9908880
#> 70    yes     35 0.9404188 0.7883264 0.9881642
#> 71     no     36 0.9630496 0.8559804 0.9922239
#> 72    yes     36 0.9462263 0.7988592 0.9898864
#> 73     no     37 0.9667293 0.8640021 0.9933651
#> 74    yes     37 0.9514970 0.8089947 0.9913602
#> 75     no     38 0.9700539 0.8711970 0.9943397
#> 76    yes     38 0.9562749 0.8187352 0.9926208
#> 77     no     39 0.9730556 0.8780651 0.9951717
#> 78    yes     39 0.9606016 0.8280842 0.9936986
#> 79     no     40 0.9757639 0.8846155 0.9958819
#> 80    yes     40 0.9645161 0.8370467 0.9946199
#> 81     no     41 0.9782061 0.8908578 0.9964879
#> 82    yes     41 0.9680546 0.8456287 0.9954070
#> 83     no     42 0.9804072 0.8968018 0.9970049
#> 84    yes     42 0.9712507 0.8538373 0.9960795
#> 85     no     43 0.9823899 0.9024576 0.9974460
#> 86    yes     43 0.9741357 0.8616803 0.9966538
#> 87     no     44 0.9841753 0.9078353 0.9978222
#> 88    yes     44 0.9767380 0.8691666 0.9971443
#> 89     no     45 0.9857822 0.9129452 0.9981431
#> 90    yes     45 0.9790842 0.8763053 0.9975630
#> 91     no     46 0.9872281 0.9177974 0.9984167
#> 92    yes     46 0.9811982 0.8831277 0.9979205
#> 93     no     47 0.9885287 0.9224022 0.9986501
#> 94    yes     47 0.9831023 0.8896655 0.9982256
#> 95     no     48 0.9896982 0.9267696 0.9988491
#> 96    yes     48 0.9848165 0.8958802 0.9984860
#> 97     no     49 0.9907496 0.9309096 0.9990187
#> 98    yes     49 0.9863593 0.9017831 0.9987082
#> 99     no     50 0.9916946 0.9348321 0.9991634
#> 100   yes     50 0.9877472 0.9073855 0.9988979
#> 101    no     51 0.9925437 0.9385466 0.9992859
#> 102   yes     51 0.9889955 0.9126988 0.9990597
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)