Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3141000 0.4226929
#> 2     yes      1 0.2814584 0.2189469 0.3589927
#> 3      no      2 0.3927672 0.3425986 0.4405785
#> 4     yes      2 0.3039568 0.2440438 0.3772271
#> 5      no      3 0.4189750 0.3707055 0.4625478
#> 6     yes      3 0.3274340 0.2682252 0.3929848
#> 7      no      4 0.4456481 0.3973972 0.4886920
#> 8     yes      4 0.3518080 0.2942761 0.4119540
#> 9      no      5 0.4726379 0.4248294 0.5206578
#> 10    yes      5 0.3769795 0.3193461 0.4380962
#> 11     no      6 0.4997887 0.4522979 0.5487874
#> 12    yes      6 0.4028326 0.3481044 0.4647589
#> 13     no      7 0.5269407 0.4750896 0.5782666
#> 14    yes      7 0.4292372 0.3758854 0.4975935
#> 15     no      8 0.5539343 0.4935880 0.6111049
#> 16    yes      8 0.4560509 0.3987037 0.5369442
#> 17     no      9 0.5806134 0.5099217 0.6444648
#> 18    yes      9 0.4831217 0.4158064 0.5772880
#> 19     no     10 0.6068295 0.5262301 0.6777877
#> 20    yes     10 0.5102919 0.4327010 0.6161706
#> 21     no     11 0.6324443 0.5424746 0.7110708
#> 22    yes     11 0.5374015 0.4487782 0.6580032
#> 23     no     12 0.6573325 0.5577732 0.7447515
#> 24    yes     12 0.5642918 0.4649648 0.6962954
#> 25     no     13 0.6813846 0.5727143 0.7757086
#> 26    yes     13 0.5908091 0.4812266 0.7312514
#> 27     no     14 0.7045073 0.5875135 0.8038590
#> 28    yes     14 0.6168084 0.4977824 0.7633357
#> 29     no     15 0.7266252 0.6021446 0.8292117
#> 30    yes     15 0.6421564 0.5144385 0.7926355
#> 31     no     16 0.7476805 0.6165823 0.8518497
#> 32    yes     16 0.6667336 0.5310615 0.8191244
#> 33     no     17 0.7676327 0.6308033 0.8719115
#> 34    yes     17 0.6904365 0.5476143 0.8428576
#> 35     no     18 0.7864578 0.6454704 0.8895734
#> 36    yes     18 0.7131789 0.5640602 0.8639532
#> 37     no     19 0.8041469 0.6601859 0.9050339
#> 38    yes     19 0.7348919 0.5814387 0.8825742
#> 39     no     20 0.8207047 0.6745928 0.9185009
#> 40    yes     20 0.7555247 0.5992431 0.8986678
#> 41     no     21 0.8361480 0.6886713 0.9301821
#> 42    yes     21 0.7750433 0.6151732 0.9127483
#> 43     no     22 0.8505034 0.7024039 0.9402781
#> 44    yes     22 0.7934296 0.6297637 0.9250190
#> 45     no     23 0.8638059 0.7157753 0.9489776
#> 46    yes     23 0.8106802 0.6449054 0.9356715
#> 47     no     24 0.8760972 0.7287726 0.9564546
#> 48    yes     24 0.8268046 0.6597541 0.9448890
#> 49     no     25 0.8874238 0.7413851 0.9628672
#> 50    yes     25 0.8418238 0.6742796 0.9528424
#> 51     no     26 0.8978357 0.7536044 0.9683571
#> 52    yes     26 0.8557676 0.6884593 0.9596887
#> 53     no     27 0.9073851 0.7654244 0.9730500
#> 54    yes     27 0.8686741 0.7022737 0.9655701
#> 55     no     28 0.9161253 0.7768406 0.9770568
#> 56    yes     28 0.8805867 0.7156243 0.9706139
#> 57     no     29 0.9241097 0.7878509 0.9804743
#> 58    yes     29 0.8915537 0.7289371 0.9749332
#> 59     no     30 0.9313909 0.7984548 0.9833868
#> 60    yes     30 0.9016260 0.7418453 0.9786277
#> 61     no     31 0.9380204 0.8086535 0.9858674
#> 62    yes     31 0.9108563 0.7543399 0.9817845
#> 63     no     32 0.9440478 0.8184498 0.9879789
#> 64    yes     32 0.9192981 0.7664141 0.9844796
#> 65     no     33 0.9495206 0.8278478 0.9897756
#> 66    yes     33 0.9270045 0.7780635 0.9867790
#> 67     no     34 0.9544839 0.8368531 0.9913040
#> 68    yes     34 0.9340278 0.7892858 0.9887398
#> 69     no     35 0.9589802 0.8454721 0.9926038
#> 70    yes     35 0.9404188 0.8000807 0.9904109
#> 71     no     36 0.9630496 0.8537126 0.9937090
#> 72    yes     36 0.9462263 0.8104500 0.9918347
#> 73     no     37 0.9667293 0.8615829 0.9946487
#> 74    yes     37 0.9514970 0.8203968 0.9930474
#> 75     no     38 0.9700539 0.8690922 0.9954476
#> 76    yes     38 0.9562749 0.8299260 0.9940802
#> 77     no     39 0.9730556 0.8762503 0.9961269
#> 78    yes     39 0.9606016 0.8390440 0.9949595
#> 79     no     40 0.9757639 0.8830675 0.9967044
#> 80    yes     40 0.9645161 0.8477581 0.9957081
#> 81     no     41 0.9782061 0.8895545 0.9971954
#> 82    yes     41 0.9680546 0.8560768 0.9963454
#> 83     no     42 0.9804072 0.8957222 0.9976130
#> 84    yes     42 0.9712507 0.8640098 0.9968878
#> 85     no     43 0.9823899 0.9015819 0.9979680
#> 86    yes     43 0.9741357 0.8715670 0.9973496
#> 87     no     44 0.9841753 0.9071449 0.9982700
#> 88    yes     44 0.9767380 0.8787594 0.9977427
#> 89     no     45 0.9857822 0.9124225 0.9985269
#> 90    yes     45 0.9790842 0.8855983 0.9980773
#> 91     no     46 0.9872281 0.9174261 0.9987455
#> 92    yes     46 0.9811982 0.8920955 0.9983621
#> 93     no     47 0.9885287 0.9221669 0.9989314
#> 94    yes     47 0.9831023 0.8982629 0.9986046
#> 95     no     48 0.9896982 0.9266561 0.9990897
#> 96    yes     48 0.9848165 0.9041127 0.9988111
#> 97     no     49 0.9907496 0.9309047 0.9992244
#> 98    yes     49 0.9863593 0.9096573 0.9989870
#> 99     no     50 0.9916946 0.9349234 0.9993390
#> 100   yes     50 0.9877472 0.9149088 0.9991367
#> 101    no     51 0.9925437 0.9387229 0.9994365
#> 102   yes     51 0.9889955 0.9198795 0.9992642
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)