Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3126088 0.4286942
#> 2     yes      1 0.2814584 0.2151961 0.3509266
#> 3      no      2 0.3927672 0.3386161 0.4441202
#> 4     yes      2 0.3039568 0.2363014 0.3683275
#> 5      no      3 0.4189750 0.3672387 0.4656407
#> 6     yes      3 0.3274340 0.2605434 0.3909169
#> 7      no      4 0.4456481 0.3970530 0.4919141
#> 8     yes      4 0.3518080 0.2881908 0.4144472
#> 9      no      5 0.4726379 0.4269822 0.5182329
#> 10    yes      5 0.3769795 0.3175127 0.4460673
#> 11     no      6 0.4997887 0.4570593 0.5521342
#> 12    yes      6 0.4028326 0.3450671 0.4827685
#> 13     no      7 0.5269407 0.4792146 0.5939807
#> 14    yes      7 0.4292372 0.3722195 0.5205250
#> 15     no      8 0.5539343 0.4961224 0.6290079
#> 16    yes      8 0.4560509 0.3958690 0.5580483
#> 17     no      9 0.5806134 0.5110914 0.6671176
#> 18    yes      9 0.4831217 0.4165692 0.5949201
#> 19     no     10 0.6068295 0.5277057 0.7031603
#> 20    yes     10 0.5102919 0.4335384 0.6307508
#> 21     no     11 0.6324443 0.5464311 0.7368402
#> 22    yes     11 0.5374015 0.4506810 0.6651953
#> 23     no     12 0.6573325 0.5623988 0.7679599
#> 24    yes     12 0.5642918 0.4727393 0.6989679
#> 25     no     13 0.6813846 0.5762847 0.7964164
#> 26    yes     13 0.5908091 0.4910475 0.7317199
#> 27     no     14 0.7045073 0.5900470 0.8221911
#> 28    yes     14 0.6168084 0.5087401 0.7618061
#> 29     no     15 0.7266252 0.6036645 0.8453363
#> 30    yes     15 0.6421564 0.5248511 0.7894825
#> 31     no     16 0.7476805 0.6171168 0.8659599
#> 32    yes     16 0.6667336 0.5391689 0.8147180
#> 33     no     17 0.7676327 0.6303847 0.8842101
#> 34    yes     17 0.6904365 0.5532819 0.8375436
#> 35     no     18 0.7864578 0.6434501 0.9002616
#> 36    yes     18 0.7131789 0.5673014 0.8580400
#> 37     no     19 0.8041469 0.6562964 0.9143035
#> 38    yes     19 0.7348919 0.5819713 0.8771560
#> 39     no     20 0.8207047 0.6689080 0.9265201
#> 40    yes     20 0.7555247 0.5965229 0.8940367
#> 41     no     21 0.8361480 0.6817883 0.9371073
#> 42    yes     21 0.7750433 0.6109011 0.9088388
#> 43     no     22 0.8505034 0.6945942 0.9462575
#> 44    yes     22 0.7934296 0.6250834 0.9217542
#> 45     no     23 0.8638059 0.7071014 0.9541416
#> 46    yes     23 0.8106802 0.6390483 0.9329747
#> 47     no     24 0.8760972 0.7192987 0.9609168
#> 48    yes     24 0.8268046 0.6527759 0.9426863
#> 49     no     25 0.8874238 0.7311765 0.9667259
#> 50    yes     25 0.8418238 0.6662481 0.9510645
#> 51     no     26 0.8978357 0.7427267 0.9716970
#> 52    yes     26 0.8557676 0.6794480 0.9582722
#> 53     no     27 0.9073851 0.7539432 0.9759438
#> 54    yes     27 0.8686741 0.6923607 0.9644580
#> 55     no     28 0.9161253 0.7648209 0.9795668
#> 56    yes     28 0.8805867 0.7049730 0.9697557
#> 57     no     29 0.9241097 0.7753566 0.9826539
#> 58    yes     29 0.8915537 0.7172732 0.9742849
#> 59     no     30 0.9313909 0.7855484 0.9852815
#> 60    yes     30 0.9016260 0.7292514 0.9781510
#> 61     no     31 0.9380204 0.7953956 0.9875161
#> 62    yes     31 0.9108563 0.7408995 0.9814470
#> 63     no     32 0.9440478 0.8048990 0.9894151
#> 64    yes     32 0.9192981 0.7522109 0.9842538
#> 65     no     33 0.9495206 0.8140602 0.9910279
#> 66    yes     33 0.9270045 0.7631806 0.9866417
#> 67     no     34 0.9544839 0.8228822 0.9923968
#> 68    yes     34 0.9340278 0.7738051 0.9886716
#> 69     no     35 0.9589802 0.8313687 0.9935582
#> 70    yes     35 0.9404188 0.7840825 0.9903961
#> 71     no     36 0.9630496 0.8395245 0.9945432
#> 72    yes     36 0.9462263 0.7940121 0.9918602
#> 73     no     37 0.9667293 0.8473550 0.9953782
#> 74    yes     37 0.9514970 0.8035944 0.9931027
#> 75     no     38 0.9700539 0.8548663 0.9960860
#> 76    yes     38 0.9562749 0.8128313 0.9941566
#> 77     no     39 0.9730556 0.8620652 0.9966857
#> 78    yes     39 0.9606016 0.8217256 0.9950503
#> 79     no     40 0.9757639 0.8689590 0.9971938
#> 80    yes     40 0.9645161 0.8302811 0.9958079
#> 81     no     41 0.9782061 0.8755553 0.9976242
#> 82    yes     41 0.9680546 0.8385026 0.9964499
#> 83     no     42 0.9804072 0.8818623 0.9979887
#> 84    yes     42 0.9712507 0.8463955 0.9969939
#> 85     no     43 0.9823899 0.8878881 0.9982974
#> 86    yes     43 0.9741357 0.8539662 0.9974548
#> 87     no     44 0.9841753 0.8936415 0.9985588
#> 88    yes     44 0.9767380 0.8612214 0.9978451
#> 89     no     45 0.9857822 0.8991310 0.9987801
#> 90    yes     45 0.9790842 0.8681685 0.9981757
#> 91     no     46 0.9872281 0.9043655 0.9989674
#> 92    yes     46 0.9811982 0.8748153 0.9984557
#> 93     no     47 0.9885287 0.9093540 0.9991260
#> 94    yes     47 0.9831023 0.8811698 0.9986927
#> 95     no     48 0.9896982 0.9141052 0.9992603
#> 96    yes     48 0.9848165 0.8872406 0.9988934
#> 97     no     49 0.9907496 0.9186280 0.9993739
#> 98    yes     49 0.9863593 0.8930363 0.9990633
#> 99     no     50 0.9916946 0.9229312 0.9994701
#> 100   yes     50 0.9877472 0.8985657 0.9992072
#> 101    no     51 0.9925437 0.9270235 0.9995515
#> 102   yes     51 0.9889955 0.9038378 0.9993290
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)