Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3015169 0.4231144
#> 2     yes      1 0.2814584 0.2103209 0.3501833
#> 3      no      2 0.3927672 0.3342038 0.4451778
#> 4     yes      2 0.3039568 0.2343653 0.3698971
#> 5      no      3 0.4189750 0.3672077 0.4653836
#> 6     yes      3 0.3274340 0.2605561 0.3956537
#> 7      no      4 0.4456481 0.3992297 0.4986789
#> 8     yes      4 0.3518080 0.2886306 0.4246461
#> 9      no      5 0.4726379 0.4242323 0.5313392
#> 10    yes      5 0.3769795 0.3160933 0.4566604
#> 11     no      6 0.4997887 0.4456744 0.5616297
#> 12    yes      6 0.4028326 0.3380160 0.4890369
#> 13     no      7 0.5269407 0.4690154 0.5924216
#> 14    yes      7 0.4292372 0.3614881 0.5215057
#> 15     no      8 0.5539343 0.4939922 0.6256322
#> 16    yes      8 0.4560509 0.3886998 0.5535538
#> 17     no      9 0.5806134 0.5189960 0.6596948
#> 18    yes      9 0.4831217 0.4166426 0.5843240
#> 19     no     10 0.6068295 0.5406852 0.6963921
#> 20    yes     10 0.5102919 0.4440982 0.6182983
#> 21     no     11 0.6324443 0.5598373 0.7307479
#> 22    yes     11 0.5374015 0.4670071 0.6517902
#> 23     no     12 0.6573325 0.5788129 0.7625419
#> 24    yes     12 0.5642918 0.4872802 0.6861665
#> 25     no     13 0.6813846 0.5978427 0.7916522
#> 26    yes     13 0.5908091 0.5078832 0.7186133
#> 27     no     14 0.7045073 0.6177242 0.8180454
#> 28    yes     14 0.6168084 0.5283903 0.7489321
#> 29     no     15 0.7266252 0.6381067 0.8417633
#> 30    yes     15 0.6421564 0.5481659 0.7769970
#> 31     no     16 0.7476805 0.6566465 0.8629076
#> 32    yes     16 0.6667336 0.5668195 0.8027502
#> 33     no     17 0.7676327 0.6736264 0.8816236
#> 34    yes     17 0.6904365 0.5852670 0.8261939
#> 35     no     18 0.7864578 0.6904612 0.8980801
#> 36    yes     18 0.7131789 0.6034786 0.8473803
#> 37     no     19 0.8041469 0.7067954 0.9124428
#> 38    yes     19 0.7348919 0.6214079 0.8664012
#> 39     no     20 0.8207047 0.7226022 0.9249504
#> 40    yes     20 0.7555247 0.6390116 0.8833773
#> 41     no     21 0.8361480 0.7378597 0.9357968
#> 42    yes     21 0.7750433 0.6559165 0.8984487
#> 43     no     22 0.8505034 0.7525513 0.9451685
#> 44    yes     22 0.7934296 0.6721708 0.9117666
#> 45     no     23 0.8638059 0.7659729 0.9532404
#> 46    yes     23 0.8106802 0.6882058 0.9234863
#> 47     no     24 0.8760972 0.7785123 0.9600048
#> 48    yes     24 0.8268046 0.7025216 0.9337622
#> 49     no     25 0.8874238 0.7905592 0.9658179
#> 50    yes     25 0.8418238 0.7164368 0.9427434
#> 51     no     26 0.8978357 0.8021130 0.9708008
#> 52    yes     26 0.8557676 0.7299890 0.9505711
#> 53     no     27 0.9073851 0.8131759 0.9750614
#> 54    yes     27 0.8686741 0.7431758 0.9573769
#> 55     no     28 0.9161253 0.8237520 0.9787136
#> 56    yes     28 0.8805867 0.7559336 0.9632816
#> 57     no     29 0.9241097 0.8338478 0.9818408
#> 58    yes     29 0.8915537 0.7682550 0.9683953
#> 59     no     30 0.9313909 0.8434713 0.9845156
#> 60    yes     30 0.9016260 0.7801349 0.9728168
#> 61     no     31 0.9380204 0.8526322 0.9868015
#> 62    yes     31 0.9108563 0.7915707 0.9766345
#> 63     no     32 0.9440478 0.8614156 0.9887537
#> 64    yes     32 0.9192981 0.8025618 0.9799271
#> 65     no     33 0.9495206 0.8698648 0.9904198
#> 66    yes     33 0.9270045 0.8131098 0.9827638
#> 67     no     34 0.9544839 0.8777920 0.9918410
#> 68    yes     34 0.9340278 0.8232180 0.9852057
#> 69     no     35 0.9589802 0.8851584 0.9930528
#> 70    yes     35 0.9404188 0.8328915 0.9873060
#> 71     no     36 0.9630496 0.8921354 0.9940856
#> 72    yes     36 0.9462263 0.8421368 0.9891114
#> 73     no     37 0.9667293 0.8987370 0.9949656
#> 74    yes     37 0.9514970 0.8509618 0.9906625
#> 75     no     38 0.9700539 0.9049777 0.9957151
#> 76    yes     38 0.9562749 0.8593755 0.9919943
#> 77     no     39 0.9730556 0.9108720 0.9963534
#> 78    yes     39 0.9606016 0.8673878 0.9931375
#> 79     no     40 0.9757639 0.9164344 0.9968969
#> 80    yes     40 0.9645161 0.8750097 0.9941199
#> 81     no     41 0.9782061 0.9216796 0.9973596
#> 82    yes     41 0.9680546 0.8822526 0.9949985
#> 83     no     42 0.9804072 0.9266219 0.9977534
#> 84    yes     42 0.9712507 0.8891287 0.9957463
#> 85     no     43 0.9823899 0.9312756 0.9980886
#> 86    yes     43 0.9741357 0.8956505 0.9963826
#> 87     no     44 0.9841753 0.9356548 0.9983738
#> 88    yes     44 0.9767380 0.9018307 0.9969240
#> 89     no     45 0.9857822 0.9397729 0.9986165
#> 90    yes     45 0.9790842 0.9076823 0.9973845
#> 91     no     46 0.9872281 0.9436434 0.9988230
#> 92    yes     46 0.9811982 0.9132184 0.9977762
#> 93     no     47 0.9885287 0.9472792 0.9989987
#> 94    yes     47 0.9831023 0.9184521 0.9981093
#> 95     no     48 0.9896982 0.9506926 0.9991482
#> 96    yes     48 0.9848165 0.9233964 0.9983926
#> 97     no     49 0.9907496 0.9538958 0.9992754
#> 98    yes     49 0.9863593 0.9280642 0.9986334
#> 99     no     50 0.9916946 0.9569004 0.9993836
#> 100   yes     50 0.9877472 0.9324682 0.9988382
#> 101    no     51 0.9925437 0.9597175 0.9994756
#> 102   yes     51 0.9889955 0.9366208 0.9990123
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)