Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3073610 0.4326796
#> 2     yes      1 0.2814584 0.2062562 0.3419266
#> 3      no      2 0.3927672 0.3345529 0.4562488
#> 4     yes      2 0.3039568 0.2314267 0.3581949
#> 5      no      3 0.4189750 0.3682312 0.4825254
#> 6     yes      3 0.3274340 0.2593081 0.3779511
#> 7      no      4 0.4456481 0.3986416 0.5088933
#> 8     yes      4 0.3518080 0.2872219 0.4010796
#> 9      no      5 0.4726379 0.4171900 0.5378448
#> 10    yes      5 0.3769795 0.3129376 0.4287288
#> 11     no      6 0.4997887 0.4397385 0.5702790
#> 12    yes      6 0.4028326 0.3351514 0.4606132
#> 13     no      7 0.5269407 0.4654213 0.6021598
#> 14    yes      7 0.4292372 0.3567037 0.5008222
#> 15     no      8 0.5539343 0.4913664 0.6356472
#> 16    yes      8 0.4560509 0.3721576 0.5386283
#> 17     no      9 0.5806134 0.5146072 0.6677491
#> 18    yes      9 0.4831217 0.3922430 0.5756547
#> 19     no     10 0.6068295 0.5330282 0.6974226
#> 20    yes     10 0.5102919 0.4156981 0.6118495
#> 21     no     11 0.6324443 0.5500793 0.7261285
#> 22    yes     11 0.5374015 0.4400476 0.6468488
#> 23     no     12 0.6573325 0.5656253 0.7538434
#> 24    yes     12 0.5642918 0.4615705 0.6803384
#> 25     no     13 0.6813846 0.5814387 0.7827890
#> 26    yes     13 0.5908091 0.4814569 0.7127671
#> 27     no     14 0.7045073 0.5977654 0.8111123
#> 28    yes     14 0.6168084 0.5014032 0.7462734
#> 29     no     15 0.7266252 0.6138594 0.8361873
#> 30    yes     15 0.6421564 0.5213460 0.7767627
#> 31     no     16 0.7476805 0.6285532 0.8576852
#> 32    yes     16 0.6667336 0.5412218 0.8042485
#> 33     no     17 0.7676327 0.6422311 0.8761595
#> 34    yes     17 0.6904365 0.5598171 0.8290754
#> 35     no     18 0.7864578 0.6556685 0.8925334
#> 36    yes     18 0.7131789 0.5782409 0.8513158
#> 37     no     19 0.8041469 0.6704190 0.9069697
#> 38    yes     19 0.7348919 0.5964488 0.8710928
#> 39     no     20 0.8207047 0.6853631 0.9196390
#> 40    yes     20 0.7555247 0.6147704 0.8885653
#> 41     no     21 0.8361480 0.6999342 0.9307127
#> 42    yes     21 0.7750433 0.6327858 0.9036566
#> 43     no     22 0.8505034 0.7141132 0.9403575
#> 44    yes     22 0.7934296 0.6504343 0.9166629
#> 45     no     23 0.8638059 0.7278839 0.9487322
#> 46    yes     23 0.8106802 0.6676758 0.9280503
#> 47     no     24 0.8760972 0.7412326 0.9559845
#> 48    yes     24 0.8268046 0.6844746 0.9379845
#> 49     no     25 0.8874238 0.7541486 0.9622506
#> 50    yes     25 0.8418238 0.7005617 0.9466240
#> 51     no     26 0.8978357 0.7666239 0.9676538
#> 52    yes     26 0.8557676 0.7151428 0.9541171
#> 53     no     27 0.9073851 0.7786527 0.9723050
#> 54    yes     27 0.8686741 0.7292881 0.9606006
#> 55     no     28 0.9161253 0.7902320 0.9763030
#> 56    yes     28 0.8805867 0.7429832 0.9661992
#> 57     no     29 0.9241097 0.8013607 0.9797353
#> 58    yes     29 0.8915537 0.7562168 0.9710252
#> 59     no     30 0.9313909 0.8120400 0.9826787
#> 60    yes     30 0.9016260 0.7689808 0.9751791
#> 61     no     31 0.9380204 0.8222731 0.9852006
#> 62    yes     31 0.9108563 0.7812696 0.9787497
#> 63     no     32 0.9440478 0.8320648 0.9873597
#> 64    yes     32 0.9192981 0.7930805 0.9818157
#> 65     no     33 0.9495206 0.8414214 0.9892069
#> 66    yes     33 0.9270045 0.8044131 0.9844458
#> 67     no     34 0.9544839 0.8503509 0.9907864
#> 68    yes     34 0.9340278 0.8152695 0.9867003
#> 69     no     35 0.9589802 0.8588622 0.9921364
#> 70    yes     35 0.9404188 0.8256538 0.9886313
#> 71     no     36 0.9630496 0.8669655 0.9932897
#> 72    yes     36 0.9462263 0.8355721 0.9902845
#> 73     no     37 0.9667293 0.8746716 0.9942746
#> 74    yes     37 0.9514970 0.8450319 0.9916990
#> 75     no     38 0.9700539 0.8819923 0.9951156
#> 76    yes     38 0.9562749 0.8540424 0.9929088
#> 77     no     39 0.9730556 0.8889399 0.9958334
#> 78    yes     39 0.9606016 0.8626142 0.9939432
#> 79     no     40 0.9757639 0.8955270 0.9964460
#> 80    yes     40 0.9645161 0.8705903 0.9948273
#> 81     no     41 0.9782061 0.9017668 0.9969688
#> 82    yes     41 0.9680546 0.8781160 0.9955828
#> 83     no     42 0.9804072 0.9076723 0.9974147
#> 84    yes     42 0.9712507 0.8852619 0.9962282
#> 85     no     43 0.9823899 0.9132571 0.9977952
#> 86    yes     43 0.9741357 0.8920404 0.9967795
#> 87     no     44 0.9841753 0.9185344 0.9981197
#> 88    yes     44 0.9767380 0.8984645 0.9972504
#> 89     no     45 0.9857822 0.9235177 0.9983964
#> 90    yes     45 0.9790842 0.9045473 0.9976525
#> 91     no     46 0.9872281 0.9282199 0.9986325
#> 92    yes     46 0.9811982 0.9103021 0.9979959
#> 93     no     47 0.9885287 0.9326541 0.9988338
#> 94    yes     47 0.9831023 0.9157423 0.9982891
#> 95     no     48 0.9896982 0.9368331 0.9990055
#> 96    yes     48 0.9848165 0.9208812 0.9985394
#> 97     no     49 0.9907496 0.9407692 0.9991519
#> 98    yes     49 0.9863593 0.9257322 0.9987531
#> 99     no     50 0.9916946 0.9444746 0.9992767
#> 100   yes     50 0.9877472 0.9303082 0.9989355
#> 101    no     51 0.9925437 0.9479610 0.9993832
#> 102   yes     51 0.9889955 0.9346222 0.9990912
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)