Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3011325 0.4189030
#> 2     yes      1 0.2814584 0.2262932 0.3483794
#> 3      no      2 0.3927672 0.3350759 0.4374842
#> 4     yes      2 0.3039568 0.2486023 0.3749508
#> 5      no      3 0.4189750 0.3694915 0.4586766
#> 6     yes      3 0.3274340 0.2692829 0.4043167
#> 7      no      4 0.4456481 0.3955212 0.4838242
#> 8     yes      4 0.3518080 0.2882578 0.4314780
#> 9      no      5 0.4726379 0.4219177 0.5145519
#> 10    yes      5 0.3769795 0.3095489 0.4558718
#> 11     no      6 0.4997887 0.4452615 0.5429325
#> 12    yes      6 0.4028326 0.3372525 0.4836361
#> 13     no      7 0.5269407 0.4644301 0.5800282
#> 14    yes      7 0.4292372 0.3605838 0.5132440
#> 15     no      8 0.5539343 0.4823269 0.6134868
#> 16    yes      8 0.4560509 0.3800616 0.5427629
#> 17     no      9 0.5806134 0.5017628 0.6453160
#> 18    yes      9 0.4831217 0.3985870 0.5731779
#> 19     no     10 0.6068295 0.5237164 0.6791639
#> 20    yes     10 0.5102919 0.4170713 0.6081508
#> 21     no     11 0.6324443 0.5455787 0.7127030
#> 22    yes     11 0.5374015 0.4357916 0.6426615
#> 23     no     12 0.6573325 0.5672664 0.7462927
#> 24    yes     12 0.5642918 0.4546961 0.6788099
#> 25     no     13 0.6813846 0.5866343 0.7779218
#> 26    yes     13 0.5908091 0.4737305 0.7147104
#> 27     no     14 0.7045073 0.6037823 0.8066289
#> 28    yes     14 0.6168084 0.4928391 0.7480892
#> 29     no     15 0.7266252 0.6201545 0.8323678
#> 30    yes     15 0.6421564 0.5119656 0.7787730
#> 31     no     16 0.7476805 0.6363278 0.8548114
#> 32    yes     16 0.6667336 0.5306862 0.8066864
#> 33     no     17 0.7676327 0.6521986 0.8741609
#> 34    yes     17 0.6904365 0.5466796 0.8318391
#> 35     no     18 0.7864578 0.6677382 0.8912495
#> 36    yes     18 0.7131789 0.5625796 0.8543109
#> 37     no     19 0.8041469 0.6829208 0.9062643
#> 38    yes     19 0.7348919 0.5783539 0.8742341
#> 39     no     20 0.8207047 0.6977234 0.9193915
#> 40    yes     20 0.7555247 0.5939709 0.8910823
#> 41     no     21 0.8361480 0.7121260 0.9308189
#> 42    yes     21 0.7750433 0.6094005 0.9057714
#> 43     no     22 0.8505034 0.7261114 0.9407291
#> 44    yes     22 0.7934296 0.6246141 0.9185249
#> 45     no     23 0.8638059 0.7396655 0.9492956
#> 46    yes     23 0.8106802 0.6395846 0.9295824
#> 47     no     24 0.8760972 0.7527769 0.9566798
#> 48    yes     24 0.8268046 0.6543479 0.9392313
#> 49     no     25 0.8874238 0.7654369 0.9630294
#> 50    yes     25 0.8418238 0.6690782 0.9476261
#> 51     no     26 0.8978357 0.7776399 0.9684781
#> 52    yes     26 0.8557676 0.6834819 0.9549110
#> 53     no     27 0.9073851 0.7893823 0.9731454
#> 54    yes     27 0.8686741 0.6975391 0.9612189
#> 55     no     28 0.9161253 0.8006632 0.9771373
#> 56    yes     28 0.8805867 0.7112324 0.9666707
#> 57     no     29 0.9241097 0.8114839 0.9805472
#> 58    yes     29 0.8915537 0.7245469 0.9713750
#> 59     no     30 0.9313909 0.8218475 0.9834566
#> 60    yes     30 0.9016260 0.7374702 0.9754289
#> 61     no     31 0.9380204 0.8317592 0.9859367
#> 62    yes     31 0.9108563 0.7499922 0.9789183
#> 63     no     32 0.9440478 0.8412257 0.9880492
#> 64    yes     32 0.9192981 0.7621053 0.9822079
#> 65     no     33 0.9495206 0.8502551 0.9898903
#> 66    yes     33 0.9270045 0.7738040 0.9850203
#> 67     no     34 0.9544839 0.8588568 0.9914535
#> 68    yes     34 0.9340278 0.7850851 0.9873938
#> 69     no     35 0.9589802 0.8670415 0.9927768
#> 70    yes     35 0.9404188 0.7959474 0.9893953
#> 71     no     36 0.9630496 0.8748204 0.9938965
#> 72    yes     36 0.9462263 0.8063915 0.9910819
#> 73     no     37 0.9667293 0.8822058 0.9948435
#> 74    yes     37 0.9514970 0.8164197 0.9925022
#> 75     no     38 0.9700539 0.8892104 0.9956442
#> 76    yes     38 0.9562749 0.8260361 0.9936927
#> 77     no     39 0.9730556 0.8958474 0.9963210
#> 78    yes     39 0.9606016 0.8352461 0.9946933
#> 79     no     40 0.9757639 0.9021304 0.9968930
#> 80    yes     40 0.9645161 0.8440562 0.9955358
#> 81     no     41 0.9782061 0.9080730 0.9973762
#> 82    yes     41 0.9680546 0.8524742 0.9962449
#> 83     no     42 0.9804072 0.9136891 0.9977845
#> 84    yes     42 0.9712507 0.8605089 0.9968417
#> 85     no     43 0.9823899 0.9189926 0.9981294
#> 86    yes     43 0.9741357 0.8681697 0.9973439
#> 87     no     44 0.9841753 0.9239971 0.9984206
#> 88    yes     44 0.9767380 0.8754670 0.9977664
#> 89     no     45 0.9857822 0.9287162 0.9986666
#> 90    yes     45 0.9790842 0.8824115 0.9981217
#> 91     no     46 0.9872281 0.9331634 0.9988743
#> 92    yes     46 0.9811982 0.8890144 0.9984206
#> 93     no     47 0.9885287 0.9373517 0.9990497
#> 94    yes     47 0.9831023 0.8952873 0.9986720
#> 95     no     48 0.9896982 0.9412940 0.9991978
#> 96    yes     48 0.9848165 0.9012419 0.9988834
#> 97     no     49 0.9907496 0.9450026 0.9993228
#> 98    yes     49 0.9863593 0.9068902 0.9990611
#> 99     no     50 0.9916946 0.9484896 0.9994283
#> 100   yes     50 0.9877472 0.9122442 0.9992106
#> 101    no     51 0.9925437 0.9517667 0.9995174
#> 102   yes     51 0.9889955 0.9173157 0.9993363
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)