Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3163206 0.4291381
#> 2     yes      1 0.2814584 0.2080330 0.3595074
#> 3      no      2 0.3927672 0.3449225 0.4477323
#> 4     yes      2 0.3039568 0.2348616 0.3788955
#> 5      no      3 0.4189750 0.3725919 0.4727949
#> 6     yes      3 0.3274340 0.2646620 0.3997836
#> 7      no      4 0.4456481 0.3956149 0.5005069
#> 8     yes      4 0.3518080 0.2904902 0.4234011
#> 9      no      5 0.4726379 0.4218524 0.5326010
#> 10    yes      5 0.3769795 0.3169038 0.4493182
#> 11     no      6 0.4997887 0.4438255 0.5656984
#> 12    yes      6 0.4028326 0.3453725 0.4755130
#> 13     no      7 0.5269407 0.4641295 0.5972715
#> 14    yes      7 0.4292372 0.3607270 0.5030324
#> 15     no      8 0.5539343 0.4843920 0.6375241
#> 16    yes      8 0.4560509 0.3781060 0.5318001
#> 17     no      9 0.5806134 0.5007880 0.6770259
#> 18    yes      9 0.4831217 0.3963274 0.5603560
#> 19     no     10 0.6068295 0.5171845 0.7141470
#> 20    yes     10 0.5102919 0.4143465 0.5900539
#> 21     no     11 0.6324443 0.5338147 0.7485787
#> 22    yes     11 0.5374015 0.4311565 0.6241715
#> 23     no     12 0.6573325 0.5507168 0.7801322
#> 24    yes     12 0.5642918 0.4481250 0.6598984
#> 25     no     13 0.6813846 0.5674985 0.8087297
#> 26    yes     13 0.5908091 0.4652117 0.6960022
#> 27     no     14 0.7045073 0.5841207 0.8343899
#> 28    yes     14 0.6168084 0.4823750 0.7299123
#> 29     no     15 0.7266252 0.6005456 0.8572085
#> 30    yes     15 0.6421564 0.4995730 0.7615339
#> 31     no     16 0.7476805 0.6157339 0.8773393
#> 32    yes     16 0.6667336 0.5167631 0.7886511
#> 33     no     17 0.7676327 0.6306770 0.8949744
#> 34    yes     17 0.6904365 0.5339030 0.8156934
#> 35     no     18 0.7864578 0.6453682 0.9103287
#> 36    yes     18 0.7131789 0.5509511 0.8400224
#> 37     no     19 0.8041469 0.6590625 0.9236259
#> 38    yes     19 0.7348919 0.5649930 0.8616152
#> 39     no     20 0.8207047 0.6716000 0.9350887
#> 40    yes     20 0.7555247 0.5785408 0.8803418
#> 41     no     21 0.8361480 0.6838983 0.9449309
#> 42    yes     21 0.7750433 0.5919749 0.8973816
#> 43     no     22 0.8505034 0.6959453 0.9533530
#> 44    yes     22 0.7934296 0.6052761 0.9126087
#> 45     no     23 0.8638059 0.7077301 0.9605390
#> 46    yes     23 0.8106802 0.6184260 0.9257610
#> 47     no     24 0.8760972 0.7192430 0.9666552
#> 48    yes     24 0.8268046 0.6314073 0.9370296
#> 49     no     25 0.8874238 0.7304757 0.9718499
#> 50    yes     25 0.8418238 0.6442033 0.9466241
#> 51     no     26 0.8978357 0.7414211 0.9762541
#> 52    yes     26 0.8557676 0.6567986 0.9548253
#> 53     no     27 0.9073851 0.7520734 0.9799826
#> 54    yes     27 0.8686741 0.6691787 0.9618156
#> 55     no     28 0.9161253 0.7624279 0.9831350
#> 56    yes     28 0.8805867 0.6813304 0.9677593
#> 57     no     29 0.9241097 0.7724812 0.9857976
#> 58    yes     29 0.8915537 0.6932415 0.9728028
#> 59     no     30 0.9313909 0.7822310 0.9880444
#> 60    yes     30 0.9016260 0.7049013 0.9770749
#> 61     no     31 0.9380204 0.7916760 0.9899277
#> 62    yes     31 0.9108563 0.7163000 0.9806884
#> 63     no     32 0.9440478 0.8008159 0.9915131
#> 64    yes     32 0.9192981 0.7274291 0.9837410
#> 65     no     33 0.9495206 0.8096516 0.9928506
#> 66    yes     33 0.9270045 0.7382815 0.9863172
#> 67     no     34 0.9544839 0.8181845 0.9939784
#> 68    yes     34 0.9340278 0.7488510 0.9884861
#> 69     no     35 0.9589802 0.8264171 0.9949292
#> 70    yes     35 0.9404188 0.7591329 0.9903003
#> 71     no     36 0.9630496 0.8343526 0.9957304
#> 72    yes     36 0.9462263 0.7691234 0.9918308
#> 73     no     37 0.9667293 0.8419949 0.9964054
#> 74    yes     37 0.9514970 0.7788199 0.9931213
#> 75     no     38 0.9700539 0.8493483 0.9969740
#> 76    yes     38 0.9562749 0.7882208 0.9942090
#> 77     no     39 0.9730556 0.8564180 0.9974528
#> 78    yes     39 0.9606016 0.7973254 0.9951255
#> 79     no     40 0.9757639 0.8632095 0.9978560
#> 80    yes     40 0.9645161 0.8061342 0.9958975
#> 81     no     41 0.9782061 0.8697287 0.9981954
#> 82    yes     41 0.9680546 0.8146483 0.9965475
#> 83     no     42 0.9804072 0.8759818 0.9984812
#> 84    yes     42 0.9712507 0.8228698 0.9970948
#> 85     no     43 0.9823899 0.8819756 0.9987217
#> 86    yes     43 0.9741357 0.8308015 0.9975555
#> 87     no     44 0.9841753 0.8877169 0.9989242
#> 88    yes     44 0.9767380 0.8384467 0.9979432
#> 89     no     45 0.9857822 0.8932127 0.9990946
#> 90    yes     45 0.9790842 0.8458981 0.9982695
#> 91     no     46 0.9872281 0.8984703 0.9992380
#> 92    yes     46 0.9811982 0.8530951 0.9985441
#> 93     no     47 0.9885287 0.9034971 0.9993587
#> 94    yes     47 0.9831023 0.8600105 0.9987752
#> 95     no     48 0.9896982 0.9083003 0.9994603
#> 96    yes     48 0.9848165 0.8666503 0.9989696
#> 97     no     49 0.9907496 0.9128875 0.9995458
#> 98    yes     49 0.9863593 0.8730205 0.9991331
#> 99     no     50 0.9916946 0.9172662 0.9996178
#> 100   yes     50 0.9877472 0.8791278 0.9992707
#> 101    no     51 0.9925437 0.9214436 0.9996783
#> 102   yes     51 0.9889955 0.8849790 0.9993865
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)