Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3015925 0.4198003
#> 2     yes      1 0.2814584 0.2068059 0.3545018
#> 3      no      2 0.3927672 0.3397528 0.4355246
#> 4     yes      2 0.3039568 0.2352974 0.3725993
#> 5      no      3 0.4189750 0.3692029 0.4561062
#> 6     yes      3 0.3274340 0.2624280 0.3932666
#> 7      no      4 0.4456481 0.3981446 0.4874937
#> 8     yes      4 0.3518080 0.2849908 0.4151610
#> 9      no      5 0.4726379 0.4225385 0.5232776
#> 10    yes      5 0.3769795 0.3080116 0.4405131
#> 11     no      6 0.4997887 0.4471358 0.5593056
#> 12    yes      6 0.4028326 0.3339378 0.4681653
#> 13     no      7 0.5269407 0.4687097 0.5972246
#> 14    yes      7 0.4292372 0.3586070 0.5044587
#> 15     no      8 0.5539343 0.4886706 0.6340194
#> 16    yes      8 0.4560509 0.3839408 0.5435870
#> 17     no      9 0.5806134 0.5048674 0.6693162
#> 18    yes      9 0.4831217 0.4043739 0.5821553
#> 19     no     10 0.6068295 0.5254074 0.7028053
#> 20    yes     10 0.5102919 0.4248642 0.6190333
#> 21     no     11 0.6324443 0.5462578 0.7342491
#> 22    yes     11 0.5374015 0.4469126 0.6563327
#> 23     no     12 0.6573325 0.5669472 0.7664205
#> 24    yes     12 0.5642918 0.4675241 0.6919085
#> 25     no     13 0.6813846 0.5863060 0.7974485
#> 26    yes     13 0.5908091 0.4838175 0.7236778
#> 27     no     14 0.7045073 0.6042116 0.8243320
#> 28    yes     14 0.6168084 0.4994851 0.7531815
#> 29     no     15 0.7266252 0.6216238 0.8478609
#> 30    yes     15 0.6421564 0.5155417 0.7803091
#> 31     no     16 0.7476805 0.6371417 0.8687131
#> 32    yes     16 0.6667336 0.5306835 0.8082715
#> 33     no     17 0.7676327 0.6521511 0.8867881
#> 34    yes     17 0.6904365 0.5451973 0.8337635
#> 35     no     18 0.7864578 0.6668603 0.9026440
#> 36    yes     18 0.7131789 0.5604205 0.8564158
#> 37     no     19 0.8041469 0.6812467 0.9164793
#> 38    yes     19 0.7348919 0.5762514 0.8763823
#> 39     no     20 0.8207047 0.6952902 0.9284961
#> 40    yes     20 0.7555247 0.5919243 0.8933660
#> 41     no     21 0.8361480 0.7089732 0.9388918
#> 42    yes     21 0.7750433 0.6074087 0.9082498
#> 43     no     22 0.8505034 0.7222805 0.9478546
#> 44    yes     22 0.7934296 0.6226755 0.9217290
#> 45     no     23 0.8638059 0.7351995 0.9555594
#> 46    yes     23 0.8106802 0.6376977 0.9335435
#> 47     no     24 0.8760972 0.7477201 0.9623309
#> 48    yes     24 0.8268046 0.6524499 0.9436858
#> 49     no     25 0.8874238 0.7598344 0.9682559
#> 50    yes     25 0.8418238 0.6669090 0.9523609
#> 51     no     26 0.8978357 0.7715368 0.9732746
#> 52    yes     26 0.8557676 0.6810540 0.9597579
#> 53     no     27 0.9073851 0.7828241 0.9775180
#> 54    yes     27 0.8686741 0.6948662 0.9660482
#> 55     no     28 0.9161253 0.7936948 0.9811005
#> 56    yes     28 0.8805867 0.7083291 0.9713852
#> 57     no     29 0.9241097 0.8041495 0.9841212
#> 58    yes     29 0.8915537 0.7214287 0.9759047
#> 59     no     30 0.9313909 0.8141905 0.9866656
#> 60    yes     30 0.9016260 0.7341532 0.9797256
#> 61     no     31 0.9380204 0.8238217 0.9888068
#> 62    yes     31 0.9108563 0.7464930 0.9829514
#> 63     no     32 0.9440478 0.8330485 0.9906073
#> 64    yes     32 0.9192981 0.7584407 0.9856716
#> 65     no     33 0.9495206 0.8413220 0.9921204
#> 66    yes     33 0.9270045 0.7699911 0.9879632
#> 67     no     34 0.9544839 0.8492546 0.9933913
#> 68    yes     34 0.9340278 0.7811410 0.9898921
#> 69     no     35 0.9589802 0.8568565 0.9944583
#> 70    yes     35 0.9404188 0.7918890 0.9915146
#> 71     no     36 0.9630496 0.8641349 0.9953538
#> 72    yes     36 0.9462263 0.8022353 0.9928785
#> 73     no     37 0.9667293 0.8710975 0.9961051
#> 74    yes     37 0.9514970 0.8119024 0.9940245
#> 75     no     38 0.9700539 0.8777525 0.9967353
#> 76    yes     38 0.9562749 0.8209869 0.9949870
#> 77     no     39 0.9730556 0.8841083 0.9972638
#> 78    yes     39 0.9606016 0.8297227 0.9957951
#> 79     no     40 0.9757639 0.8901739 0.9977069
#> 80    yes     40 0.9645161 0.8381143 0.9964734
#> 81     no     41 0.9782061 0.8959583 0.9980783
#> 82    yes     41 0.9680546 0.8461674 0.9970426
#> 83     no     42 0.9804072 0.9014708 0.9983897
#> 84    yes     42 0.9712507 0.8538880 0.9975202
#> 85     no     43 0.9823899 0.9067206 0.9986507
#> 86    yes     43 0.9741357 0.8612831 0.9979207
#> 87     no     44 0.9841753 0.9117173 0.9988694
#> 88    yes     44 0.9767380 0.8683603 0.9982567
#> 89     no     45 0.9857822 0.9164701 0.9990527
#> 90    yes     45 0.9790842 0.8751273 0.9985384
#> 91     no     46 0.9872281 0.9209884 0.9992062
#> 92    yes     46 0.9811982 0.8815927 0.9987747
#> 93     no     47 0.9885287 0.9252816 0.9993349
#> 94    yes     47 0.9831023 0.8877651 0.9989728
#> 95     no     48 0.9896982 0.9293586 0.9994428
#> 96    yes     48 0.9848165 0.8936534 0.9991389
#> 97     no     49 0.9907496 0.9332287 0.9995331
#> 98    yes     49 0.9863593 0.8992669 0.9992782
#> 99     no     50 0.9916946 0.9369005 0.9996088
#> 100   yes     50 0.9877472 0.9046147 0.9993949
#> 101    no     51 0.9925437 0.9403827 0.9996723
#> 102   yes     51 0.9889955 0.9097062 0.9994928
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)