Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3071876 0.4243639
#> 2     yes      1 0.2814584 0.2003871 0.3536301
#> 3      no      2 0.3927672 0.3418143 0.4465518
#> 4     yes      2 0.3039568 0.2248117 0.3670896
#> 5      no      3 0.4189750 0.3758804 0.4739177
#> 6     yes      3 0.3274340 0.2463667 0.3875159
#> 7      no      4 0.4456481 0.4051007 0.4971282
#> 8     yes      4 0.3518080 0.2693351 0.4105434
#> 9      no      5 0.4726379 0.4312972 0.5190324
#> 10    yes      5 0.3769795 0.2936670 0.4347541
#> 11     no      6 0.4997887 0.4508426 0.5475083
#> 12    yes      6 0.4028326 0.3200927 0.4602139
#> 13     no      7 0.5269407 0.4720817 0.5760260
#> 14    yes      7 0.4292372 0.3499761 0.4901424
#> 15     no      8 0.5539343 0.4907548 0.6077664
#> 16    yes      8 0.4560509 0.3768598 0.5260310
#> 17     no      9 0.5806134 0.5096614 0.6402447
#> 18    yes      9 0.4831217 0.3987471 0.5612449
#> 19     no     10 0.6068295 0.5289931 0.6740228
#> 20    yes     10 0.5102919 0.4149331 0.5922526
#> 21     no     11 0.6324443 0.5462714 0.7036235
#> 22    yes     11 0.5374015 0.4337613 0.6225441
#> 23     no     12 0.6573325 0.5619923 0.7348924
#> 24    yes     12 0.5642918 0.4544699 0.6519090
#> 25     no     13 0.6813846 0.5775901 0.7634532
#> 26    yes     13 0.5908091 0.4695260 0.6801641
#> 27     no     14 0.7045073 0.5918514 0.7891526
#> 28    yes     14 0.6168084 0.4849422 0.7071567
#> 29     no     15 0.7266252 0.6060200 0.8131938
#> 30    yes     15 0.6421564 0.5008914 0.7327668
#> 31     no     16 0.7476805 0.6203000 0.8363320
#> 32    yes     16 0.6667336 0.5168356 0.7591686
#> 33     no     17 0.7676327 0.6343713 0.8571079
#> 34    yes     17 0.6904365 0.5338296 0.7851795
#> 35     no     18 0.7864578 0.6481356 0.8756382
#> 36    yes     18 0.7131789 0.5508887 0.8105891
#> 37     no     19 0.8041469 0.6607715 0.8920675
#> 38    yes     19 0.7348919 0.5678237 0.8335907
#> 39     no     20 0.8207047 0.6727775 0.9065573
#> 40    yes     20 0.7555247 0.5845955 0.8542982
#> 41     no     21 0.8361480 0.6845609 0.9192770
#> 42    yes     21 0.7750433 0.6011669 0.8728191
#> 43     no     22 0.8505034 0.6961111 0.9303975
#> 44    yes     22 0.7934296 0.6166228 0.8892880
#> 45     no     23 0.8638059 0.7074185 0.9400380
#> 46    yes     23 0.8106802 0.6302591 0.9038567
#> 47     no     24 0.8760972 0.7184748 0.9484063
#> 48    yes     24 0.8268046 0.6436925 0.9166855
#> 49     no     25 0.8874238 0.7292726 0.9556612
#> 50    yes     25 0.8418238 0.6569052 0.9279369
#> 51     no     26 0.8978357 0.7398056 0.9619363
#> 52    yes     26 0.8557676 0.6698807 0.9377702
#> 53     no     27 0.9073851 0.7500687 0.9673533
#> 54    yes     27 0.8686741 0.6826038 0.9463377
#> 55     no     28 0.9161253 0.7600577 0.9720213
#> 56    yes     28 0.8805867 0.6950608 0.9537824
#> 57     no     29 0.9241097 0.7697695 0.9760380
#> 58    yes     29 0.8915537 0.7072393 0.9602365
#> 59     no     30 0.9313909 0.7792018 0.9794900
#> 60    yes     30 0.9016260 0.7191285 0.9658205
#> 61     no     31 0.9380204 0.7883534 0.9824533
#> 62    yes     31 0.9108563 0.7307191 0.9706436
#> 63     no     32 0.9440478 0.7972239 0.9849948
#> 64    yes     32 0.9192981 0.7420031 0.9748030
#> 65     no     33 0.9495206 0.8058137 0.9871728
#> 66    yes     33 0.9270045 0.7529742 0.9783856
#> 67     no     34 0.9544839 0.8141239 0.9890380
#> 68    yes     34 0.9340278 0.7636273 0.9814680
#> 69     no     35 0.9589802 0.8221565 0.9906345
#> 70    yes     35 0.9404188 0.7739586 0.9841175
#> 71     no     36 0.9630496 0.8299141 0.9920002
#> 72    yes     36 0.9462263 0.7839658 0.9863931
#> 73     no     37 0.9667293 0.8373997 0.9931680
#> 74    yes     37 0.9514970 0.7936475 0.9883461
#> 75     no     38 0.9700539 0.8446172 0.9941662
#> 76    yes     38 0.9562749 0.8030039 0.9900214
#> 77     no     39 0.9730556 0.8515707 0.9950193
#> 78    yes     39 0.9606016 0.8120358 0.9914577
#> 79     no     40 0.9757639 0.8582648 0.9957481
#> 80    yes     40 0.9645161 0.8207454 0.9926886
#> 81     no     41 0.9782061 0.8647047 0.9963706
#> 82    yes     41 0.9680546 0.8291356 0.9937431
#> 83     no     42 0.9804072 0.8708956 0.9969022
#> 84    yes     42 0.9712507 0.8372101 0.9946462
#> 85     no     43 0.9823899 0.8768432 0.9973561
#> 86    yes     43 0.9741357 0.8450381 0.9954194
#> 87     no     44 0.9841753 0.8825534 0.9977436
#> 88    yes     44 0.9767380 0.8525983 0.9960812
#> 89     no     45 0.9857822 0.8880324 0.9980744
#> 90    yes     45 0.9790842 0.8598499 0.9966477
#> 91     no     46 0.9872281 0.8932865 0.9983567
#> 92    yes     46 0.9811982 0.8667996 0.9971325
#> 93     no     47 0.9885287 0.8983219 0.9985977
#> 94    yes     47 0.9831023 0.8729417 0.9975472
#> 95     no     48 0.9896982 0.9031452 0.9988034
#> 96    yes     48 0.9848165 0.8787755 0.9979021
#> 97     no     49 0.9907496 0.9077630 0.9989789
#> 98    yes     49 0.9863593 0.8843767 0.9982056
#> 99     no     50 0.9916946 0.9121818 0.9991287
#> 100   yes     50 0.9877472 0.8897514 0.9984653
#> 101    no     51 0.9925437 0.9164082 0.9992565
#> 102   yes     51 0.9889955 0.8949059 0.9986874
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)