Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3071283 0.4242495
#> 2     yes      1 0.2814584 0.2168522 0.3432152
#> 3      no      2 0.3927672 0.3345418 0.4431901
#> 4     yes      2 0.3039568 0.2455755 0.3615987
#> 5      no      3 0.4189750 0.3658445 0.4663921
#> 6     yes      3 0.3274340 0.2740591 0.3817717
#> 7      no      4 0.4456481 0.3975279 0.4926434
#> 8     yes      4 0.3518080 0.3007862 0.4056788
#> 9      no      5 0.4726379 0.4243674 0.5252028
#> 10    yes      5 0.3769795 0.3224665 0.4358019
#> 11     no      6 0.4997887 0.4484860 0.5535267
#> 12    yes      6 0.4028326 0.3432172 0.4722972
#> 13     no      7 0.5269407 0.4679535 0.5853532
#> 14    yes      7 0.4292372 0.3650723 0.5071968
#> 15     no      8 0.5539343 0.4871557 0.6174235
#> 16    yes      8 0.4560509 0.3890453 0.5380544
#> 17     no      9 0.5806134 0.5055890 0.6516570
#> 18    yes      9 0.4831217 0.4116077 0.5686114
#> 19     no     10 0.6068295 0.5240061 0.6882843
#> 20    yes     10 0.5102919 0.4328954 0.5997659
#> 21     no     11 0.6324443 0.5419809 0.7199929
#> 22    yes     11 0.5374015 0.4523211 0.6355062
#> 23     no     12 0.6573325 0.5572373 0.7489265
#> 24    yes     12 0.5642918 0.4714119 0.6698166
#> 25     no     13 0.6813846 0.5723864 0.7757878
#> 26    yes     13 0.5908091 0.4890751 0.7024091
#> 27     no     14 0.7045073 0.5874384 0.8005260
#> 28    yes     14 0.6168084 0.5079347 0.7330620
#> 29     no     15 0.7266252 0.6031084 0.8231420
#> 30    yes     15 0.6421564 0.5260912 0.7612831
#> 31     no     16 0.7476805 0.6198579 0.8442367
#> 32    yes     16 0.6667336 0.5433859 0.7872353
#> 33     no     17 0.7676327 0.6363267 0.8633292
#> 34    yes     17 0.6904365 0.5609570 0.8110551
#> 35     no     18 0.7864578 0.6524819 0.8805877
#> 36    yes     18 0.7131789 0.5783805 0.8343663
#> 37     no     19 0.8041469 0.6683543 0.8964349
#> 38    yes     19 0.7348919 0.5956086 0.8560125
#> 39     no     20 0.8207047 0.6838851 0.9103937
#> 40    yes     20 0.7555247 0.6120800 0.8752520
#> 41     no     21 0.8361480 0.6990158 0.9226340
#> 42    yes     21 0.7750433 0.6278831 0.8922434
#> 43     no     22 0.8505034 0.7137252 0.9333252
#> 44    yes     22 0.7934296 0.6434165 0.9071648
#> 45     no     23 0.8638059 0.7279952 0.9426312
#> 46    yes     23 0.8106802 0.6586528 0.9202035
#> 47     no     24 0.8760972 0.7418108 0.9507072
#> 48    yes     24 0.8268046 0.6735672 0.9315191
#> 49     no     25 0.8874238 0.7551604 0.9576975
#> 50    yes     25 0.8418238 0.6881373 0.9412850
#> 51     no     26 0.8978357 0.7680352 0.9637344
#> 52    yes     26 0.8557676 0.7023431 0.9495653
#> 53     no     27 0.9073851 0.7804294 0.9689379
#> 54    yes     27 0.8686741 0.7161675 0.9565762
#> 55     no     28 0.9161253 0.7923401 0.9734155
#> 56    yes     28 0.8805867 0.7295957 0.9626504
#> 57     no     29 0.9241097 0.8037668 0.9772629
#> 58    yes     29 0.8915537 0.7426156 0.9678736
#> 59     no     30 0.9313909 0.8147113 0.9805646
#> 60    yes     30 0.9016260 0.7552177 0.9723865
#> 61     no     31 0.9380204 0.8251778 0.9833950
#> 62    yes     31 0.9108563 0.7673947 0.9762808
#> 63     no     32 0.9440478 0.8351724 0.9858192
#> 64    yes     32 0.9192981 0.7791419 0.9796371
#> 65     no     33 0.9495206 0.8447027 0.9878939
#> 66    yes     33 0.9270045 0.7904566 0.9825267
#> 67     no     34 0.9544839 0.8537781 0.9896683
#> 68    yes     34 0.9340278 0.8013383 0.9850124
#> 69     no     35 0.9589802 0.8619878 0.9911849
#> 70    yes     35 0.9404188 0.8117883 0.9871490
#> 71     no     36 0.9630496 0.8697111 0.9924806
#> 72    yes     36 0.9462263 0.8218097 0.9889842
#> 73     no     37 0.9667293 0.8770636 0.9935871
#> 74    yes     37 0.9514970 0.8314073 0.9905597
#> 75     no     38 0.9700539 0.8840562 0.9945316
#> 76    yes     38 0.9562749 0.8405873 0.9919117
#> 77     no     39 0.9730556 0.8907003 0.9953377
#> 78    yes     39 0.9606016 0.8493570 0.9930713
#> 79     no     40 0.9757639 0.8970077 0.9960255
#> 80    yes     40 0.9645161 0.8577250 0.9940655
#> 81     no     41 0.9782061 0.9029904 0.9966121
#> 82    yes     41 0.9680546 0.8657009 0.9949178
#> 83     no     42 0.9804072 0.9086607 0.9971124
#> 84    yes     42 0.9712507 0.8732949 0.9956481
#> 85     no     43 0.9823899 0.9140308 0.9975390
#> 86    yes     43 0.9741357 0.8805179 0.9962738
#> 87     no     44 0.9841753 0.9191130 0.9979027
#> 88    yes     44 0.9767380 0.8873816 0.9968099
#> 89     no     45 0.9857822 0.9239250 0.9982128
#> 90    yes     45 0.9790842 0.8938979 0.9972689
#> 91     no     46 0.9872281 0.9285092 0.9984771
#> 92    yes     46 0.9811982 0.9000789 0.9976621
#> 93     no     47 0.9885287 0.9328369 0.9987023
#> 94    yes     47 0.9831023 0.9059372 0.9979987
#> 95     no     48 0.9896982 0.9369202 0.9988943
#> 96    yes     48 0.9848165 0.9114852 0.9982870
#> 97     no     49 0.9907496 0.9407707 0.9990579
#> 98    yes     49 0.9863593 0.9167356 0.9985337
#> 99     no     50 0.9916946 0.9443998 0.9991973
#> 100   yes     50 0.9877472 0.9217008 0.9987450
#> 101    no     51 0.9925437 0.9478186 0.9993161
#> 102   yes     51 0.9889955 0.9263932 0.9989258
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)