Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3125181 0.4260705
#> 2     yes      1 0.2814584 0.2136428 0.3422407
#> 3      no      2 0.3927672 0.3433337 0.4441944
#> 4     yes      2 0.3039568 0.2367677 0.3594541
#> 5      no      3 0.4189750 0.3752775 0.4640133
#> 6     yes      3 0.3274340 0.2587821 0.3770615
#> 7      no      4 0.4456481 0.4040745 0.4856788
#> 8     yes      4 0.3518080 0.2820950 0.3950343
#> 9      no      5 0.4726379 0.4297568 0.5189031
#> 10    yes      5 0.3769795 0.3047797 0.4183570
#> 11     no      6 0.4997887 0.4519875 0.5539981
#> 12    yes      6 0.4028326 0.3287310 0.4534815
#> 13     no      7 0.5269407 0.4743035 0.5885577
#> 14    yes      7 0.4292372 0.3541245 0.4846435
#> 15     no      8 0.5539343 0.4967230 0.6250000
#> 16    yes      8 0.4560509 0.3773774 0.5263682
#> 17     no      9 0.5806134 0.5183736 0.6634136
#> 18    yes      9 0.4831217 0.3982898 0.5682661
#> 19     no     10 0.6068295 0.5393564 0.6988868
#> 20    yes     10 0.5102919 0.4200996 0.6092119
#> 21     no     11 0.6324443 0.5580798 0.7296811
#> 22    yes     11 0.5374015 0.4431428 0.6479161
#> 23     no     12 0.6573325 0.5739064 0.7584140
#> 24    yes     12 0.5642918 0.4630165 0.6831228
#> 25     no     13 0.6813846 0.5894969 0.7849921
#> 26    yes     13 0.5908091 0.4853785 0.7159737
#> 27     no     14 0.7045073 0.6049038 0.8101628
#> 28    yes     14 0.6168084 0.5039332 0.7466787
#> 29     no     15 0.7266252 0.6200976 0.8339365
#> 30    yes     15 0.6421564 0.5212135 0.7751038
#> 31     no     16 0.7476805 0.6350506 0.8551234
#> 32    yes     16 0.6667336 0.5391953 0.8011856
#> 33     no     17 0.7676327 0.6515634 0.8738798
#> 34    yes     17 0.6904365 0.5572392 0.8249229
#> 35     no     18 0.7864578 0.6679401 0.8905129
#> 36    yes     18 0.7131789 0.5751335 0.8463670
#> 37     no     19 0.8041469 0.6835747 0.9051847
#> 38    yes     19 0.7348919 0.5928331 0.8656101
#> 39     no     20 0.8207047 0.6973249 0.9180661
#> 40    yes     20 0.7555247 0.6101749 0.8827747
#> 41     no     21 0.8361480 0.7107276 0.9293294
#> 42    yes     21 0.7750433 0.6250230 0.8980035
#> 43     no     22 0.8505034 0.7237678 0.9391428
#> 44    yes     22 0.7934296 0.6401643 0.9116806
#> 45     no     23 0.8638059 0.7364329 0.9476665
#> 46    yes     23 0.8106802 0.6550238 0.9241844
#> 47     no     24 0.8760972 0.7488147 0.9550503
#> 48    yes     24 0.8268046 0.6695768 0.9350440
#> 49     no     25 0.8874238 0.7611262 0.9614320
#> 50    yes     25 0.8418238 0.6838010 0.9444413
#> 51     no     26 0.8978357 0.7730094 0.9669366
#> 52    yes     26 0.8557676 0.6976768 0.9525476
#> 53     no     27 0.9073851 0.7844606 0.9717336
#> 54    yes     27 0.8686741 0.7111871 0.9595214
#> 55     no     28 0.9161253 0.7954782 0.9759167
#> 56    yes     28 0.8805867 0.7243174 0.9655070
#> 57     no     29 0.9241097 0.8060627 0.9794935
#> 58    yes     29 0.8915537 0.7370559 0.9706342
#> 59     no     30 0.9313909 0.8162166 0.9825481
#> 60    yes     30 0.9016260 0.7493143 0.9750185
#> 61     no     31 0.9380204 0.8259442 0.9851544
#> 62    yes     31 0.9108563 0.7610613 0.9787622
#> 63     no     32 0.9440478 0.8352514 0.9873763
#> 64    yes     32 0.9192981 0.7724107 0.9819550
#> 65     no     33 0.9495206 0.8441453 0.9892690
#> 66    yes     33 0.9270045 0.7833599 0.9846750
#> 67     no     34 0.9544839 0.8526343 0.9908804
#> 68    yes     34 0.9340278 0.7939085 0.9869901
#> 69     no     35 0.9589802 0.8607281 0.9922515
#> 70    yes     35 0.9404188 0.8040575 0.9889495
#> 71     no     36 0.9630496 0.8684369 0.9934178
#> 72    yes     36 0.9462263 0.8138097 0.9906134
#> 73     no     37 0.9667293 0.8757717 0.9944094
#> 74    yes     37 0.9514970 0.8231694 0.9920286
#> 75     no     38 0.9700539 0.8827444 0.9952523
#> 76    yes     38 0.9562749 0.8321420 0.9932318
#> 77     no     39 0.9730556 0.8893669 0.9959685
#> 78    yes     39 0.9606016 0.8407342 0.9942543
#> 79     no     40 0.9757639 0.8956517 0.9965770
#> 80    yes     40 0.9645161 0.8489536 0.9951230
#> 81     no     41 0.9782061 0.9016114 0.9970938
#> 82    yes     41 0.9680546 0.8568087 0.9958608
#> 83     no     42 0.9804072 0.9072588 0.9975328
#> 84    yes     42 0.9712507 0.8643086 0.9964873
#> 85     no     43 0.9823899 0.9126065 0.9979056
#> 86    yes     43 0.9741357 0.8714633 0.9970193
#> 87     no     44 0.9841753 0.9176673 0.9982221
#> 88    yes     44 0.9767380 0.8782830 0.9974708
#> 89     no     45 0.9857822 0.9224537 0.9984908
#> 90    yes     45 0.9790842 0.8847783 0.9978540
#> 91     no     46 0.9872281 0.9269782 0.9987189
#> 92    yes     46 0.9811982 0.8909601 0.9981792
#> 93     no     47 0.9885287 0.9312527 0.9989126
#> 94    yes     47 0.9831023 0.8968395 0.9984552
#> 95     no     48 0.9896982 0.9352893 0.9990770
#> 96    yes     48 0.9848165 0.9024277 0.9986894
#> 97     no     49 0.9907496 0.9390994 0.9992166
#> 98    yes     49 0.9863593 0.9077360 0.9988881
#> 99     no     50 0.9916946 0.9426943 0.9993350
#> 100   yes     50 0.9877472 0.9127755 0.9990567
#> 101    no     51 0.9925437 0.9460848 0.9994356
#> 102   yes     51 0.9889955 0.9175572 0.9991997
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)