Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3124276 0.4286047
#> 2     yes      1 0.2814584 0.2179414 0.3491211
#> 3      no      2 0.3927672 0.3450571 0.4445562
#> 4     yes      2 0.3039568 0.2428351 0.3668474
#> 5      no      3 0.4189750 0.3724858 0.4604865
#> 6     yes      3 0.3274340 0.2694128 0.3852088
#> 7      no      4 0.4456481 0.4015733 0.4883196
#> 8     yes      4 0.3518080 0.2934943 0.4052822
#> 9      no      5 0.4726379 0.4243966 0.5117128
#> 10    yes      5 0.3769795 0.3124107 0.4277222
#> 11     no      6 0.4997887 0.4437179 0.5464992
#> 12    yes      6 0.4028326 0.3355465 0.4578237
#> 13     no      7 0.5269407 0.4622583 0.5839798
#> 14    yes      7 0.4292372 0.3574023 0.4929129
#> 15     no      8 0.5539343 0.4809038 0.6180031
#> 16    yes      8 0.4560509 0.3744147 0.5250934
#> 17     no      9 0.5806134 0.4996028 0.6502500
#> 18    yes      9 0.4831217 0.3908075 0.5592236
#> 19     no     10 0.6068295 0.5183029 0.6839400
#> 20    yes     10 0.5102919 0.4074569 0.5961628
#> 21     no     11 0.6324443 0.5369520 0.7182902
#> 22    yes     11 0.5374015 0.4243264 0.6317273
#> 23     no     12 0.6573325 0.5554984 0.7497279
#> 24    yes     12 0.5642918 0.4415261 0.6641065
#> 25     no     13 0.6813846 0.5738916 0.7787255
#> 26    yes     13 0.5908091 0.4592641 0.6951413
#> 27     no     14 0.7045073 0.5908221 0.8052219
#> 28    yes     14 0.6168084 0.4752652 0.7245654
#> 29     no     15 0.7266252 0.6061420 0.8304364
#> 30    yes     15 0.6421564 0.4895062 0.7532576
#> 31     no     16 0.7476805 0.6216887 0.8530168
#> 32    yes     16 0.6667336 0.5037649 0.7798671
#> 33     no     17 0.7676327 0.6369801 0.8730482
#> 34    yes     17 0.6904365 0.5181462 0.8043509
#> 35     no     18 0.7864578 0.6519879 0.8906982
#> 36    yes     18 0.7131789 0.5326816 0.8269929
#> 37     no     19 0.8041469 0.6649995 0.9061568
#> 38    yes     19 0.7348919 0.5489433 0.8487814
#> 39     no     20 0.8207047 0.6777056 0.9196255
#> 40    yes     20 0.7555247 0.5652763 0.8682638
#> 41     no     21 0.8361480 0.6901512 0.9313068
#> 42    yes     21 0.7750433 0.5816640 0.8855740
#> 43     no     22 0.8505034 0.7023240 0.9413980
#> 44    yes     22 0.7934296 0.5978725 0.9008678
#> 45     no     23 0.8638059 0.7142132 0.9500856
#> 46    yes     23 0.8106802 0.6138686 0.9143133
#> 47     no     24 0.8760972 0.7258094 0.9575429
#> 48    yes     24 0.8268046 0.6296208 0.9260825
#> 49     no     25 0.8874238 0.7371047 0.9639280
#> 50    yes     25 0.8418238 0.6451001 0.9363454
#> 51     no     26 0.8978357 0.7480927 0.9693833
#> 52    yes     26 0.8557676 0.6602792 0.9452652
#> 53     no     27 0.9073851 0.7587681 0.9740355
#> 54    yes     27 0.8686741 0.6751336 0.9530231
#> 55     no     28 0.9161253 0.7691272 0.9779966
#> 56    yes     28 0.8805867 0.6896410 0.9600040
#> 57     no     29 0.9241097 0.7791674 0.9813648
#> 58    yes     29 0.8915537 0.7037821 0.9660373
#> 59     no     30 0.9313909 0.7888873 0.9842255
#> 60    yes     30 0.9016260 0.7175397 0.9711875
#> 61     no     31 0.9380204 0.7989311 0.9866530
#> 62    yes     31 0.9108563 0.7308996 0.9755762
#> 63     no     32 0.9440478 0.8086682 0.9887110
#> 64    yes     32 0.9192981 0.7438500 0.9793104
#> 65     no     33 0.9495206 0.8180418 0.9904547
#> 66    yes     33 0.9270045 0.7563818 0.9824838
#> 67     no     34 0.9544839 0.8270550 0.9919312
#> 68    yes     34 0.9340278 0.7684880 0.9851776
#> 69     no     35 0.9589802 0.8357120 0.9931808
#> 70    yes     35 0.9404188 0.7793388 0.9874624
#> 71     no     36 0.9630496 0.8440179 0.9942380
#> 72    yes     36 0.9462263 0.7897707 0.9893986
#> 73     no     37 0.9667293 0.8519787 0.9951320
#> 74    yes     37 0.9514970 0.7998365 0.9910384
#> 75     no     38 0.9700539 0.8596013 0.9958878
#> 76    yes     38 0.9562749 0.8095370 0.9924265
#> 77     no     39 0.9730556 0.8668930 0.9965267
#> 78    yes     39 0.9606016 0.8188741 0.9936009
#> 79     no     40 0.9757639 0.8738620 0.9970666
#> 80    yes     40 0.9645161 0.8278511 0.9945941
#> 81     no     41 0.9782061 0.8805168 0.9975227
#> 82    yes     41 0.9680546 0.8364723 0.9954338
#> 83     no     42 0.9804072 0.8868662 0.9979080
#> 84    yes     42 0.9712507 0.8447429 0.9961435
#> 85     no     43 0.9823899 0.8929195 0.9982335
#> 86    yes     43 0.9741357 0.8526692 0.9967433
#> 87     no     44 0.9841753 0.8986862 0.9985084
#> 88    yes     44 0.9767380 0.8602579 0.9972500
#> 89     no     45 0.9857822 0.9041759 0.9987406
#> 90    yes     45 0.9790842 0.8675167 0.9976780
#> 91     no     46 0.9872281 0.9093983 0.9989367
#> 92    yes     46 0.9811982 0.8744536 0.9980395
#> 93     no     47 0.9885287 0.9143632 0.9991022
#> 94    yes     47 0.9831023 0.8810772 0.9983449
#> 95     no     48 0.9896982 0.9190804 0.9992420
#> 96    yes     48 0.9848165 0.8873965 0.9986027
#> 97     no     49 0.9907496 0.9235596 0.9993600
#> 98    yes     49 0.9863593 0.8934208 0.9988204
#> 99     no     50 0.9916946 0.9278104 0.9994597
#> 100   yes     50 0.9877472 0.8991595 0.9990042
#> 101    no     51 0.9925437 0.9318425 0.9995438
#> 102   yes     51 0.9889955 0.9046222 0.9991593
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)