Bootstrap simulation for model prediction

Generate model predictions against a specified set of explanatory levels with bootstrapped confidence intervals.

Usage

bootPredict(fit, newdata, R = 100, type = "response", ...)

Arguments

fit

An object of class lm or glm

newdata

A data.frame

R

Number of simulations. Note default R=100 is very low.

type

he type of prediction required, see predict.glm. The default for glm models is on the scale of the response variable. Thus for a binomial model the default predictions are predicted probabilities.

...

Further arguments to be passed to boot::boot

Value

An object of class "data.frame"

Examples

data(GBSG2,package="TH.data")
fit=glm(cens~horTh+pnodes,data=GBSG2,family="binomial")
newdata=expand.grid(horTh=factor(c(1,2),labels=c("no","yes")),pnodes=1:51)
bootPredict(fit,newdata)
#>     horTh pnodes  estimate     lower     upper
#> 1      no      1 0.3671628 0.3024052 0.4348178
#> 2     yes      1 0.2814584 0.2143347 0.3562284
#> 3      no      2 0.3927672 0.3337834 0.4553914
#> 4     yes      2 0.3039568 0.2372001 0.3749931
#> 5      no      3 0.4189750 0.3645758 0.4761225
#> 6     yes      3 0.3274340 0.2623082 0.3941445
#> 7      no      4 0.4456481 0.3994901 0.4967043
#> 8     yes      4 0.3518080 0.2879614 0.4177982
#> 9      no      5 0.4726379 0.4285585 0.5264804
#> 10    yes      5 0.3769795 0.3108784 0.4427623
#> 11     no      6 0.4997887 0.4536769 0.5571338
#> 12    yes      6 0.4028326 0.3313663 0.4693986
#> 13     no      7 0.5269407 0.4744257 0.5912753
#> 14    yes      7 0.4292372 0.3518514 0.5002668
#> 15     no      8 0.5539343 0.4952640 0.6300968
#> 16    yes      8 0.4560509 0.3728977 0.5369023
#> 17     no      9 0.5806134 0.5154907 0.6669169
#> 18    yes      9 0.4831217 0.3928661 0.5727934
#> 19     no     10 0.6068295 0.5337426 0.7017847
#> 20    yes     10 0.5102919 0.4140060 0.6076424
#> 21     no     11 0.6324443 0.5556633 0.7344288
#> 22    yes     11 0.5374015 0.4357755 0.6412390
#> 23     no     12 0.6573325 0.5749686 0.7646677
#> 24    yes     12 0.5642918 0.4584529 0.6756466
#> 25     no     13 0.6813846 0.5942256 0.7919154
#> 26    yes     13 0.5908091 0.4802209 0.7103002
#> 27     no     14 0.7045073 0.6131930 0.8161979
#> 28    yes     14 0.6168084 0.5007451 0.7425683
#> 29     no     15 0.7266252 0.6312700 0.8382192
#> 30    yes     15 0.6421564 0.5212626 0.7723732
#> 31     no     16 0.7476805 0.6473002 0.8580554
#> 32    yes     16 0.6667336 0.5393966 0.7996355
#> 33     no     17 0.7676327 0.6630058 0.8760557
#> 34    yes     17 0.6904365 0.5564220 0.8243520
#> 35     no     18 0.7864578 0.6783599 0.8921423
#> 36    yes     18 0.7131789 0.5733182 0.8458420
#> 37     no     19 0.8041469 0.6933383 0.9063635
#> 38    yes     19 0.7348919 0.5904408 0.8650194
#> 39     no     20 0.8207047 0.7079197 0.9188795
#> 40    yes     20 0.7555247 0.6075984 0.8821401
#> 41     no     21 0.8361480 0.7220858 0.9298516
#> 42    yes     21 0.7750433 0.6244918 0.8973442
#> 43     no     22 0.8505034 0.7358214 0.9394371
#> 44    yes     22 0.7934296 0.6415837 0.9107829
#> 45     no     23 0.8638059 0.7491141 0.9477861
#> 46    yes     23 0.8106802 0.6589480 0.9226120
#> 47     no     24 0.8760972 0.7619544 0.9550389
#> 48    yes     24 0.8268046 0.6761121 0.9329865
#> 49     no     25 0.8874238 0.7743355 0.9613252
#> 50    yes     25 0.8418238 0.6915764 0.9420561
#> 51     no     26 0.8978357 0.7862533 0.9667629
#> 52    yes     26 0.8557676 0.7066223 0.9499629
#> 53     no     27 0.9073851 0.7977061 0.9714586
#> 54    yes     27 0.8686741 0.7212295 0.9568392
#> 55     no     28 0.9161253 0.8086946 0.9755076
#> 56    yes     28 0.8805867 0.7353809 0.9628067
#> 57     no     29 0.9241097 0.8192214 0.9789945
#> 58    yes     29 0.8915537 0.7491445 0.9679760
#> 59     no     30 0.9313909 0.8292911 0.9819940
#> 60    yes     30 0.9016260 0.7625135 0.9724468
#> 61     no     31 0.9380204 0.8389101 0.9845719
#> 62    yes     31 0.9108563 0.7753823 0.9763083
#> 63     no     32 0.9440478 0.8480864 0.9867856
#> 64    yes     32 0.9192981 0.7877466 0.9796394
#> 65     no     33 0.9495206 0.8568290 0.9886852
#> 66    yes     33 0.9270045 0.7996047 0.9825102
#> 67     no     34 0.9544839 0.8651484 0.9903143
#> 68    yes     34 0.9340278 0.8109579 0.9849822
#> 69     no     35 0.9589802 0.8730559 0.9917108
#> 70    yes     35 0.9404188 0.8218097 0.9871090
#> 71     no     36 0.9630496 0.8805699 0.9929074
#> 72    yes     36 0.9462263 0.8321659 0.9889379
#> 73     no     37 0.9667293 0.8877222 0.9939322
#> 74    yes     37 0.9514970 0.8420345 0.9905095
#> 75     no     38 0.9700539 0.8944974 0.9948097
#> 76    yes     38 0.9562749 0.8514249 0.9918596
#> 77     no     39 0.9730556 0.9009092 0.9955608
#> 78    yes     39 0.9606016 0.8603481 0.9930188
#> 79     no     40 0.9757639 0.9069719 0.9962036
#> 80    yes     40 0.9645161 0.8688167 0.9940138
#> 81     no     41 0.9782061 0.9126994 0.9967536
#> 82    yes     41 0.9680546 0.8768438 0.9948676
#> 83     no     42 0.9804072 0.9181060 0.9972241
#> 84    yes     42 0.9712507 0.8844437 0.9956001
#> 85     no     43 0.9823899 0.9232059 0.9976266
#> 86    yes     43 0.9741357 0.8916313 0.9962283
#> 87     no     44 0.9841753 0.9280130 0.9979708
#> 88    yes     44 0.9767380 0.8984219 0.9967671
#> 89     no     45 0.9857822 0.9325411 0.9982652
#> 90    yes     45 0.9790842 0.9048313 0.9972291
#> 91     no     46 0.9872281 0.9368037 0.9985169
#> 92    yes     46 0.9811982 0.9108753 0.9976252
#> 93     no     47 0.9885287 0.9408139 0.9987321
#> 94    yes     47 0.9831023 0.9165698 0.9979647
#> 95     no     48 0.9896982 0.9445847 0.9989161
#> 96    yes     48 0.9848165 0.9219306 0.9982557
#> 97     no     49 0.9907496 0.9481284 0.9990735
#> 98    yes     49 0.9863593 0.9269735 0.9985052
#> 99     no     50 0.9916946 0.9514570 0.9992080
#> 100   yes     50 0.9877472 0.9317139 0.9987190
#> 101    no     51 0.9925437 0.9545822 0.9993230
#> 102   yes     51 0.9889955 0.9361669 0.9989022
library(survival)
fit=coxph(Surv(time,cens)~age+horTh+progrec+pnodes,data=GBSG2)