Make a list of univariable model with multivariable regression model

Make a list of univariable model with multivariable regression model

Usage

fit2list(fit)

Arguments

fit

An object of class "lm" or "glm"

Value

An object of class "fitlist" which is a list of objects of class "lm" or "glm"

Examples

library(survival)
data(cancer)
fit=glm(status~rx+sex+age+obstruct+nodes,data=colon,family="binomial")
fit2list(fit)
#> [[1]]
#> 
#> Call:  glm(formula = as.formula(myformula), family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)        rxLev    rxLev+5FU  
#>     0.19290     -0.06774     -0.62670  
#> 
#> Degrees of Freedom: 1821 Total (i.e. Null);  1819 Residual
#> Null Deviance:	    2525 
#> Residual Deviance: 2491 	AIC: 2497
#> 
#> [[2]]
#> 
#> Call:  glm(formula = as.formula(myformula), family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)          sex  
#>   -0.004598    -0.050038  
#> 
#> Degrees of Freedom: 1821 Total (i.e. Null);  1820 Residual
#> Null Deviance:	    2525 
#> Residual Deviance: 2525 	AIC: 2529
#> 
#> [[3]]
#> 
#> Call:  glm(formula = as.formula(myformula), family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)          age  
#>    0.211300    -0.004049  
#> 
#> Degrees of Freedom: 1821 Total (i.e. Null);  1820 Residual
#> Null Deviance:	    2525 
#> Residual Deviance: 2524 	AIC: 2528
#> 
#> [[4]]
#> 
#> Call:  glm(formula = as.formula(myformula), family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)     obstruct  
#>    -0.07895      0.24982  
#> 
#> Degrees of Freedom: 1821 Total (i.e. Null);  1820 Residual
#> Null Deviance:	    2525 
#> Residual Deviance: 2521 	AIC: 2525
#> 
#> [[5]]
#> 
#> Call:  glm(formula = as.formula(myformula), family = family, data = data)
#> 
#> Coefficients:
#> (Intercept)        nodes  
#>     -0.6886       0.1891  
#> 
#> Degrees of Freedom: 1821 Total (i.e. Null);  1820 Residual
#> Null Deviance:	    2525 
#> Residual Deviance: 2382 	AIC: 2386
#> 
#> attr(,"class")
#> [1] "fitlist"
fit=lm(mpg~wt*hp+am,data=mtcars)
fit2list(fit)
#> [[1]]
#> 
#> Call:
#> lm(formula = as.formula(myformula), data = data)
#> 
#> Coefficients:
#> (Intercept)           wt  
#>      37.285       -5.344  
#> 
#> 
#> [[2]]
#> 
#> Call:
#> lm(formula = as.formula(myformula), data = data)
#> 
#> Coefficients:
#> (Intercept)           hp  
#>    30.09886     -0.06823  
#> 
#> 
#> [[3]]
#> 
#> Call:
#> lm(formula = as.formula(myformula), data = data)
#> 
#> Coefficients:
#> (Intercept)           am  
#>      17.147        7.245  
#> 
#> 
#> [[4]]
#> 
#> Call:
#> lm(formula = as.formula(myformula), data = data)
#> 
#> Coefficients:
#> (Intercept)        wt:hp  
#>    27.74564     -0.01487  
#> 
#> 
#> attr(,"class")
#> [1] "fitlist"