Produce table for descriptive statistics

Produce table for descriptive statistics by groups for several variables easily. Depending on the nature of these variables, different descriptive statistical methods were used(t-test, ANOVA, Kruskal-Wallis, chi-squared, Fisher's,...)

Usage

gaze(x, ...)

# S3 method for formula
gaze(x, ...)

# S3 method for data.frame
gaze(x, ...)

Arguments

x

An R object, formula or data.frame

...

arguments to be passed to gaze.data.frame or gaze.formula

Methods (by class)

formula: S3 method for formula

data.frame: default S3 method

Examples

library(moonBook)
library(dplyr)
gaze(acs)
#> —————————————————————————————————————————————————
#>        name            levels          stats     
#> —————————————————————————————————————————————————
#> age               Mean ± SD           63.3 ± 11.7 
#> sex               Female              287 (33.5%) 
#>                   Male                570 (66.5%) 
#> cardiogenicShock  No                  805 (93.9%) 
#>                   Yes                   52 (6.1%) 
#> entry             Femoral             312 (36.4%) 
#>                   Radial              545 (63.6%) 
#> Dx                NSTEMI              153 (17.9%) 
#>                   STEMI               304 (35.5%) 
#>                   Unstable Angina     400 (46.7%) 
#> EF                Mean ± SD            55.8 ± 9.6 
#> height            Mean ± SD           163.2 ± 9.1 
#> weight            Mean ± SD           64.8 ± 11.4 
#> BMI               Mean ± SD            24.3 ± 3.3 
#> obesity           No                  567 (66.2%) 
#>                   Yes                 290 (33.8%) 
#> TC                Mean ± SD          185.2 ± 47.8 
#> LDLC              Mean ± SD          116.6 ± 41.1 
#> HDLC              Mean ± SD           38.2 ± 11.1 
#> TG                Mean ± SD          125.2 ± 90.9 
#> DM                No                  553 (64.5%) 
#>                   Yes                 304 (35.5%) 
#> HBP               No                  356 (41.5%) 
#>                   Yes                 501 (58.5%) 
#> smoking           Ex-smoker           204 (23.8%) 
#>                   Never               332 (38.7%) 
#>                   Smoker              321 (37.5%) 
#> —————————————————————————————————————————————————
gaze(~age+sex,data=acs)
#> ——————————————————————————————
#>  name   levels       stats    
#> ——————————————————————————————
#> age   Mean ± SD    63.3 ± 11.7 
#> sex   Female       287 (33.5%) 
#>       Male         570 (66.5%) 
#> ——————————————————————————————
gaze(sex~.,data=acs,digits=1,method=1,show.p=TRUE) %>% myft()
#> a flextable object.
#> col_keys: `name`, `levels`, `Female (N=287)`, `Male (N=570)`, `p` 
#> header has 1 row(s) 
#> body has 25 row(s) 
#> original dataset sample: 
#> ——————————————————————————————————————————————————————————————————
#>   Dependent:sex     levels        Female          Male        p   
#>        (N)                       (N=287)        (N=570)           
#> ——————————————————————————————————————————————————————————————————
#> age               Mean ± SD       68.7 ± 10.7   60.6 ± 11.2  <.001 
#> cardiogenicShock  No              275 (95.8%)     530 (93%)   .136 
#>                   Yes               12 (4.2%)       40 (7%)        
#> entry             Femoral         119 (41.5%)   193 (33.9%)   .035 
#>                   Radial          168 (58.5%)   377 (66.1%)        
#> ——————————————————————————————————————————————————————————————————
# \donttest{
gaze(sex~age+Dx,data=acs)
#> —————————————————————————————————————————————————————————————————————
#>  Dependent:sex      levels           Female          Male        p   
#>       (N)                           (N=287)        (N=570)           
#> —————————————————————————————————————————————————————————————————————
#> age            Mean ± SD             68.7 ± 10.7   60.6 ± 11.2  <.001 
#> Dx             NSTEMI                 50 (17.4%)   103 (18.1%)   .012 
#>                STEMI                  84 (29.3%)   220 (38.6%)        
#>                Unstable Angina       153 (53.3%)   247 (43.3%)        
#> —————————————————————————————————————————————————————————————————————
gaze(EF~.,data=acs) %>% myft()
#> a flextable object.
#> col_keys: `name`, `levels`, `unit`, `value` 
#> header has 1 row(s) 
#> body has 26 row(s) 
#> original dataset sample: 
#> ——————————————————————————————————————————————————————————
#>    Dependent:EF        levels        unit        value    
#> ——————————————————————————————————————————————————————————
#> age               [28,91]           Mean ± SD  63.3 ± 11.7 
#> sex               Female (N=287)    Mean ± SD  56.3 ± 10.1 
#>                   Male   (N=570)    Mean ± SD   55.6 ± 9.4 
#> cardiogenicShock  No  (N=805)       Mean ± SD   56.1 ± 9.4 
#>                   Yes (N=52)        Mean ± SD  50.2 ± 11.6 
#> ——————————————————————————————————————————————————————————
gaze(sex+Dx~.,data=acs,show.p=TRUE) %>% myft()
#> a flextable object.
#> col_keys: `name`, `levels`, `NSTEMI (N=50)`, `STEMI (N=84)`, `Unstable Angina (N=153)`, `p`, `NSTEMI (N=103)`, `STEMI (N=220)`, `Unstable Angina (N=247)`, `p  ` 
#> header has 2 row(s) 
#> body has 22 row(s) 
#> original dataset sample: 
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>            sex (N)                                  Female (N=287)                                                 Male (N=570)                           
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>    Dependent:Dx     levels       NSTEMI         STEMI          Unstable Angina        p        NSTEMI          STEMI          Unstable Angina        p   
#>        (N)                       (N=50)         (N=84)             (N=153)                    (N=103)         (N=220)             (N=247)                
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> age               Mean ± SD      70.9 ± 11.4   69.1 ± 10.4              67.7 ± 10.7   .177     61.1 ± 11.6    59.4 ± 11.7              61.4 ± 10.6   .133 
#> cardiogenicShock  No                49 (98%)    73 (86.9%)               153 (100%)  <.001     100 (97.1%)    183 (83.2%)               247 (100%)  <.001 
#>                   Yes                 1 (2%)    11 (13.1%)                   0 (0%)               3 (2.9%)     37 (16.8%)                   0 (0%)        
#> entry             Femoral           22 (44%)    45 (53.6%)                 52 (34%)   .013        36 (35%)       88 (40%)               69 (27.9%)   .022 
#>                   Radial            28 (56%)    39 (46.4%)                101 (66%)               67 (65%)      132 (60%)              178 (72.1%)        
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
gaze(sex+Dx~.,data=acs)
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>            sex (N)                                  Female (N=287)                                                 Male (N=570)                           
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>    Dependent:Dx     levels       NSTEMI         STEMI          Unstable Angina        p        NSTEMI          STEMI          Unstable Angina        p   
#>        (N)                       (N=50)         (N=84)             (N=153)                    (N=103)         (N=220)             (N=247)                
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> age               Mean ± SD      70.9 ± 11.4   69.1 ± 10.4              67.7 ± 10.7   .177     61.1 ± 11.6    59.4 ± 11.7              61.4 ± 10.6   .133 
#> cardiogenicShock  No                49 (98%)    73 (86.9%)               153 (100%)  <.001     100 (97.1%)    183 (83.2%)               247 (100%)  <.001 
#>                   Yes                 1 (2%)    11 (13.1%)                   0 (0%)               3 (2.9%)     37 (16.8%)                   0 (0%)        
#> entry             Femoral           22 (44%)    45 (53.6%)                 52 (34%)   .013        36 (35%)       88 (40%)               69 (27.9%)   .022 
#>                   Radial            28 (56%)    39 (46.4%)                101 (66%)               67 (65%)      132 (60%)              178 (72.1%)        
#> EF                Mean ± SD       54.8 ± 9.1   52.3 ± 10.9               59.4 ± 8.8  <.001      55.1 ± 9.4     52.4 ± 8.9               59.1 ± 8.7  <.001 
#> height            Mean ± SD      154.2 ± 5.1   155.7 ± 5.4              152.6 ± 6.7   .002     167.5 ± 5.7    168.7 ± 6.0              167.3 ± 6.4   .055 
#> weight            Mean ± SD      57.2 ± 10.3    57.4 ± 9.0               57.1 ± 9.1   .978      67.5 ± 8.4    68.8 ± 10.9              69.0 ± 10.6   .479 
#> BMI               Mean ± SD       24.1 ± 4.3    23.6 ± 3.2               24.5 ± 3.5   .215      24.1 ± 2.6     24.1 ± 3.4               24.6 ± 3.4   .205 
#> obesity           No                35 (70%)    60 (71.4%)               99 (64.7%)   .528      71 (68.9%)    149 (67.7%)              153 (61.9%)   .301 
#>                   Yes               15 (30%)    24 (28.6%)               54 (35.3%)             32 (31.1%)     71 (32.3%)               94 (38.1%)        
#> TC                Mean ± SD     196.3 ± 52.7  180.7 ± 45.7             191.1 ± 53.1   .192    192.6 ± 54.3   184.1 ± 42.6             178.7 ± 44.6   .036 
#> LDLC              Mean ± SD     127.7 ± 39.5  111.0 ± 40.0             118.3 ± 41.8   .088    125.4 ± 47.1   118.9 ± 39.1             109.5 ± 39.2   .002 
#> HDLC              Mean ± SD      40.1 ± 13.8   39.5 ± 11.2              38.5 ± 10.8   .627     38.4 ± 10.9    38.1 ± 10.9              37.4 ± 10.9   .655 
#> TG                Mean ± SD     112.5 ± 51.1  112.3 ± 87.2             126.3 ± 76.0   .316   138.0 ± 100.2   104.3 ± 65.5            144.3 ± 114.2  <.001 
#> DM                No                25 (50%)    54 (64.3%)               94 (61.4%)   .240      71 (68.9%)      154 (70%)              155 (62.8%)   .219 
#>                   Yes               25 (50%)    30 (35.7%)               59 (38.6%)             32 (31.1%)       66 (30%)               92 (37.2%)        
#> HBP               No                19 (38%)    28 (33.3%)               36 (23.5%)   .084      43 (41.7%)    122 (55.5%)              108 (43.7%)   .016 
#>                   Yes               31 (62%)    56 (66.7%)              117 (76.5%)             60 (58.3%)     98 (44.5%)              139 (56.3%)        
#> smoking           Ex-smoker          8 (16%)    13 (15.5%)               28 (18.3%)   .184        34 (33%)     53 (24.1%)               68 (27.5%)   .002 
#>                   Never             37 (74%)    57 (67.9%)              115 (75.2%)             13 (12.6%)     40 (18.2%)               70 (28.3%)        
#>                   Smoker             5 (10%)    14 (16.7%)                10 (6.5%)             56 (54.4%)    127 (57.7%)              109 (44.1%)        
#> —————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
gaze(Dx+sex~cardiogenicShock,data=acs,show.p=TRUE) %>% myft()
#> a flextable object.
#> col_keys: `name`, `levels`, `Female (N=50)`, `Male (N=103)`, `p`, `Female (N=84)`, `Male (N=220)`, `p  `, `Female (N=153)`, `Male (N=247)`, `p   ` 
#> header has 2 row(s) 
#> body has 2 row(s) 
#> original dataset sample: 
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>           Dx (N)                     NSTEMI (N=153)                      STEMI (N=304)                  Unstable Angina (N=400)        
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Dependent:sex    levels     Female          Male        p       Female          Male       p        Female          Male        p   
#>        (N)                    (N=50)        (N=103)               (N=84)        (N=220)              (N=153)        (N=247)           
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> cardiogenicShock  No             49 (98%)   100 (97.1%)  1.000     73 (86.9%)   183 (83.2%)  .535      153 (100%)    247 (100%)  1.000 
#>                   Yes              1 (2%)      3 (2.9%)            11 (13.1%)    37 (16.8%)                0 (0%)        0 (0%)        
#> ——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
gaze(Dx+sex+HBP~cardiogenicShock,data=acs,show.p=TRUE)
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>             Dx                        NSTEMI                        NSTEMI                         STEMI                          STEMI                      Unstable Angina                  Unstable Angina           
#>            sex                        Female                         Male                          Female                          Male                           Female                            Male                
#>            (N)                        (N=50)                        (N=103)                        (N=84)                        (N=220)                         (N=153)                          (N=247)               
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>   Dependent:HBP    levels     No         Yes        p        No         Yes        p        No         Yes       p        No          Yes        p       No          Yes        p         No          Yes         p    
#>        (N)                  (N=19)      (N=31)             (N=43)      (N=60)             (N=28)      (N=56)            (N=122)      (N=98)            (N=36)      (N=117)             (N=108)      (N=139)            
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> cardiogenicShock  No        19 (100%)  30 (96.8%)  1.000  42 (97.7%)  58 (96.7%)  1.000  23 (82.1%)  50 (89.3%)  .567  106 (86.9%)  77 (78.6%)   .145  36 (100%)   117 (100%)   1.000  108 (100%)   139 (100%)    1.000 
#>                   Yes          0 (0%)    1 (3.2%)           1 (2.3%)    2 (3.3%)          5 (17.9%)   6 (10.7%)         16 (13.1%)  21 (21.4%)            0 (0%)       0 (0%)              0 (0%)       0 (0%)          
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
gaze(~mpg+cyl,data=mtcars)
#> —————————————————————————————
#>  name   levels      stats    
#> —————————————————————————————
#> mpg   Mean ± SD    20.1 ± 6.0 
#> cyl   Mean ± SD     6.2 ± 1.8 
#> —————————————————————————————
gaze(~.,data=mtcars)
#> ————————————————————————————————
#>  name   levels        stats     
#> ————————————————————————————————
#> mpg   Mean ± SD       20.1 ± 6.0 
#> cyl   Mean ± SD        6.2 ± 1.8 
#> disp  Mean ± SD    230.7 ± 123.9 
#> hp    Mean ± SD     146.7 ± 68.6 
#> drat  Mean ± SD        3.6 ± 0.5 
#> wt    Mean ± SD        3.2 ± 1.0 
#> qsec  Mean ± SD       17.8 ± 1.8 
#> vs    Mean ± SD        0.4 ± 0.5 
#> am    Mean ± SD        0.4 ± 0.5 
#> gear  Mean ± SD        3.7 ± 0.7 
#> carb  Mean ± SD        2.8 ± 1.6 
#> ————————————————————————————————
gaze(cyl~.,data=mtcars,show.p=TRUE)
#> ———————————————————————————————————————————————————————————————————————————
#>  Dependent:cyl   levels         4             6             8          p   
#>       (N)                     (N=11)        (N=7)         (N=14)           
#> ———————————————————————————————————————————————————————————————————————————
#> mpg            Mean ± SD      26.7 ± 4.5    19.7 ± 1.5    15.1 ± 2.6  <.001 
#> disp           Mean ± SD    105.1 ± 26.9  183.3 ± 41.6  353.1 ± 67.8  <.001 
#> hp             Mean ± SD     82.6 ± 20.9  122.3 ± 24.3  209.2 ± 51.0  <.001 
#> drat           Mean ± SD       4.1 ± 0.4     3.6 ± 0.5     3.2 ± 0.4  <.001 
#> wt             Mean ± SD       2.3 ± 0.6     3.1 ± 0.4     4.0 ± 0.8  <.001 
#> qsec           Mean ± SD      19.1 ± 1.7    18.0 ± 1.7    16.8 ± 1.2   .002 
#> vs             Mean ± SD       0.9 ± 0.3     0.6 ± 0.5     0.0 ± 0.0  <.001 
#> am             Mean ± SD       0.7 ± 0.5     0.4 ± 0.5     0.1 ± 0.4   .010 
#> gear           Mean ± SD       4.1 ± 0.5     3.9 ± 0.7     3.3 ± 0.7   .015 
#> carb           Mean ± SD       1.5 ± 0.5     3.4 ± 1.8     3.5 ± 1.6   .003 
#> ———————————————————————————————————————————————————————————————————————————
gaze(hp~.,data=mtcars)
#> —————————————————————————————————————————————————————
#>  Dependent:hp    levels       unit         value     
#> —————————————————————————————————————————————————————
#> mpg           [10.4,33.9]    Mean ± SD     20.1 ± 6.0 
#> cyl           [4,8]          Mean ± SD      6.2 ± 1.8 
#> disp          [71.1,472]     Mean ± SD  230.7 ± 123.9 
#> drat          [2.8,4.9]      Mean ± SD      3.6 ± 0.5 
#> wt            [1.5,5.4]      Mean ± SD      3.2 ± 1.0 
#> qsec          [14.5,22.9]    Mean ± SD     17.8 ± 1.8 
#> vs            [0,1]          Mean ± SD      0.4 ± 0.5 
#> am            [0,1]          Mean ± SD      0.4 ± 0.5 
#> gear          [3,5]          Mean ± SD      3.7 ± 0.7 
#> carb          [1,8]          Mean ± SD      2.8 ± 1.6 
#> —————————————————————————————————————————————————————
gaze(cyl+am~.,data=mtcars)
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>          cyl (N)                     4 (N=11)                          6 (N=7)                           8 (N=14)               
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#>  Dependent:am   levels         0             1        p         0             1         p         0             1          p   
#>      (N)                     (N=3)         (N=8)              (N=4)         (N=3)               (N=12)        (N=2)            
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
#> mpg           Mean ± SD      22.9 ± 1.5   28.1 ± 4.5  .089    19.1 ± 1.6    20.6 ± 0.8  .221    15.1 ± 2.8    15.4 ± 0.6   .866 
#> disp          Mean ± SD    135.9 ± 14.0  93.6 ± 20.5  .010  204.6 ± 44.7   155.0 ± 8.7  .124  357.6 ± 71.8  326.0 ± 35.4   .563 
#> hp            Mean ± SD     84.7 ± 19.7  81.9 ± 22.7  .856   115.2 ± 9.2  131.7 ± 37.5  .530  194.2 ± 33.4  299.5 ± 50.2   .002 
#> drat          Mean ± SD       3.8 ± 0.1    4.2 ± 0.4  .094     3.4 ± 0.6     3.8 ± 0.2  .330     3.1 ± 0.2     3.9 ± 0.5   .002 
#> wt            Mean ± SD       2.9 ± 0.4    2.0 ± 0.4  .010     3.4 ± 0.1     2.8 ± 0.1  .001     4.1 ± 0.8     3.4 ± 0.3   .218 
#> qsec          Mean ± SD      21.0 ± 1.7   18.4 ± 1.1  .017    19.2 ± 0.8    16.3 ± 0.8  .005    17.1 ± 0.8    14.6 ± 0.1  <.001 
#> vs            Mean ± SD       1.0 ± 0.0    0.9 ± 0.4  .351     1.0 ± 0.0     0.0 ± 0.0           0.0 ± 0.0     0.0 ± 0.0        
#> gear          Mean ± SD       3.7 ± 0.6    4.2 ± 0.5  .113     3.5 ± 0.6     4.3 ± 0.6  .117     3.0 ± 0.0     5.0 ± 0.0        
#> carb          Mean ± SD       1.7 ± 0.6    1.5 ± 0.5  .662     2.5 ± 1.7     4.7 ± 1.2  .122     3.1 ± 0.9     6.0 ± 2.8   .379 
#> ———————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
# }